% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/activate.R
\name{activate_family_group_record}
\alias{activate_family_group_record}
\title{Activate a Family Group record}
\usage{
activate_family_group_record(gedcom, family)
}
\arguments{
\item{gedcom}{A tidyged object.}

\item{family}{The xref of the Family Group record to be activated.}
}
\value{
The same tidyged object with "active_record" attribute set to the specific
Family group record to allow easy editing.
}
\description{
This allows a mechanism to easily edit particular records. For Family group records,
the explicit xref is required.
}
