% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/activate.R
\name{activate_indi}
\alias{activate_indi}
\alias{activate_famg}
\alias{activate_subm}
\alias{activate_media}
\alias{activate_note}
\alias{activate_sour}
\alias{activate_repo}
\title{Activate a record}
\usage{
activate_indi(gedcom, record)

activate_famg(gedcom, record)

activate_subm(gedcom, record)

activate_media(gedcom, record)

activate_note(gedcom, record)

activate_sour(gedcom, record)

activate_repo(gedcom, record)
}
\arguments{
\item{gedcom}{A tidyged object.}

\item{record}{The xref or attribute of the record to be activated.}
}
\value{
The same tidyged object with "active_record" attribute set to the specific
record to allow easy editing.
}
\description{
Set a specific record to be the active record.
}
\details{
This allows a mechanism to easily edit particular records. Either a specific record attribute can be provided
(can be a regex pattern) or the xref can be provided explicitly. Providing an attribute instead of the
xref can make your code more readable, provided that it returns a unique record!

The attributes that can be provided for each type of record are:
\itemize{
\item Individuals: Full name (NAME tag), phonetic variation (FONE), romanised variation (ROMN);
\item Multimedia: File reference (FILE);
\item Note: User text (NOTE);
\item Source: Title (TITL);
\item Repository: Name (NAME).
}
}
