% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/activate.R
\name{activate_multimedia_record}
\alias{activate_multimedia_record}
\title{Activate a Multimedia record}
\usage{
activate_multimedia_record(gedcom, multimedia)
}
\arguments{
\item{gedcom}{A tidyged object.}

\item{multimedia}{The xref or file reference of the Multimedia record to be activated.}
}
\value{
The same tidyged object with "active_record" attribute set to the specific
Multimedia record to allow easy editing.
}
\description{
This allows a mechanism to easily edit particular records. Either the file reference can be provided
or the xref can be provided explicitly. Providing the file reference instead of the
xref can make your code more readable, provided that it returns a unique record!
}
