% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/activate.R
\name{activate_note_record}
\alias{activate_note_record}
\title{Activate a Note record}
\usage{
activate_note_record(gedcom, note)
}
\arguments{
\item{gedcom}{A tidygedcom object.}

\item{note}{The xref or an excerpt of the Note record to be activated.}
}
\value{
The same tidygedcom object with "active_record" attribute set to the specific
Note record to allow easy editing.
}
\description{
This allows a mechanism to easily edit particular records. Either a note excerpt can be provided,
or the xref can be provided explicitly. Providing a note excerpt instead of the
xref can make your code more readable, provided that the excerpt returns a unique record!
}
