% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/activate.R
\name{activate_repository_record}
\alias{activate_repository_record}
\title{Activate a Repository record}
\usage{
activate_repository_record(gedcom, repository)
}
\arguments{
\item{gedcom}{A tidygedcom object.}

\item{repository}{The xref or name of the Repository record to be activated.}
}
\value{
The same tidygedcom object with "active_record" attribute set to the specific
Repository record to allow easy editing.
}
\description{
This allows a mechanism to easily edit particular records. Either the repository name can be provided
(can be a partial name) or the xref can be provided explicitly. Providing the name instead of the
xref can make your code more readable, provided that the name returns a unique record!
}
