% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/activate.R
\name{activate_source_record}
\alias{activate_source_record}
\title{Activate a Source record}
\usage{
activate_source_record(gedcom, source_title = character(), xref = character())
}
\arguments{
\item{gedcom}{A tidygedcom object}

\item{source_title}{The title of the Source record to be activated}

\item{xref}{The xref of the Source record to activate}
}
\value{
The same tidygedcom object with "active_record" attribute set to the specific
Source record to allow easy editing.
}
\description{
This allows a mechanism to easily edit particular records. Either the source title can be provided
or the xref can be provided explicitly. Providing the source title instead of the
xref can make your code more readable, provided that the name returns a unique record!
}
