% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/activate.R
\name{activate_submitter_record}
\alias{activate_submitter_record}
\title{Activate a Submitter record}
\usage{
activate_submitter_record(gedcom, submitter)
}
\arguments{
\item{gedcom}{A tidygedcom object.}

\item{submitter}{The xref or name of the Submitter record to be activated.}
}
\value{
The same tidygedcom object with "active_record" attribute set to the specific
Submitter record to allow easy editing.
}
\description{
This allows a mechanism to easily edit particular records. Either the submitter name can be provided
(can be a partial name) or the xref can be provided explicitly. Providing the name instead of the
xref can make your code more readable, provided that the name returns a unique record!
}
