% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/activate.R
\name{active_record}
\alias{active_record}
\title{Get the active record in a tidyged object}
\usage{
active_record(gedcom)
}
\arguments{
\item{gedcom}{A tidyged object.}
}
\value{
The xref of the active record.
}
\description{
Get the active record in a tidyged object
}
