% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrappers.R
\name{add_children}
\alias{add_children}
\title{Create multiple children for a Family Group}
\usage{
add_children(gedcom, xref = character(), sexes = NULL)
}
\arguments{
\item{gedcom}{A tidyged object.}

\item{xref}{The xref of a Family Group record to act on if one is not activated (will override active record).}

\item{sexes}{A character string giving the sexes of each child. For example,
"FFM" to add two daughters and one son. See the help for \code{add_indi()}
for possible codes.}
}
\value{
A tidyged object with additional child records.
}
\description{
Create multiple children for a Family Group
}
