% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/family_group.R
\name{add_famg}
\alias{add_famg}
\title{Add a Family Group record to a tidyged object}
\usage{
add_famg(
  gedcom,
  husband = character(),
  wife = character(),
  children = character(),
  number_of_children = character(),
  user_reference_numbers = character(),
  family_notes = character(),
  multimedia_links = character()
)
}
\arguments{
\item{gedcom}{A tidyged object.}

\item{husband}{An xref identifying the husband of this family.}

\item{wife}{An xref identifying the wife of this family.}

\item{children}{A character vector of xrefs identifying the children of this family. These are
assumed to be biological children ("birth"). If non-biological children are to be defined, use
a named vector to define the relationships using a name of either "adopted" or "foster".
For example: children = c("@I4@", adopted = "@I7@", adopted = "@I10@") defines a single biological
child and two adopted children.}

\item{number_of_children}{The reported number of children known to belong to this family,
regardless of whether the associated children are represented here.}

\item{user_reference_numbers}{A unique user-defined number or text that the submitter
uses to identify this record. You can supply more than one in a vector. You can also define a
user reference type by using a named vector (e.g c(id1 = "123A", id2 = "456B")).}

\item{family_notes}{A character vector of notes accompanying this Family group record. These could be
xrefs to existing Note records.}

\item{multimedia_links}{A character vector of Multimedia record xrefs accompanying this record.}
}
\value{
An updated tidyged object including the Family group record.
}
\description{
Add a Family Group record to a tidyged object
}
\details{
If you need to add further information about this family (i.e. events), use the
\code{add_famg_event()} function.

The function will automatically add links to this family to the respective Individual
records of the wife, husband, and children.
}
