% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/family_events.R
\name{add_famg_event}
\alias{add_famg_event}
\title{Add a family event to a Family Group record}
\usage{
add_famg_event(
  gedcom,
  type,
  descriptor = "",
  classification = character(),
  date = character(),
  husband_age = character(),
  wife_age = character(),
  cause = character(),
  event_place = place(),
  event_address = address(),
  notes = character(),
  responsible_agency = character(),
  religious_affiliation = character(),
  multimedia_links = character(),
  xref = character(),
  update_date_changed = TRUE
)
}
\arguments{
\item{gedcom}{A tidyged object.}

\item{type}{A (case-insensitive) three-letter code giving the type of event. See Details.}

\item{descriptor}{A short description of an 'other' event.}

\item{classification}{A descriptive word or phrase used to further classify this
event. This should be used whenever the 'other' event is used (but can also be used
with others). Recommended values for non-marriage relationships can be found in Details.}

\item{date}{A date_calendar(), date_approximated(), date_period(), or date_range()
object giving the timing of the event..}

\item{husband_age}{A character string that indicates the age in years, months, and days
that the husband was at the time of the event. Any combination of these is permitted.
Any labels must come after their corresponding number, for example; "4y 8m 10d".}

\item{wife_age}{A character string that indicates the age in years, months, and days
that the wife was at the time of the event. Any combination of these is permitted.
Any labels must come after their corresponding number, for example; "4y 8m 10d".}

\item{cause}{Used in special cases to record the reasons which precipitated an event.}

\item{event_place}{A place() object giving the place associated with this event.}

\item{event_address}{An address() object giving the address associated with this event.}

\item{notes}{A character vector of notes accompanying the event. These could be xrefs to
existing Note records.}

\item{responsible_agency}{The organisation, institution, corporation, person, or other
entity that has responsibility for the event.}

\item{religious_affiliation}{A name of the religion with which this event was affiliated.}

\item{multimedia_links}{A character vector of Multimedia record xrefs accompanying this
record.}

\item{xref}{The xref of a record to act on if one is not activated (will override active record).}

\item{update_date_changed}{Whether to add/update the change date for the record.}
}
\value{
An updated tidyged object with an expanded Family group record including
this event.
}
\description{
Add a family event to a Family Group record
}
\details{
The three-letter codes used for the type parameter are:

ann(ulment), cen(sus), div(orce), dif (divorce filed), eng(agement), mab (marriage banns),
mac (marriage contract), mal (marriage license), rel(ationship), mas (marriage settlement),
res(idence).

Alternatively eve (for any other event).

Example classifications of non-marriage relationships are:

not married, civil, living together, living apart together. See page 59 of the GEDCOM
Specification for more examples.
}
