% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/edit_family_events.R
\name{add_family_event}
\alias{add_family_event}
\alias{add_family_event_annulment}
\alias{add_family_event_census}
\alias{add_family_event_divorce}
\alias{add_family_event_divorce_filed}
\alias{add_family_event_engagement}
\alias{add_family_event_marriage_banns}
\alias{add_family_event_marriage_contract}
\alias{add_family_event_marriage}
\alias{add_family_event_marriage_license}
\alias{add_family_event_marriage_settlement}
\alias{add_family_event_residence}
\alias{add_family_event_other}
\title{Add a family event to a family record}
\usage{
add_family_event(
  gedcom,
  family_xref = character(),
  event_type = character(),
  event_subtype = character(),
  event_descriptor = character(),
  husband_age_at_event = character(),
  wife_age_at_event = character(),
  event_notes = character(),
  event_date = date_value(),
  event_cause = character(),
  place_name = character(),
  place_hierarchy = character(),
  place_phonetic_variation = character(),
  phonetic_type = character(),
  place_romanized_variation = character(),
  romanized_type = character(),
  place_latitude = character(),
  place_longitude = character(),
  place_notes = character(),
  address_first_line = character(),
  city = character(),
  state = character(),
  postal_code = character(),
  country = character(),
  phone_number = character(),
  email = character(),
  fax = character(),
  web_page = character(),
  responsible_agency = character(),
  religious_affiliation = character(),
  cause_of_event = character(),
  restriction_notice = character()
)

add_family_event_annulment(
  gedcom,
  family_xref = character(),
  event_type = "ANUL",
  event_subtype = character(),
  event_descriptor = character(),
  husband_age_at_event = character(),
  wife_age_at_event = character(),
  event_notes = character(),
  event_date = date_value(),
  event_cause = character(),
  place_name = character(),
  place_hierarchy = character(),
  place_phonetic_variation = character(),
  phonetic_type = character(),
  place_romanized_variation = character(),
  romanized_type = character(),
  place_latitude = character(),
  place_longitude = character(),
  place_notes = character(),
  address_first_line = character(),
  city = character(),
  state = character(),
  postal_code = character(),
  country = character(),
  phone_number = character(),
  email = character(),
  fax = character(),
  web_page = character(),
  responsible_agency = character(),
  religious_affiliation = character(),
  cause_of_event = character(),
  restriction_notice = character()
)

add_family_event_census(
  gedcom,
  family_xref = character(),
  event_type = "CENS",
  event_subtype = character(),
  event_descriptor = character(),
  husband_age_at_event = character(),
  wife_age_at_event = character(),
  event_notes = character(),
  event_date = date_value(),
  event_cause = character(),
  place_name = character(),
  place_hierarchy = character(),
  place_phonetic_variation = character(),
  phonetic_type = character(),
  place_romanized_variation = character(),
  romanized_type = character(),
  place_latitude = character(),
  place_longitude = character(),
  place_notes = character(),
  address_first_line = character(),
  city = character(),
  state = character(),
  postal_code = character(),
  country = character(),
  phone_number = character(),
  email = character(),
  fax = character(),
  web_page = character(),
  responsible_agency = character(),
  religious_affiliation = character(),
  cause_of_event = character(),
  restriction_notice = character()
)

add_family_event_divorce(
  gedcom,
  family_xref = character(),
  event_type = "DIV",
  event_subtype = character(),
  event_descriptor = character(),
  husband_age_at_event = character(),
  wife_age_at_event = character(),
  event_notes = character(),
  event_date = date_value(),
  event_cause = character(),
  place_name = character(),
  place_hierarchy = character(),
  place_phonetic_variation = character(),
  phonetic_type = character(),
  place_romanized_variation = character(),
  romanized_type = character(),
  place_latitude = character(),
  place_longitude = character(),
  place_notes = character(),
  address_first_line = character(),
  city = character(),
  state = character(),
  postal_code = character(),
  country = character(),
  phone_number = character(),
  email = character(),
  fax = character(),
  web_page = character(),
  responsible_agency = character(),
  religious_affiliation = character(),
  cause_of_event = character(),
  restriction_notice = character()
)

add_family_event_divorce_filed(
  gedcom,
  family_xref = character(),
  event_type = "DIVF",
  event_subtype = character(),
  event_descriptor = character(),
  husband_age_at_event = character(),
  wife_age_at_event = character(),
  event_notes = character(),
  event_date = date_value(),
  event_cause = character(),
  place_name = character(),
  place_hierarchy = character(),
  place_phonetic_variation = character(),
  phonetic_type = character(),
  place_romanized_variation = character(),
  romanized_type = character(),
  place_latitude = character(),
  place_longitude = character(),
  place_notes = character(),
  address_first_line = character(),
  city = character(),
  state = character(),
  postal_code = character(),
  country = character(),
  phone_number = character(),
  email = character(),
  fax = character(),
  web_page = character(),
  responsible_agency = character(),
  religious_affiliation = character(),
  cause_of_event = character(),
  restriction_notice = character()
)

add_family_event_engagement(
  gedcom,
  family_xref = character(),
  event_type = "ENGA",
  event_subtype = character(),
  event_descriptor = character(),
  husband_age_at_event = character(),
  wife_age_at_event = character(),
  event_notes = character(),
  event_date = date_value(),
  event_cause = character(),
  place_name = character(),
  place_hierarchy = character(),
  place_phonetic_variation = character(),
  phonetic_type = character(),
  place_romanized_variation = character(),
  romanized_type = character(),
  place_latitude = character(),
  place_longitude = character(),
  place_notes = character(),
  address_first_line = character(),
  city = character(),
  state = character(),
  postal_code = character(),
  country = character(),
  phone_number = character(),
  email = character(),
  fax = character(),
  web_page = character(),
  responsible_agency = character(),
  religious_affiliation = character(),
  cause_of_event = character(),
  restriction_notice = character()
)

add_family_event_marriage_banns(
  gedcom,
  family_xref = character(),
  event_type = "MARB",
  event_subtype = character(),
  event_descriptor = character(),
  husband_age_at_event = character(),
  wife_age_at_event = character(),
  event_notes = character(),
  event_date = date_value(),
  event_cause = character(),
  place_name = character(),
  place_hierarchy = character(),
  place_phonetic_variation = character(),
  phonetic_type = character(),
  place_romanized_variation = character(),
  romanized_type = character(),
  place_latitude = character(),
  place_longitude = character(),
  place_notes = character(),
  address_first_line = character(),
  city = character(),
  state = character(),
  postal_code = character(),
  country = character(),
  phone_number = character(),
  email = character(),
  fax = character(),
  web_page = character(),
  responsible_agency = character(),
  religious_affiliation = character(),
  cause_of_event = character(),
  restriction_notice = character()
)

add_family_event_marriage_contract(
  gedcom,
  family_xref = character(),
  event_type = "MARC",
  event_subtype = character(),
  event_descriptor = character(),
  husband_age_at_event = character(),
  wife_age_at_event = character(),
  event_notes = character(),
  event_date = date_value(),
  event_cause = character(),
  place_name = character(),
  place_hierarchy = character(),
  place_phonetic_variation = character(),
  phonetic_type = character(),
  place_romanized_variation = character(),
  romanized_type = character(),
  place_latitude = character(),
  place_longitude = character(),
  place_notes = character(),
  address_first_line = character(),
  city = character(),
  state = character(),
  postal_code = character(),
  country = character(),
  phone_number = character(),
  email = character(),
  fax = character(),
  web_page = character(),
  responsible_agency = character(),
  religious_affiliation = character(),
  cause_of_event = character(),
  restriction_notice = character()
)

add_family_event_marriage(
  gedcom,
  family_xref = character(),
  event_type = "MARR",
  event_subtype = character(),
  event_descriptor = character(),
  husband_age_at_event = character(),
  wife_age_at_event = character(),
  event_notes = character(),
  event_date = date_value(),
  event_cause = character(),
  place_name = character(),
  place_hierarchy = character(),
  place_phonetic_variation = character(),
  phonetic_type = character(),
  place_romanized_variation = character(),
  romanized_type = character(),
  place_latitude = character(),
  place_longitude = character(),
  place_notes = character(),
  address_first_line = character(),
  city = character(),
  state = character(),
  postal_code = character(),
  country = character(),
  phone_number = character(),
  email = character(),
  fax = character(),
  web_page = character(),
  responsible_agency = character(),
  religious_affiliation = character(),
  cause_of_event = character(),
  restriction_notice = character()
)

add_family_event_marriage_license(
  gedcom,
  family_xref = character(),
  event_type = "MARL",
  event_subtype = character(),
  event_descriptor = character(),
  husband_age_at_event = character(),
  wife_age_at_event = character(),
  event_notes = character(),
  event_date = date_value(),
  event_cause = character(),
  place_name = character(),
  place_hierarchy = character(),
  place_phonetic_variation = character(),
  phonetic_type = character(),
  place_romanized_variation = character(),
  romanized_type = character(),
  place_latitude = character(),
  place_longitude = character(),
  place_notes = character(),
  address_first_line = character(),
  city = character(),
  state = character(),
  postal_code = character(),
  country = character(),
  phone_number = character(),
  email = character(),
  fax = character(),
  web_page = character(),
  responsible_agency = character(),
  religious_affiliation = character(),
  cause_of_event = character(),
  restriction_notice = character()
)

add_family_event_marriage_settlement(
  gedcom,
  family_xref = character(),
  event_type = "MARS",
  event_subtype = character(),
  event_descriptor = character(),
  husband_age_at_event = character(),
  wife_age_at_event = character(),
  event_notes = character(),
  event_date = date_value(),
  event_cause = character(),
  place_name = character(),
  place_hierarchy = character(),
  place_phonetic_variation = character(),
  phonetic_type = character(),
  place_romanized_variation = character(),
  romanized_type = character(),
  place_latitude = character(),
  place_longitude = character(),
  place_notes = character(),
  address_first_line = character(),
  city = character(),
  state = character(),
  postal_code = character(),
  country = character(),
  phone_number = character(),
  email = character(),
  fax = character(),
  web_page = character(),
  responsible_agency = character(),
  religious_affiliation = character(),
  cause_of_event = character(),
  restriction_notice = character()
)

add_family_event_residence(
  gedcom,
  family_xref = character(),
  event_type = "RESI",
  event_subtype = character(),
  event_descriptor = character(),
  husband_age_at_event = character(),
  wife_age_at_event = character(),
  event_notes = character(),
  event_date = date_value(),
  event_cause = character(),
  place_name = character(),
  place_hierarchy = character(),
  place_phonetic_variation = character(),
  phonetic_type = character(),
  place_romanized_variation = character(),
  romanized_type = character(),
  place_latitude = character(),
  place_longitude = character(),
  place_notes = character(),
  address_first_line = character(),
  city = character(),
  state = character(),
  postal_code = character(),
  country = character(),
  phone_number = character(),
  email = character(),
  fax = character(),
  web_page = character(),
  responsible_agency = character(),
  religious_affiliation = character(),
  cause_of_event = character(),
  restriction_notice = character()
)

add_family_event_other(
  gedcom,
  family_xref = character(),
  event_type = "EVEN",
  event_subtype = character(),
  event_descriptor = character(),
  husband_age_at_event = character(),
  wife_age_at_event = character(),
  event_notes = character(),
  event_date = date_value(),
  event_cause = character(),
  place_name = character(),
  place_hierarchy = character(),
  place_phonetic_variation = character(),
  phonetic_type = character(),
  place_romanized_variation = character(),
  romanized_type = character(),
  place_latitude = character(),
  place_longitude = character(),
  place_notes = character(),
  address_first_line = character(),
  city = character(),
  state = character(),
  postal_code = character(),
  country = character(),
  phone_number = character(),
  email = character(),
  fax = character(),
  web_page = character(),
  responsible_agency = character(),
  religious_affiliation = character(),
  cause_of_event = character(),
  restriction_notice = character()
)
}
\arguments{
\item{gedcom}{}

\item{family_xref}{}

\item{event_type}{}

\item{event_subtype}{}

\item{event_descriptor}{}

\item{husband_age_at_event}{}

\item{wife_age_at_event}{}

\item{event_notes}{}

\item{event_date}{}

\item{event_cause}{}

\item{place_name}{}

\item{place_hierarchy}{}

\item{place_phonetic_variation}{}

\item{phonetic_type}{}

\item{place_romanized_variation}{}

\item{romanized_type}{}

\item{place_latitude}{}

\item{place_longitude}{}

\item{place_notes}{}

\item{address_first_line}{}

\item{city}{}

\item{state}{}

\item{postal_code}{}

\item{country}{}

\item{phone_number}{}

\item{email}{}

\item{fax}{}

\item{web_page}{}

\item{responsible_agency}{}

\item{religious_affiliation}{}

\item{cause_of_event}{}

\item{restriction_notice}{}
}
\value{

}
\description{
Add a family event to a family record
}
