% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/edit_family_group.R
\name{add_family_group}
\alias{add_family_group}
\title{Add a Family Group record to a tidygedcom object}
\usage{
add_family_group(
  gedcom,
  husband = character(),
  wife = character(),
  children = character(),
  child_linkage_types = character(),
  number_of_children = character(),
  user_reference_number = character(),
  user_reference_type = character(),
  automated_record_id = character(),
  family_notes = character(),
  multimedia_links = character()
)
}
\arguments{
\item{gedcom}{A tidygedcom object.}

\item{husband}{A character string identifying the husband of this family. This can either
be an xref or a regular expression to match to an individual name.}

\item{wife}{A character string identifying the wife of this family. This can either
be an xref or a regular expression to match to an individual name.}

\item{children}{A character vector of other Individual records that are children of this
family. Children can either be referenced by an xref or by a regular expression
to match to an individual name.}

\item{child_linkage_types}{Codes used to indicate the child to family relationships. If defined,
this must be a character vector the same size as children. Values must be one of:
adopted, birth, foster, sealing.}

\item{number_of_children}{The reported number of children known to belong to this family,
regardless of whether the associated children are represented here.}

\item{user_reference_number}{A user-defined number or text that the submitter uses to identify
this record. See the Gedcom 5.5.5 Specification for more details.}

\item{user_reference_type}{A user-defined definition of the user_reference_number.}

\item{automated_record_id}{A unique record identification number assigned to the record by
the source system.}

\item{family_notes}{A character vector of notes accompanying this Family record. These could be
xrefs to existing Note records.}

\item{multimedia_links}{A character vector of multimedia file references accompanying this
Family group record. These could be xrefs to existing Multimedia records.}
}
\value{
An updated tidygedcom object including the Family group record.
}
\description{
This function adds a record containing details of a family group.
}
\details{
This function will automatically assign a unique xref for this record. Most users
will only need to use the husband, wife, children, and family_notes parameters (and of course gedcom).

If you need to add further information about this family (e.g. events), use the
add_family_event_*() functions.

The function will automatically add links to this family to the respective Individual
records of the wife, husband, and children.
}
