% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/individual.R
\name{add_indi}
\alias{add_indi}
\title{Add an Individual record to a tidyged object}
\usage{
add_indi(
  gedcom,
  sex = "U",
  user_reference_numbers = character(),
  indi_notes = character(),
  multimedia_links = character(),
  qn = character()
)
}
\arguments{
\item{gedcom}{A tidyged object.}

\item{sex}{The sex of the individual. Either "M" (male), "F" (female), "U" (undetermined),
"X" (intersex), or "N" (not recorded).}

\item{user_reference_numbers}{A unique user-defined number or text that the submitter
uses to identify this record. You can supply more than one in a vector. You can also define a
user reference type by using a named vector (e.g c(id1 = "123A", id2 = "456B")).}

\item{indi_notes}{A character vector of notes accompanying this Individual record.
These could be xrefs to existing Note records.}

\item{multimedia_links}{A character vector of Multimedia record xrefs accompanying this
record.}

\item{qn}{A shortcut to quickly define a name for this individual. This is a shortcut for
the add_indi_names() function (which you should really use instead), but this is useful
for quick demonstrations or tests.}
}
\value{
An updated tidyged object including the Individual record.
}
\description{
Add an Individual record to a tidyged object
}
\details{
If you need to add further information about this individual (e.g. names), use the
add_indi_* functions.
}
