% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/individual_links.R
\name{add_indi_association}
\alias{add_indi_association}
\title{Add an association with another individual}
\usage{
add_indi_association(
  gedcom,
  associated_with,
  association,
  association_notes = character(),
  xref = character(),
  update_date_changed = TRUE
)
}
\arguments{
\item{gedcom}{A tidyged object.}

\item{associated_with}{An xref identifying the associated individual.}

\item{association}{A word or phrase stating the nature of the association.}

\item{association_notes}{A character vector of notes accompanying this association.
These could be xrefs to existing Note records.}

\item{xref}{The xref of a record to act on if one is not activated (will override active record).}

\item{update_date_changed}{Whether to add/update the change date for the record.}
}
\value{
An updated tidyged object with an expanded Individual record including
this association.
}
\description{
Add an association with another individual
}
