% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/individual_facts.R
\name{add_indi_fact}
\alias{add_indi_fact}
\title{Add a fact associated with an individual}
\usage{
add_indi_fact(
  gedcom,
  type,
  descriptor = "",
  classification = character(),
  date = character(),
  age = character(),
  cause = character(),
  user_reference_type = character(),
  fact_place = place(),
  fact_address = address(),
  notes = character(),
  responsible_agency = character(),
  religious_affiliation = character(),
  adoptive_family_xref = character(),
  adopting_parent = character(),
  multimedia_links = character(),
  xref = character(),
  update_date_changed = TRUE
)
}
\arguments{
\item{gedcom}{A tidyged object.}

\item{type}{A (case-insensitive) three-letter code giving the type of event or attribute. See Details.}

\item{descriptor}{A short description of the attribute (which is not a residence) or 'other' event.}

\item{classification}{A descriptive word or phrase used to further classify this
fact. This should be used whenever the 'other' event/attribute is used (but can also be used
with others).}

\item{date}{A date_calendar(), date_approximated(), date_period(), or date_range()
object giving the timing of the fact.}

\item{age}{A character string that indicates the age in years, months, and days
that the individual was at the time of the fact. Any combination of these is permitted.
Any labels must come after their corresponding number, for example; "4y 8m 10d".}

\item{cause}{Used in special cases to record the reasons which precipitated an event.
Normally this will be used for a death event to show cause of death, such as might be listed
on a death certificate.}

\item{user_reference_type}{A user-defined type to associate with an attribute. This argument
is mandatory for type = nid and type = att.}

\item{fact_place}{A place() object giving the place associated with this fact.}

\item{fact_address}{An address() object giving the address associated with this fact.}

\item{notes}{A character vector of notes accompanying the fact. These could be xrefs to
existing Note records.}

\item{responsible_agency}{The organisation, institution, corporation, person, or other
entity that has responsibility for the fact.}

\item{religious_affiliation}{A name of the religion with which this fact was affiliated.}

\item{adoptive_family_xref}{The xref of the family that adopted this individual.
Only used for adoption events.}

\item{adopting_parent}{A code which shows which parent in the associated family adopted this
individual. Either "HUSB", "WIFE", or "BOTH".}

\item{multimedia_links}{A character vector of Multimedia record xrefs accompanying this
record.}

\item{xref}{The xref of a record to act on if one is not activated (will override active record).}

\item{update_date_changed}{Whether to add/update the change date for the record.}
}
\value{
An updated tidyged object with an expanded Individual record including
this fact.
}
\description{
Add a fact associated with an individual
}
\details{
The three-letter codes used for the type parameter are:

Attributes:
res(idence), occ(upation), edu(cation), pos(sessions), cas(te), phy(sical description),
rel(igion), cit(izenship or nationality), nob(ility title), nid (national ID number),
nur (number of relationships), nuc (number of children),

Events:
bir(th), dea(th), cen(sus), ado(ption), bap(tism), chr(istening), bur(ial),
adu(lt christening), wil(l), gra(duation), pro(bate),
ret(irement), cre(mation), bar(-mitvah), bas(-mitzvah),
emi(gration), imm(igration), con(firmation), fir(st communion), nat(uralization).

Alternatively eve or att (for any other event or attribute).

If attributes (except residence) or 'other' events/attributes are used then the descriptor
argument must be provided.
}
