% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/individual_links.R
\name{add_indi_family_link_as_spouse}
\alias{add_indi_family_link_as_spouse}
\title{Add a family link as a spouse}
\usage{
add_indi_family_link_as_spouse(
  gedcom,
  family_xref,
  linkage_notes = character(),
  xref = character(),
  update_date_changed = TRUE
)
}
\arguments{
\item{gedcom}{A tidyged object.}

\item{family_xref}{The xref of the family associated of which this individual is a spouse.}

\item{linkage_notes}{A character vector of notes accompanying this linkage.
These could be xrefs to existing Note records.}

\item{xref}{The xref of a record to act on if one is not activated (will override active record).}

\item{update_date_changed}{Whether to add/update the change date for the record.}
}
\value{
An updated tidyged object with an expanded Individual record including
this family link.
}
\description{
Add a family link as a spouse
}
\details{
These are only to be used by the add_famg function. The spouse is added separately.
}
