% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/individual_names.R
\name{add_indi_names}
\alias{add_indi_names}
\title{Add a personal name (and components) to an Individual record}
\usage{
add_indi_names(
  gedcom,
  names = name_pieces(),
  type = character(),
  xref = character(),
  update_date_changed = TRUE
)
}
\arguments{
\item{gedcom}{A tidyged object.}

\item{names}{A name_pieces() object defining the components of the personal name.}

\item{type}{The name type, e.g. "birth", "aka", "maiden".}

\item{xref}{The xref of a record to act on if one is not activated (will override active record).}

\item{update_date_changed}{Whether to add/update the change date for the record.}
}
\value{
An updated tidyged object with an expanded Individual record including
these names.
}
\description{
This function can be applied to an Individual record several times to record
personal names. The personal name is constructed automatically from the components.
}
