% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/individual_names.R
\name{add_indi_names_var}
\alias{add_indi_names_var}
\title{Add a variation of a personal name to an Individual record}
\usage{
add_indi_names_var(
  gedcom,
  primary_name,
  method,
  names_var = name_pieces(),
  phonetic_variation = TRUE,
  xref = character(),
  update_date_changed = TRUE
)
}
\arguments{
\item{gedcom}{A tidyged object.}

\item{primary_name}{The name for which this is a variation. This must be given exactly
(without forward slashes).}

\item{method}{Indicates the method used in transforming the text to the variation.}

\item{names_var}{A name_pieces() object defining the components of the name variation.}

\item{phonetic_variation}{Whether the name variation is a phonetic variation
(TRUE, default) or a romanised variation (FALSE).}

\item{xref}{The xref of a record to act on if one is not activated (will override active record).}

\item{update_date_changed}{Whether to add/update the change date for the record.}
}
\value{
An updated tidyged object with an expanded Individual record including
these name variants.
}
\description{
Add a variation of a personal name to an Individual record
}
