% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/edit_individual.R
\name{add_individual}
\alias{add_individual}
\title{Add an Individual record to a tidygedcom object}
\usage{
add_individual(
  gedcom,
  sex = character(),
  submitters = character(),
  aliases = character(),
  submitters_interested_in_ancestors = character(),
  submitters_interested_in_descendants = character(),
  permanent_record_file_number = character(),
  ancestral_file_number = character(),
  user_reference_number = character(),
  user_reference_type = character(),
  automated_record_id = character(),
  restriction_notice = character(),
  individual_notes = character()
)
}
\arguments{
\item{gedcom}{A tidygedcom object.}

\item{sex}{The sex of the individual. Either "M" (male), "F" (female), or "U" (undetermined).}

\item{submitters}{A character vector of submitters of this record. A submitter can either be
referenced by an xref or by a regular expression to match to a submitter name.}

\item{aliases}{A character vector of other Individual records that are aliases of this
individual. An individual can either be referenced by an xref or by a regular expression
to match to an individual name.}

\item{submitters_interested_in_ancestors}{A character vector of submitters interested in
ancestors of this individual. A submitter can either be referenced by an xref or by a
regular expression to match to a submitter name.}

\item{submitters_interested_in_descendants}{A character vector of submitters interested in
descendants of this individual. A submitter can either be referenced by an xref or by a
regular expression to match to a submitter name.}

\item{permanent_record_file_number}{The record number that uniquely identifies this record
within a registered network resource. See the Gedcom 5.5.1 Standard for more details.}

\item{ancestral_file_number}{A unique permanent record number of an individual record
contained in the Family History Department's Ancestral File.}

\item{user_reference_number}{A user-defined number or text that the submitter uses to identify
this record. See the Gedcom 5.5.1 Standard for more details.}

\item{user_reference_type}{A user-defined definition of the user_reference_number.}

\item{automated_record_id}{A unique record identification number assigned to the record by
the source system.}

\item{restriction_notice}{Only for Ancestral File usage. See the Gedcom 5.5.1 Standard for more
details.}

\item{individual_notes}{A character vector of notes accompanying this Individual record.}
}
\value{
An updated tidygedcom object including the Individual record.
}
\description{
Add an Individual record to a tidygedcom object
}
\details{
This function will automatically assign a unique xref for this record. Most users
will only need to use the sex, submitters, and individual_notes parameters.

If you need to add further information about this individual (e.g. names), use the
add_individual_* functions.
}
\seealso{
\code{\link[=add_individual_names]{add_individual_names()}}, \code{\link[=add_individual_names_var]{add_individual_names_var()}},
\code{\link[=add_individual_event]{add_individual_event()}}, \code{\link[=add_individual_attribute]{add_individual_attribute()}},
\code{\link[=add_individual_association]{add_individual_association()}}, \code{\link[=add_individual_family_link_as_spouse]{add_individual_family_link_as_spouse()}},
\code{\link[=import_gedcom]{import_gedcom()}}
}
