% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/edit_individual_attributes.R
\name{add_individual_attribute}
\alias{add_individual_attribute}
\alias{add_individual_attribute_caste}
\alias{add_individual_attribute_phys_descr}
\alias{add_individual_attribute_education}
\alias{add_individual_attribute_national_id}
\alias{add_individual_attribute_nationality}
\alias{add_individual_attribute_num_children}
\alias{add_individual_attribute_num_marriages}
\alias{add_individual_attribute_occupation}
\alias{add_individual_attribute_possessions}
\alias{add_individual_attribute_religion}
\alias{add_individual_attribute_residence}
\alias{add_individual_attribute_social_sec_num}
\alias{add_individual_attribute_nobility_title}
\alias{add_individual_attribute_other}
\title{Title}
\usage{
add_individual_attribute(
  gedcom,
  attribute_type,
  attribute_descriptor,
  fact_classification = character(),
  event_date = date_value(),
  age_at_event = character(),
  event_notes = character(),
  place_name = character(),
  place_hierarchy = character(),
  place_phonetic_variation = character(),
  phonetic_type = character(),
  place_romanized_variation = character(),
  romanized_type = character(),
  place_latitude = character(),
  place_longitude = character(),
  place_notes = character(),
  address_first_line = character(),
  city = character(),
  state = character(),
  postal_code = character(),
  country = character(),
  phone_number = character(),
  email = character(),
  fax = character(),
  web_page = character(),
  responsible_agency = character(),
  religious_affiliation = character(),
  cause_of_event = character(),
  restriction_notice = character()
)

add_individual_attribute_caste(
  gedcom,
  attribute_type = "CAST",
  attribute_descriptor,
  fact_classification = character(),
  event_date = date_value(),
  age_at_event = character(),
  event_notes = character(),
  place_name = character(),
  place_hierarchy = character(),
  place_phonetic_variation = character(),
  phonetic_type = character(),
  place_romanized_variation = character(),
  romanized_type = character(),
  place_latitude = character(),
  place_longitude = character(),
  place_notes = character(),
  address_first_line = character(),
  city = character(),
  state = character(),
  postal_code = character(),
  country = character(),
  phone_number = character(),
  email = character(),
  fax = character(),
  web_page = character(),
  responsible_agency = character(),
  religious_affiliation = character(),
  cause_of_event = character(),
  restriction_notice = character()
)

add_individual_attribute_phys_descr(
  gedcom,
  attribute_type = "DSCR",
  attribute_descriptor,
  fact_classification = character(),
  event_date = date_value(),
  age_at_event = character(),
  event_notes = character(),
  place_name = character(),
  place_hierarchy = character(),
  place_phonetic_variation = character(),
  phonetic_type = character(),
  place_romanized_variation = character(),
  romanized_type = character(),
  place_latitude = character(),
  place_longitude = character(),
  place_notes = character(),
  address_first_line = character(),
  city = character(),
  state = character(),
  postal_code = character(),
  country = character(),
  phone_number = character(),
  email = character(),
  fax = character(),
  web_page = character(),
  responsible_agency = character(),
  religious_affiliation = character(),
  cause_of_event = character(),
  restriction_notice = character()
)

add_individual_attribute_education(
  gedcom,
  attribute_type = "EDUC",
  attribute_descriptor,
  fact_classification = character(),
  event_date = date_value(),
  age_at_event = character(),
  event_notes = character(),
  place_name = character(),
  place_hierarchy = character(),
  place_phonetic_variation = character(),
  phonetic_type = character(),
  place_romanized_variation = character(),
  romanized_type = character(),
  place_latitude = character(),
  place_longitude = character(),
  place_notes = character(),
  address_first_line = character(),
  city = character(),
  state = character(),
  postal_code = character(),
  country = character(),
  phone_number = character(),
  email = character(),
  fax = character(),
  web_page = character(),
  responsible_agency = character(),
  religious_affiliation = character(),
  cause_of_event = character(),
  restriction_notice = character()
)

add_individual_attribute_national_id(
  gedcom,
  attribute_type = "IDNO",
  attribute_descriptor,
  fact_classification = character(),
  event_date = date_value(),
  age_at_event = character(),
  event_notes = character(),
  place_name = character(),
  place_hierarchy = character(),
  place_phonetic_variation = character(),
  phonetic_type = character(),
  place_romanized_variation = character(),
  romanized_type = character(),
  place_latitude = character(),
  place_longitude = character(),
  place_notes = character(),
  address_first_line = character(),
  city = character(),
  state = character(),
  postal_code = character(),
  country = character(),
  phone_number = character(),
  email = character(),
  fax = character(),
  web_page = character(),
  responsible_agency = character(),
  religious_affiliation = character(),
  cause_of_event = character(),
  restriction_notice = character()
)

add_individual_attribute_nationality(
  gedcom,
  attribute_type = "NATI",
  attribute_descriptor,
  fact_classification = character(),
  event_date = date_value(),
  age_at_event = character(),
  event_notes = character(),
  place_name = character(),
  place_hierarchy = character(),
  place_phonetic_variation = character(),
  phonetic_type = character(),
  place_romanized_variation = character(),
  romanized_type = character(),
  place_latitude = character(),
  place_longitude = character(),
  place_notes = character(),
  address_first_line = character(),
  city = character(),
  state = character(),
  postal_code = character(),
  country = character(),
  phone_number = character(),
  email = character(),
  fax = character(),
  web_page = character(),
  responsible_agency = character(),
  religious_affiliation = character(),
  cause_of_event = character(),
  restriction_notice = character()
)

add_individual_attribute_num_children(
  gedcom,
  attribute_type = "NCHI",
  attribute_descriptor,
  fact_classification = character(),
  event_date = date_value(),
  age_at_event = character(),
  event_notes = character(),
  place_name = character(),
  place_hierarchy = character(),
  place_phonetic_variation = character(),
  phonetic_type = character(),
  place_romanized_variation = character(),
  romanized_type = character(),
  place_latitude = character(),
  place_longitude = character(),
  place_notes = character(),
  address_first_line = character(),
  city = character(),
  state = character(),
  postal_code = character(),
  country = character(),
  phone_number = character(),
  email = character(),
  fax = character(),
  web_page = character(),
  responsible_agency = character(),
  religious_affiliation = character(),
  cause_of_event = character(),
  restriction_notice = character()
)

add_individual_attribute_num_marriages(
  gedcom,
  attribute_type = "NMR",
  attribute_descriptor,
  fact_classification = character(),
  event_date = date_value(),
  age_at_event = character(),
  event_notes = character(),
  place_name = character(),
  place_hierarchy = character(),
  place_phonetic_variation = character(),
  phonetic_type = character(),
  place_romanized_variation = character(),
  romanized_type = character(),
  place_latitude = character(),
  place_longitude = character(),
  place_notes = character(),
  address_first_line = character(),
  city = character(),
  state = character(),
  postal_code = character(),
  country = character(),
  phone_number = character(),
  email = character(),
  fax = character(),
  web_page = character(),
  responsible_agency = character(),
  religious_affiliation = character(),
  cause_of_event = character(),
  restriction_notice = character()
)

add_individual_attribute_occupation(
  gedcom,
  attribute_type = "OCCU",
  attribute_descriptor,
  fact_classification = character(),
  event_date = date_value(),
  age_at_event = character(),
  event_notes = character(),
  place_name = character(),
  place_hierarchy = character(),
  place_phonetic_variation = character(),
  phonetic_type = character(),
  place_romanized_variation = character(),
  romanized_type = character(),
  place_latitude = character(),
  place_longitude = character(),
  place_notes = character(),
  address_first_line = character(),
  city = character(),
  state = character(),
  postal_code = character(),
  country = character(),
  phone_number = character(),
  email = character(),
  fax = character(),
  web_page = character(),
  responsible_agency = character(),
  religious_affiliation = character(),
  cause_of_event = character(),
  restriction_notice = character()
)

add_individual_attribute_possessions(
  gedcom,
  attribute_type = "PROP",
  attribute_descriptor,
  fact_classification = character(),
  event_date = date_value(),
  age_at_event = character(),
  event_notes = character(),
  place_name = character(),
  place_hierarchy = character(),
  place_phonetic_variation = character(),
  phonetic_type = character(),
  place_romanized_variation = character(),
  romanized_type = character(),
  place_latitude = character(),
  place_longitude = character(),
  place_notes = character(),
  address_first_line = character(),
  city = character(),
  state = character(),
  postal_code = character(),
  country = character(),
  phone_number = character(),
  email = character(),
  fax = character(),
  web_page = character(),
  responsible_agency = character(),
  religious_affiliation = character(),
  cause_of_event = character(),
  restriction_notice = character()
)

add_individual_attribute_religion(
  gedcom,
  attribute_type = "RELI",
  attribute_descriptor,
  fact_classification = character(),
  event_date = date_value(),
  age_at_event = character(),
  event_notes = character(),
  place_name = character(),
  place_hierarchy = character(),
  place_phonetic_variation = character(),
  phonetic_type = character(),
  place_romanized_variation = character(),
  romanized_type = character(),
  place_latitude = character(),
  place_longitude = character(),
  place_notes = character(),
  address_first_line = character(),
  city = character(),
  state = character(),
  postal_code = character(),
  country = character(),
  phone_number = character(),
  email = character(),
  fax = character(),
  web_page = character(),
  responsible_agency = character(),
  religious_affiliation = character(),
  cause_of_event = character(),
  restriction_notice = character()
)

add_individual_attribute_residence(
  gedcom,
  attribute_type = "RESI",
  attribute_descriptor,
  fact_classification = character(),
  event_date = date_value(),
  age_at_event = character(),
  event_notes = character(),
  place_name = character(),
  place_hierarchy = character(),
  place_phonetic_variation = character(),
  phonetic_type = character(),
  place_romanized_variation = character(),
  romanized_type = character(),
  place_latitude = character(),
  place_longitude = character(),
  place_notes = character(),
  address_first_line = character(),
  city = character(),
  state = character(),
  postal_code = character(),
  country = character(),
  phone_number = character(),
  email = character(),
  fax = character(),
  web_page = character(),
  responsible_agency = character(),
  religious_affiliation = character(),
  cause_of_event = character(),
  restriction_notice = character()
)

add_individual_attribute_social_sec_num(
  gedcom,
  attribute_type = "SSN",
  attribute_descriptor,
  fact_classification = character(),
  event_date = date_value(),
  age_at_event = character(),
  event_notes = character(),
  place_name = character(),
  place_hierarchy = character(),
  place_phonetic_variation = character(),
  phonetic_type = character(),
  place_romanized_variation = character(),
  romanized_type = character(),
  place_latitude = character(),
  place_longitude = character(),
  place_notes = character(),
  address_first_line = character(),
  city = character(),
  state = character(),
  postal_code = character(),
  country = character(),
  phone_number = character(),
  email = character(),
  fax = character(),
  web_page = character(),
  responsible_agency = character(),
  religious_affiliation = character(),
  cause_of_event = character(),
  restriction_notice = character()
)

add_individual_attribute_nobility_title(
  gedcom,
  attribute_type = "TITL",
  attribute_descriptor,
  fact_classification = character(),
  event_date = date_value(),
  age_at_event = character(),
  event_notes = character(),
  place_name = character(),
  place_hierarchy = character(),
  place_phonetic_variation = character(),
  phonetic_type = character(),
  place_romanized_variation = character(),
  romanized_type = character(),
  place_latitude = character(),
  place_longitude = character(),
  place_notes = character(),
  address_first_line = character(),
  city = character(),
  state = character(),
  postal_code = character(),
  country = character(),
  phone_number = character(),
  email = character(),
  fax = character(),
  web_page = character(),
  responsible_agency = character(),
  religious_affiliation = character(),
  cause_of_event = character(),
  restriction_notice = character()
)

add_individual_attribute_other(
  gedcom,
  attribute_type = "FACT",
  attribute_descriptor,
  fact_classification = character(),
  event_date = date_value(),
  age_at_event = character(),
  event_notes = character(),
  place_name = character(),
  place_hierarchy = character(),
  place_phonetic_variation = character(),
  phonetic_type = character(),
  place_romanized_variation = character(),
  romanized_type = character(),
  place_latitude = character(),
  place_longitude = character(),
  place_notes = character(),
  address_first_line = character(),
  city = character(),
  state = character(),
  postal_code = character(),
  country = character(),
  phone_number = character(),
  email = character(),
  fax = character(),
  web_page = character(),
  responsible_agency = character(),
  religious_affiliation = character(),
  cause_of_event = character(),
  restriction_notice = character()
)
}
\arguments{
\item{gedcom}{}

\item{attribute_type}{}

\item{attribute_descriptor}{}

\item{fact_classification}{}

\item{event_date}{}

\item{age_at_event}{}

\item{event_notes}{}

\item{place_name}{}

\item{place_hierarchy}{}

\item{place_phonetic_variation}{}

\item{phonetic_type}{}

\item{place_romanized_variation}{}

\item{romanized_type}{}

\item{place_latitude}{}

\item{place_longitude}{}

\item{place_notes}{}

\item{address_first_line}{}

\item{city}{}

\item{state}{}

\item{postal_code}{}

\item{country}{}

\item{phone_number}{}

\item{email}{}

\item{fax}{}

\item{web_page}{}

\item{responsible_agency}{}

\item{religious_affiliation}{}

\item{cause_of_event}{}

\item{restriction_notice}{}
}
\value{

}
\description{
Title
}
