% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/edit_individual_links.R
\name{add_individual_family_link_as_child}
\alias{add_individual_family_link_as_child}
\title{Add a family link as a child}
\usage{
add_individual_family_link_as_child(
  gedcom,
  family_xref,
  linkage_type = "birth",
  linkage_notes = character()
)
}
\arguments{
\item{gedcom}{A tidygedcom object.}

\item{family_xref}{The xref of the family associated of which this individual is a child.}

\item{linkage_type}{A code used to indicate the child to family relationship. Either,
"birth" (default), "foster", or "adopted".}

\item{linkage_notes}{A character vector of notes accompanying this linkage.
These could be xrefs to existing Note records.}
}
\value{
An updated tidygedcom object with an expanded Individual record including
this family link.
}
\description{
Add a family link as a child
}
