% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{add_levels}
\alias{add_levels}
\title{Push a tidygedcom structure down a number of levels}
\usage{
add_levels(df, start_level)
}
\arguments{
\item{df}{A tidygedcom structure.}

\item{start_level}{How many levels to add on to the existing levels.}
}
\value{
The tidygedcom structure with modified levels.
}
\description{
Push a tidygedcom structure down a number of levels
}
