% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/edit_multimedia.R
\name{add_multimedia}
\alias{add_multimedia}
\title{Add a Multimedia record to a tidygedcom object}
\usage{
add_multimedia(
  gedcom,
  file_reference,
  format,
  source_media = character(),
  title = character(),
  user_reference_number = character(),
  user_reference_type = character(),
  automated_record_id = character(),
  multimedia_notes = character()
)
}
\arguments{
\item{gedcom}{A tidygedcom object.}

\item{file_reference}{A character vector of references for the files, typically a filepath or URL.}

\item{format}{A character vector indicating the format of each multimedia file.
Currently limited to one of: bmp, gif, jpg, ole, pcx, tif, wav.
A format must be defined for each file reference given.}

\item{source_media}{A code that indicates the type of material in which the referenced
source is stored. Must be one of: audio, book, card, electronic, fiche, film, magazine,
manuscript, map, newspaper, photo, tombstone, video. If this is defined, it must be a character
vector the same size as file_reference and format.}

\item{title}{The title of the multimedia file(s). If this is defined, it must be a character
vector the same size as file_reference and format.}

\item{user_reference_number}{A user-defined number or text that the submitter uses to identify
this record. See the Gedcom 5.5.5 Specification for more details.}

\item{user_reference_type}{A user-defined definition of the user_reference_number.}

\item{automated_record_id}{A unique record identification number assigned to the record by
the source system.}

\item{multimedia_notes}{A character vector of notes accompanying this Multimedia record. These could be
xrefs to existing Note records.}
}
\value{
An updated tidygedcom object including the Multimedia record.
}
\description{
Add a Multimedia record to a tidygedcom object
}
\details{
This function will automatically assign a unique xref for this record. Most users
will only need to use the file_reference, format, source_media, title, and
multimedia_notes parameters (and of course gedcom).
}
