% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrappers.R
\name{add_parents}
\alias{add_parents}
\title{Add parent records for an individual}
\usage{
add_parents(gedcom, xref = character(), inc_sex = TRUE)
}
\arguments{
\item{gedcom}{A tidyged object.}

\item{xref}{The xref of an Individual record to act on if one is not activated (will override active record).}

\item{inc_sex}{Whether to populate the sex of the parents. This will ensure
that there is one male and one female parent. Otherwise the sex will be
assigned as "U" (undetermined).}
}
\value{
A tidyged object with additional parent records.
}
\description{
This function adds placeholder records for an individual's parents.
}
\details{
This function may also create a Family Group record and will
not modify existing parents.
}
