% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/repository.R
\name{add_repo}
\alias{add_repo}
\title{Add a Repository record to a tidyged object}
\usage{
add_repo(
  gedcom,
  name,
  repo_address = address(),
  user_reference_numbers = character(),
  repo_notes = character()
)
}
\arguments{
\item{gedcom}{A tidyged object.}

\item{name}{The name of the repository.}

\item{repo_address}{An address() object giving the address of the repository.}

\item{user_reference_numbers}{A unique user-defined number or text that the submitter
uses to identify this record. You can supply more than one in a vector. You can also define a
user reference type by using a named vector (e.g c(id1 = "123A", id2 = "456B")).}

\item{repo_notes}{A character vector of notes accompanying this Repository record.
These could be xrefs to existing Note records.}
}
\value{
An updated tidyged object including the Repository record.
}
\description{
Add a Repository record to a tidyged object
}
