% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/edit_repository.R
\name{add_repository}
\alias{add_repository}
\title{Add a Repository record to a tidygedcom object}
\usage{
add_repository(
  gedcom,
  name,
  local_address_lines = character(),
  city = character(),
  state = character(),
  postal_code = character(),
  country = character(),
  phone_number = character(),
  email = character(),
  fax = character(),
  web_page = character(),
  user_reference_number = character(),
  user_reference_type = character(),
  automated_record_id = character(),
  repository_notes = character()
)
}
\arguments{
\item{gedcom}{A tidygedcom object.}

\item{name}{The name of the repository.}

\item{local_address_lines}{The first lines of the repository address.}

\item{city}{The city of the repository.}

\item{state}{The state/county of the repository.}

\item{postal_code}{The postal code of the repository.}

\item{country}{The country of the repository.}

\item{phone_number}{A character vector containing up to three phone numbers of the repository.}

\item{email}{A character vector containing up to three email addresses of the repository.}

\item{fax}{A character vector containing up to three fax numbers of the repository.}

\item{web_page}{A character vector containing up to three web pages of the repository.}

\item{user_reference_number}{A user-defined number or text that the submitter uses to identify
this record. See the Gedcom 5.5.5 Specification for more details.}

\item{user_reference_type}{A user-defined definition of the user_reference_number.}

\item{automated_record_id}{A unique record identification number assigned to the record by
the source system.}

\item{repository_notes}{A character vector of notes accompanying this Repository record.
These could be xrefs to existing Note records.}
}
\value{
An updated tidygedcom object including the Repository record.
}
\description{
Add a Repository record to a tidygedcom object
}
\details{
This function will automatically assign a unique xref for this record. Most users
will only need to use the sex, submitters, and individual_notes parameters (and of course gedcom).

If you need to add further information about this individual (e.g. names), use the
add_individual_* functions.
}
