% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/source.R
\name{add_sour_repo_citation}
\alias{add_sour_repo_citation}
\title{Add a source repository citation to a Source record}
\usage{
add_sour_repo_citation(
  gedcom,
  repository,
  call_number = character(),
  xref = character(),
  update_date_changed = TRUE
)
}
\arguments{
\item{gedcom}{A tidyged object.}

\item{repository}{A character string identifying the repository xref.}

\item{call_number}{An identification or reference description used to file
and retrieve items from the holdings of a repository.}

\item{xref}{The xref of a record to act on if one is not activated (will override active record).}

\item{update_date_changed}{Whether to add/update the change date for the record.}
}
\value{
An updated tidyged object with an expanded Source record including
this repository citation.
}
\description{
This structure is used within a source record to point to a name and
address record of the holder of the source document.
}
