% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/edit_source.R
\name{add_source}
\alias{add_source}
\title{Add a Source record to a tidyged object}
\usage{
add_source(
  gedcom,
  events_recorded = character(),
  date_period_covered = date_period(),
  jurisdiction = character(),
  responsible_agency = character(),
  originator = character(),
  title = character(),
  short_title = character(),
  publication_detail = character(),
  source_text = character(),
  user_reference_number = character(),
  user_reference_type = character(),
  automated_record_id = character(),
  data_notes = character(),
  source_notes = character(),
  multimedia_links = character()
)
}
\arguments{
\item{gedcom}{A tidyged object.}

\item{events_recorded}{An enumeration of the different kinds of events that were recorded
in this source. Each enumeration is separated by a comma. See the Gedcom 5.5.5 Specification for
more details.}

\item{date_period_covered}{A date_period() object indicating the time period of events
covered by this source.}

\item{jurisdiction}{A character string indicating the lowest level jurisdiction encompassing
all places named in this source. See the Gedcom 5.5.5 Specification for more details.}

\item{responsible_agency}{The organisation, institution, corporation, person, or other
entity that has responsibility for the source data.}

\item{originator}{The person, agency, or entity who created the record.}

\item{title}{The title of the source work, record, or item and, when appropriate, the title of the
larger work or series of which it is a part.}

\item{short_title}{An abbreviated or shortened version of the title (if required).}

\item{publication_detail}{When and where the source record was created. For published works,
this includes information such as the city of publication, name of the publisher, and year of
publication.}

\item{source_text}{A verbatim copy of relevant text contained within the source.
This indicates notes or text that are actually contained in the source document,
not the submitter's opinion about the source.}

\item{user_reference_number}{A user-defined number or text that the submitter uses to identify
this record. See the Gedcom 5.5.5 Specification for more details.}

\item{user_reference_type}{A user-defined definition of the user_reference_number.}

\item{automated_record_id}{A unique record identification number assigned to the record by
the source system.}

\item{data_notes}{A character vector of notes associated with the data in this Source record.
These could be xrefs to existing Note records.}

\item{source_notes}{A character vector of notes accompanying this Source record.
These could be xrefs to existing Note records.}

\item{multimedia_links}{A character vector of multimedia file references accompanying this
source. These could be xrefs to existing Multimedia records.}
}
\value{
An updated tidyged object including the Source record.
}
\description{
Add a Source record to a tidyged object
}
\details{
This function will automatically assign a unique xref for this record. Most users
will only need to use the originator, title, publication_detail, data_notes,
and source_notes parameters (and of course gedcom).

The function will automatically split the originator, title, publication_detail, source_text,
data_notes and source_notes onto separate lines if the character limit in the Gedcom
standard is exceeded.
}
