% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrappers.R
\name{add_spouse}
\alias{add_spouse}
\title{Add a spouse for an individual}
\usage{
add_spouse(gedcom, xref = character(), sex = "U")
}
\arguments{
\item{gedcom}{A tidyged object.}

\item{xref}{The xref of an Individual record to act on if one is not activated (will override active record).}

\item{sex}{The sex of the spouse.}
}
\value{
A tidyged object with additional spouse and Family Group records.
}
\description{
This creates a record for a spouse and their Family Group record.
}
