% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/edit_submission.R
\name{add_submission}
\alias{add_submission}
\title{Add a Submission record to a tidygedcom object}
\usage{
add_submission(
  gedcom,
  name_of_family_file = character(),
  temple_code = character(),
  generations_of_ancestors = character(),
  generations_of_descendants = character(),
  ordinance_process_flag = character(),
  automated_record_id = character(),
  submission_notes = character()
)
}
\arguments{
\item{gedcom}{A tidygedcom object.}

\item{name_of_family_file}{Name under which family names for ordinances are stored
in the temple's family file.}

\item{temple_code}{See the Gedcom 5.5.1 Standard for more details.}

\item{generations_of_ancestors}{The number of generations of ancestors included in this file.}

\item{generations_of_descendants}{The number of generations of descendants included in this file.}

\item{ordinance_process_flag}{See the Gedcom 5.5.1 Standard for more details.}

\item{automated_record_id}{A unique record identification number assigned to the record by
the source system.}

\item{submission_notes}{A character vector of notes accompanying this Submission record.
These could be xrefs to existing Note records.}
}
\value{
An updated tidygedcom object including the Submission record.
}
\description{
Add a Submission record to a tidygedcom object
}
\details{
The sending system uses a submission record to send instructions and
information to the receiving system. Each GEDCOM transmission file should have only one
submission record.

This function will automatically assign a unique xref for this record.

The function will automatically split the submission_notes onto separate lines if the
character limit in the Gedcom standard is exceeded.
}
