% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/edit_submitter.R
\name{add_submitter}
\alias{add_submitter}
\title{Add a Submitter record to a tidygedcom object}
\usage{
add_submitter(
  gedcom,
  name,
  address_first_line = character(),
  city = character(),
  state = character(),
  postal_code = character(),
  country = character(),
  phone_number = character(),
  email = character(),
  fax = character(),
  web_page = character(),
  language_preference = character(),
  submitter_registered_rfn = character(),
  automated_record_id = character(),
  submitter_notes = character()
)
}
\arguments{
\item{gedcom}{A tidygedcom object.}

\item{name}{The name of the submitter.}

\item{address_first_line}{The first line of the submitter address.}

\item{city}{The city of the submitter.}

\item{state}{The state/county of the submitter.}

\item{postal_code}{The postal code of the submitter.}

\item{country}{The country of the submitter.}

\item{phone_number}{A character vector containing up to three phone numbers of the submitter.}

\item{email}{A character vector containing up to three email addresses of the submitter.}

\item{fax}{A character vector containing up to three fax numbers of the submitter.}

\item{web_page}{A character vector containing up to three web pages of the submitter.}

\item{language_preference}{The language in which the submitter prefers to communicate.}

\item{submitter_registered_rfn}{See the Gedcom 5.5.1 Standard for more details.}

\item{automated_record_id}{A unique record identification number assigned to the record by
the source system.}

\item{submitter_notes}{A character vector of notes accompanying this Submitter record.
These could be xrefs to existing Note records.}
}
\value{
An updated tidygedcom object including the Submitter record.
}
\description{
Add a Submitter record to a tidygedcom object
}
\details{
The submitter record identifies an individual or organization that contributed
information contained in the GEDCOM transmission.

This function will automatically assign a unique xref for this record.

The function will automatically split the submitter_notes onto separate lines if the
character limit in the Gedcom standard is exceeded.
}
