% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/structures.R
\name{address}
\alias{address}
\title{Define an address}
\usage{
address(
  local_address_lines = character(),
  city = character(),
  state = character(),
  postal_code = character(),
  country = character(),
  phone_number = character(),
  email = character(),
  fax = character(),
  web_page = character()
)
}
\arguments{
\item{local_address_lines}{A character vector containing up to three local address lines.}

\item{city}{The city of the address.}

\item{state}{The state/county of the address.}

\item{postal_code}{The postal code of the address.}

\item{country}{The country of the address.}

\item{phone_number}{A character vector containing up to three phone numbers.}

\item{email}{A character vector containing up to three email addresses.}

\item{fax}{A character vector containing up to three fax numbers.}

\item{web_page}{A character vector containing up to three web pages.}
}
\value{
A tibble describing an address.
}
\description{
Define an address
}
