% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subrecords.R
\name{ADDRESS_STRUCTURE}
\alias{ADDRESS_STRUCTURE}
\title{Construct the ADDRESS_STRUCTURE tibble}
\usage{
ADDRESS_STRUCTURE(
  local_address_lines = character(),
  address_city = character(),
  address_state = character(),
  address_postal_code = character(),
  address_country = character(),
  phone_number = character(),
  address_email = character(),
  address_fax = character(),
  address_web_page = character()
)
}
\arguments{
\item{local_address_lines}{These address lines contain the address lines before the town/city.
This can be a vector with up to 3 elements.}

\item{address_city}{The name of the city/town used in the address. Isolated for sorting or indexing.}

\item{address_state}{The name of the US state/UK county used in the address. Isolated for sorting or indexing.}

\item{address_postal_code}{The ZIP or postal code used by the various localities in handling of mail.
Isolated for sorting or indexing.}

\item{address_country}{The name of the country that pertains to the associated address. Isolated by some systems for
sorting or indexing. Used in most cases to facilitate automatic sorting of mail.}

\item{phone_number}{A phone number.}

\item{address_email}{An electronic address that can be used for contact such as an email address.}

\item{address_fax}{A FAX telephone number appropriate for sending data facsimiles.}

\item{address_web_page}{The world wide web page address.}
}
\value{
A tidy tibble containing the ADDRESS_STRUCTURE part of a GEDCOM file.
}
\description{
This function constructs a tibble representation of the ADDRESS_STRUCTURE from the GEDCOM 5.5.5
specification.
}
