% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{assign_xref}
\alias{assign_xref}
\title{Create a new xref for a record}
\usage{
assign_xref(type = "", ref = 0, gedcom = tibble::tibble())
}
\arguments{
\item{type}{An alphabetic sequence to be used as a prefix for the xref identifier.
GEDCOM files traditionally use a single letter to denote the type of record, e.g.
I for Individual, F for Family group, etc.}

\item{ref}{An explicit reference number after the type if one is to be chosen manually.}

\item{gedcom}{A tidyged object.}
}
\value{
An xref to use for a new record.
}
\description{
This function is used to assign xrefs to new records that are created.
}
