% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/structures.R
\name{association_structure}
\alias{association_structure}
\title{Constructs the ASSOCIATION_STRUCTURE from the GEDCOM specification}
\usage{
association_structure(
  xref_indi,
  relation_is_descriptor,
  source_citations = list(),
  notes = list()
)
}
\arguments{
\item{xref_indi}{Either a pointer or an unique cross-reference identifier.}

\item{relation_is_descriptor}{A word or phrase that states object 1's relation is object 2.}
}
\value{

}
\description{
Constructs the ASSOCIATION_STRUCTURE from the GEDCOM specification
}
