% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_write_create.R
\name{check_line_lengths}
\alias{check_line_lengths}
\title{Check the line lengths of a GEDCOM file}
\usage{
check_line_lengths(lines, limit)
}
\arguments{
\item{lines}{A character vector of GEDCOM lines.}

\item{limit}{The maximum length of a line allowed.}
}
\value{
The input character vector
}
\description{
Check the line lengths of a GEDCOM file
}
