% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/structures.R
\name{CHILD_TO_FAMILY_LINK}
\alias{CHILD_TO_FAMILY_LINK}
\title{Construct the CHILD_TO_FAMILY_LINK tibble}
\usage{
CHILD_TO_FAMILY_LINK(
  xref_fam,
  pedigree_linkage_type = character(),
  child_linkage_status = character(),
  notes = list()
)
}
\arguments{
\item{xref_fam}{An xref ID of a family record.}

\item{pedigree_linkage_type}{A code used to indicate the child to family relationship for pedigree navigation
purposes.}

\item{child_linkage_status}{A status code that allows passing on the users opinion of the
status of a child to family link.
"challenged": Linking this child to this family is suspect, but the linkage has been
neither proven nor disproven.
"disproven": There has been a claim by some that this child belongs to this family, but
the linkage has been disproven.
"proven" = There has been a claim by some that this child does not belongs to this family,
but the linkage has been proven.}

\item{notes}{A list of NOTE_STRUCTURE() objects.}
}
\value{
A tidy tibble containing the CHILD_TO_FAMILY_LINK part of a GEDCOM file.
}
\description{
This function constructs a tibble representation of the CHILD_TO_FAMILY_LINK from the GEDCOM 5.5.1
specification.
}
