% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_write_create.R
\name{combine_gedcom_values}
\alias{combine_gedcom_values}
\title{Convert the GEDCOM grammar to the GEDCOM form}
\usage{
combine_gedcom_values(gedcom)
}
\arguments{
\item{gedcom}{A tidygedcom object.}
}
\value{
A tidygedcom object in the GEDCOM form.
}
\description{
This function applies concatenation indicated by CONC/CONT lines.
}
\details{
The function works by collapsing CONC/CONT lines using group-by/summarise.
}
