% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/edit_note.R
\name{consolidate_notes}
\alias{consolidate_notes}
\title{Consolidate duplicated notes}
\usage{
consolidate_notes(gedcom, min_occurences = 2)
}
\arguments{
\item{gedcom}{A tidygedcom object.}

\item{min_occurences}{How many duplicates to prompt creating a new Note record.}
}
\value{
A tidygedcom object with all notes consolidated.
}
\description{
Consolidate duplicated notes
}
