% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/globals.R
\name{date_approximated}
\alias{date_approximated}
\title{Construct a DATE_APPROXIMATED string}
\description{
See \code{tidyged.internals::\link[tidyged.internals:date_approximated]{date_approximated}} for details.
}
