% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers_dates.R
\name{date_exact}
\alias{date_exact}
\title{Construct a DATE_EXACT string}
\usage{
date_exact(day = numeric(), month = numeric(), year = numeric())
}
\arguments{
\item{day}{The day of the month.}

\item{month}{The month of the year.}

\item{year}{The year.}
}
\value{
A DATE_EXACT string
}
\description{
Construct a DATE_EXACT string
}
