% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers_dates.R
\name{date_exact_pattern}
\alias{date_exact_pattern}
\title{Construct the regex pattern for DATE_EXACT values}
\usage{
date_exact_pattern()
}
\value{
A regex string
}
\description{
Construct the regex pattern for DATE_EXACT values
}
