% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dates.R
\name{date_period}
\alias{date_period}
\title{Construct a DATE_PERIOD string}
\usage{
date_period(
  start_year = numeric(),
  start_month = numeric(),
  start_day = numeric(),
  end_year = numeric(),
  end_month = numeric(),
  end_day = numeric(),
  to = FALSE
)
}
\arguments{
\item{start_year}{The start year of the period. This value is required.}

\item{start_month}{The start month of the period.}

\item{start_day}{The start day of the period.}

\item{end_year}{The end year of the period}

\item{end_month}{The end month of the period.}

\item{end_day}{The end day of the period.}

\item{to}{Whether the date given by start_year/month/day should be considered the end of a period.}
}
\value{
A DATE_PERIOD string
}
\description{
Construct a DATE_PERIOD string
}
