% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dates.R
\name{date_value}
\alias{date_value}
\title{Construct a DATE_VALUE string}
\usage{
date_value(
  start_year = numeric(),
  start_month = numeric(),
  start_day = numeric(),
  end_year = numeric(),
  end_month = numeric(),
  end_day = numeric(),
  from = FALSE,
  to = FALSE,
  before = FALSE,
  after = FALSE,
  between = FALSE,
  about = FALSE,
  calc = FALSE,
  est = FALSE
)
}
\arguments{
\item{start_year}{The year (or start year if a date range/period is provided). This value is required.}

\item{start_month}{The month of the year (or start month if a date range/period is provided)}

\item{start_day}{The day of the month (or start day if a date range/period is provided)}

\item{end_year}{The end year}

\item{end_month}{The end month of the year}

\item{end_day}{The end day of the month}

\item{from}{Whether the date given by start_year/month/day should be considered the beginning of a period.
If 'to' is FALSE then the period is semi-infinite.}

\item{to}{Whether the date given by end_year/month/day should be considered the end of a period.
If 'from' is FALSE then the period is semi-infinite.}

\item{before}{Whether the date given by start_year/month/day should be considered the end of a range.}

\item{after}{Whether the date given by start_year/month/day should be considered the beginning of a range.}

\item{between}{Whether the start and end dates should be considered as the bounds of a date range.}

\item{about}{Whether the date given by start_year/month/day is approximate.}

\item{calc}{Whether the date given by start_year/month/day is calculated from other values.}

\item{est}{Whether the date given by start_year/month/day is estimated.}
}
\value{
A DATE_VALUE string
}
\description{
Construct a DATE_VALUE string
}
