% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regex_patterns.R
\name{date_value_pattern}
\alias{date_value_pattern}
\title{Construct the regex pattern for DATE_VALUE values}
\usage{
date_value_pattern()
}
\value{
Either a single regex string or a vector of them
}
\description{
Construct the regex pattern for DATE_VALUE values
}
