% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/describe.R
\name{describe_famg}
\alias{describe_famg}
\alias{describe_indi}
\alias{describe_media}
\alias{describe_sour}
\alias{describe_repo}
\alias{describe_note}
\alias{describe_subm}
\title{Get a description of a record}
\usage{
describe_famg(gedcom, xref, short_desc = FALSE)

describe_indi(gedcom, xref, name_only = FALSE, short_desc = FALSE)

describe_media(gedcom, xref, file_ref_only = FALSE, short_desc = FALSE)

describe_sour(gedcom, xref, title_only = FALSE, short_desc = FALSE)

describe_repo(gedcom, xref, name_only = FALSE, short_desc = FALSE)

describe_note(gedcom, xref, short_desc = FALSE)

describe_subm(gedcom, xref, name_only = FALSE, short_desc = FALSE)
}
\arguments{
\item{gedcom}{A tidyged object.}

\item{xref}{An xref of a record.}

\item{short_desc}{Whether to return a shorter description.}

\item{name_only}{Whether to return the individual/repository name only. If none is found, the xref
is returned.}

\item{file_ref_only}{Whether to return the multimedia file reference only. If none is found, the xref
is returned.}

\item{title_only}{Whether to return the source title only. If none is found, the xref
is returned.}
}
\value{
A character string describing the record.
}
\description{
Get descriptions of a record at various degrees of detail.
}
\details{
This function offers three levels of detail. For example, individual records can be:

"Joe Bloggs" (name_only = TRUE)
"Individual @I1@, Joe Bloggs" (short_desc = TRUE)
"Individual @I1@, Joe Bloggs, child of X and Y, born on x/x/x in place, died on x/x/x in place" (short_desc = FALSE)
}
