% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interrogate.R
\name{describe_individual}
\alias{describe_individual}
\title{Get a description of an individual}
\usage{
describe_individual(gedcom, xref, abb = FALSE)
}
\arguments{
\item{gedcom}{A tidyged object.}

\item{xref}{An xref of an Individual record.}

\item{abb}{Whether to abbreviate the output.}
}
\value{
A character string describing the individual.
}
\description{
Get a description of an individual
}
