% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interrogate.R
\name{describe_records}
\alias{describe_records}
\title{Get descriptions for records}
\usage{
describe_records(gedcom, xrefs, short = FALSE)
}
\arguments{
\item{gedcom}{A tidyged object.}

\item{xrefs}{A vector of record xrefs.}

\item{short}{Whether to return a shorter description.}
}
\value{
A vector of record descriptions.
}
\description{
Get descriptions for records
}
