% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interrogate.R
\name{df_indi}
\alias{df_indi}
\alias{df_famg}
\alias{df_media}
\alias{df_sour}
\alias{df_repo}
\alias{df_note}
\title{Summarise records in a GEDCOM file}
\usage{
df_indi(gedcom)

df_famg(gedcom)

df_media(gedcom)

df_sour(gedcom)

df_repo(gedcom)

df_note(gedcom)
}
\arguments{
\item{gedcom}{A tidyged object.}
}
\value{
A tibble summarising records where every row is a record.
}
\description{
These functions give a summary of key information of individuals/families/notes etc.
in the GEDCOM file.
}
