% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarise.R
\name{df_indi_facts}
\alias{df_indi_facts}
\alias{df_famg_facts}
\title{Create a table summarising all individual/family facts}
\usage{
df_indi_facts(gedcom, xref)

df_famg_facts(gedcom, xref)
}
\arguments{
\item{gedcom}{A tidyged object.}

\item{xref}{The xref of the Individual or Family Group record.}
}
\value{
A tibble containing a row for each fact.
}
\description{
This function creates a tidy table making it easy to extract fact details for an individual or family group.
}
\details{
Notes and source citations are not included in the summary, as well as other more obscure fields.
}
