% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interrogate.R
\name{df_individuals}
\alias{df_individuals}
\alias{df_families}
\alias{df_multimedia}
\alias{df_sources}
\alias{df_repositories}
\alias{df_notes}
\title{Summarise records in a GEDCOM file}
\usage{
df_individuals(gedcom)

df_families(gedcom)

df_multimedia(gedcom)

df_sources(gedcom)

df_repositories(gedcom)

df_notes(gedcom)
}
\arguments{
\item{gedcom}{A tidygedcom object.}
}
\value{
A tibble summarising records where every row is a record.
}
\description{
These functions give a summary of key information of individuals/families/notes etc.
in the GEDCOM file.
}
