% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_export_create.R
\name{export_gedcom}
\alias{export_gedcom}
\title{Save a tidygedcom object to disk as a GEDCOM file}
\usage{
export_gedcom(gedcom, filepath)
}
\arguments{
\item{filepath}{The full filepath to write to}

\item{gedcom_df}{A tidygedcom object}
}
\value{
Nothing
}
\description{
Save a tidygedcom object to disk as a GEDCOM file
}
