% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarise.R
\name{fact_summary_indi}
\alias{fact_summary_indi}
\alias{fact_summary_famg}
\title{Create a table summarising all individual/family facts}
\usage{
fact_summary_indi(gedcom, xref)

fact_summary_famg(gedcom, xref)
}
\arguments{
\item{gedcom}{A tidyged object.}

\item{xref}{The xref of the Individual or Family Group record.}
}
\value{
A tibble containing a row for each fact. A column is created for each tag used, and
each tag is appropriately namespaced.
}
\description{
This function creates a tidy table making it easy to extract fact details for an individual.
}
\details{
Notes and source citations are not included in the summary.
}
