% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/structures.R
\name{family_event_detail}
\alias{family_event_detail}
\title{Construct the FAMILY_EVENT_DETAIL from the GEDCOM specification}
\usage{
family_event_detail(
  husband_age_at_event = character(),
  wife_age_at_event = character(),
  event_details = event_detail()
)
}
\arguments{
\item{husband_age_at_event}{A string that indicates the age in years, months, and days that the
principal was at the time of the associated event. Any labels must come after their
corresponding number, for example; 4y 8m 10d.}

\item{wife_age_at_event}{A string that indicates the age in years, months, and days that the
principal was at the time of the associated event. Any labels must come after their
corresponding number, for example; 4y 8m 10d.}

\item{event_details}{An event_detail() object giving details of the event.}
}
\value{
A tidy tibble containing the FAMILY_EVENT_DETAIL part of a GEDCOM file.
}
\description{
Construct the FAMILY_EVENT_DETAIL from the GEDCOM specification
}
