% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/structures.R
\name{family_event_structure}
\alias{family_event_structure}
\title{Constructs the FAMILY_EVENT_STRUCTURE from the GEDCOM specification}
\usage{
family_event_structure(
  event_type_family,
  event_descriptor = character(),
  family_event_details = family_event_detail()
)
}
\arguments{
\item{event_type_family}{A code used to indicate the type of family event.}

\item{event_descriptor}{Text describing a particular event pertaining to the individual or family.
This event value is usually assigned to the EVEN tag. The classification as to the difference
between this specific event and other occurrences of the EVENt tag is indicated by the use of
a subordinate TYPE tag selected from the EVENT_DETAIL structure.}
}
\value{

}
\description{
Constructs the FAMILY_EVENT_STRUCTURE from the GEDCOM specification
}
