% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/records.R
\name{FAMILY_GROUP_RECORD}
\alias{FAMILY_GROUP_RECORD}
\title{Construct the FAMILY_RECORD tibble}
\usage{
FAMILY_GROUP_RECORD(
  xref_fam,
  events = list(),
  xref_husb = character(),
  xref_wife = character(),
  xrefs_chil = character(),
  count_of_children = character(),
  user_reference_number = character(),
  user_reference_type = character(),
  automated_record_id = character(),
  date_changed = CHANGE_DATE(),
  notes = list(),
  source_citations = list(),
  multimedia_links = list()
)
}
\arguments{
\item{xref_fam}{An xref ID of a family group record.}

\item{events}{A list of FAMILY_EVENT_STRUCTURE() objects giving events associated with this family.}

\item{xref_husb}{An xref ID of the husband.}

\item{xref_wife}{An xref ID of the wife.}

\item{xrefs_chil}{A vector of xref IDs of children in this family.}

\item{count_of_children}{The known number of children of this individual from all marriages or,
if subordinate to a family group record, the reported number of children known to belong to this
family, regardless of whether the associated children are represented in the corresponding
structure. This is not necessarily the count of children listed in a family structure.}

\item{user_reference_number}{A user-defined number or text that the submitter uses to identify this record.}

\item{user_reference_type}{A user-defined definition of the user_reference_number.}

\item{automated_record_id}{A unique record identification number assigned to the record by
the source system. This number is intended to serve as a more sure means of identification
of a record for reconciling differences in data between two interfacing systems.}

\item{date_changed}{A CHANGE_DATE() object giving the time this record was last modified. If not provided,
the current date is used.}

\item{notes}{A list of NOTE_STRUCTURE() objects.}

\item{source_citations}{A list of SOURCE_CITATION() objects.}

\item{multimedia_links}{A list of MULTIMEDIA_LINK() objects}
}
\value{
A tidy tibble containing a FAMILY_RECORD part of a GEDCOM file.
}
\description{
This function constructs a tibble representation of the FAMILY_RECORD from the GEDCOM 5.5.1
specification.
}
