% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_xref.R
\name{find_indi_refn}
\alias{find_indi_refn}
\alias{find_indi_name}
\alias{find_indi_name_all}
\alias{find_repo_refn}
\alias{find_repo_name}
\alias{find_repo_name_all}
\alias{find_note_refn}
\alias{find_note_text}
\alias{find_note_text_all}
\alias{find_media_refn}
\alias{find_media_fileref}
\alias{find_media_fileref_all}
\alias{find_sour_refn}
\alias{find_sour_titl}
\alias{find_sour_titl_all}
\alias{find_famg_refn}
\title{Helper functions to locate record xrefs}
\usage{
find_indi_refn(gedcom, pattern, ignore_case = FALSE)

find_indi_name(gedcom, pattern, ignore_case = FALSE)

find_indi_name_all(gedcom, pattern, ignore_case = FALSE)

find_repo_refn(gedcom, pattern, ignore_case = FALSE)

find_repo_name(gedcom, pattern, ignore_case = FALSE)

find_repo_name_all(gedcom, pattern, ignore_case = FALSE)

find_note_refn(gedcom, pattern, ignore_case = FALSE)

find_note_text(gedcom, pattern, ignore_case = FALSE)

find_note_text_all(gedcom, pattern, ignore_case = FALSE)

find_media_refn(gedcom, pattern, ignore_case = FALSE)

find_media_fileref(gedcom, pattern, ignore_case = FALSE)

find_media_fileref_all(gedcom, pattern, ignore_case = FALSE)

find_sour_refn(gedcom, pattern, ignore_case = FALSE)

find_sour_titl(gedcom, pattern, ignore_case = FALSE)

find_sour_titl_all(gedcom, pattern, ignore_case = FALSE)

find_famg_refn(gedcom, pattern, ignore_case = FALSE)
}
\arguments{
\item{gedcom}{A tidyged object.}

\item{pattern}{The search pattern to use (regular expression).}

\item{ignore_case}{Should case differences be ignored in the match?}
}
\value{
A character vector of xref(s).
}
\description{
These functions act as wrappers to the \code{find_xref} function to find one or more record xrefs.
}
\details{
If you have your own specific use cases to identify records, it's easy to write your own wrapper.
It's best to name your function \verb{find_recordtype_*} and end it with \verb{_all} if it can return multiple xrefs.
If you provide more than one search pattern, you should also include the mode argument.
}
