% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/records.R
\name{FOOTER_SECTION}
\alias{FOOTER_SECTION}
\title{Construct the FOOTER_SECTION tibble}
\usage{
FOOTER_SECTION()
}
\value{
A tidy tibble containing a FOOTER_SECTION part of a GEDCOM file.
}
\description{
This function constructs a tibble representation of the FOOTER_SECTION from the GEDCOM 5.5.1
specification.
}
