% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/records.R
\name{GEDCOM_HEADER}
\alias{GEDCOM_HEADER}
\title{Construct the GEDCOM_HEADER tibble}
\usage{
GEDCOM_HEADER(
  character_encoding = "UTF-8",
  gedcom_version_number = "5.5.5",
  gedcom_form = "LINEAGE-LINKED",
  header_extension = LINEAGE_LINKED_HEADER_EXTENSION()
)
}
\arguments{
\item{character_encoding}{A code value that represents the character set to be used to
interpret this data. Allowed values: "UTF-8" and "UNICODE" (UTF-16).}

\item{gedcom_version_number}{The version number of the specification used.}

\item{gedcom_form}{A value that identifies the GEDCOM form used in this GEDCOM file. The value must be
alphanumerical string. This string is case-sensitive. Only "LINEAGE-LINKED" is supported.}

\item{header_extension}{A LINEAGE_LINKED_HEADER_EXTENSION() object providing more information
about the GEDCOM file.}
}
\value{
A tidy tibble containing a GEDCOM_HEADER part of a GEDCOM file.
}
\description{
This function constructs a tibble representation of the GEDCOM_HEADER from the GEDCOM 5.5.5
specification.
}
