% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{gedcom_value}
\alias{gedcom_value}
\title{Extract a particular value from a tidygedcom object}
\usage{
gedcom_value(gedcom, record_xref, tag, level, after_tag = NULL)
}
\arguments{
\item{gedcom}{A tidygedcom object}

\item{record_xref}{}

\item{tag}{}

\item{level}{}

\item{after_tag}{}
}
\value{

}
\description{
Extract a particular value from a tidygedcom object
}
