% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/identify.R
\name{get_ancestors}
\alias{get_ancestors}
\title{Identify all ancestors for an individual}
\usage{
get_ancestors(
  gedcom,
  indi_xref = character(),
  inc_indi = FALSE,
  inc_sibs = FALSE,
  inc_famg = FALSE,
  inc_supp = FALSE
)
}
\arguments{
\item{gedcom}{A tidyged object.}

\item{indi_xref}{The xref of an Individual record to act on if one
is not activated (will override active record).}

\item{inc_indi}{Whether to also include the individual themselves.}

\item{inc_sibs}{Whether to also include all siblings of ancestors (siblings of this individual will only be
included if the individual is included).}

\item{inc_famg}{Whether to also include all Family Group records where this individual is a child
(and all ancestors' Family Group records).}

\item{inc_supp}{Whether to also include all supporting records (Note, Source, Repository, Multimedia).}
}
\value{
A vector of xrefs of ancestors.
}
\description{
This function identifies records in an entire branch of the family tree above a certain individual.
}
\examples{
get_ancestors(sample555, "@I3@")
get_ancestors(sample555, "@I3@", inc_indi = TRUE)
get_ancestors(sample555, "@I3@", inc_indi = TRUE, inc_famg = TRUE)
}
