% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{get_children}
\alias{get_children}
\title{Get all children for an individual}
\usage{
get_children(gedcom, individual = character(), return_name = FALSE)
}
\arguments{
\item{gedcom}{A tidygedcom object.}

\item{individual}{The xref or name of an Individual record to act on if one
is not activated (will override active record).}

\item{return_name}{Whether to return the childrens name(s) instead of the xref(s).}
}
\value{
A character vector of children xrefs or names.
}
\description{
Get all children for an individual
}
