% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/identify.R
\name{get_descendants}
\alias{get_descendants}
\title{Identify all descendants for an individual}
\usage{
get_descendants(
  gedcom,
  indi_xref = character(),
  inc_indi = FALSE,
  inc_spou = FALSE,
  inc_famg = FALSE,
  inc_supp = FALSE
)
}
\arguments{
\item{gedcom}{A tidyged object.}

\item{indi_xref}{The xref of an Individual record to act on if one
is not activated (will override active record).}

\item{inc_indi}{Whether to also include the individual themselves.}

\item{inc_spou}{Whether to also include all spouses of this individual (and their descendants and
descendants' spouses).}

\item{inc_famg}{Whether to also include all Family Group records where this individual is a spouse
(and all descendants' Family Group records).}

\item{inc_supp}{Whether to also include all supporting records (Note, Source, Repository, Multimedia).}
}
\value{
A vector of xrefs of descendants.
}
\description{
This function identifies records in an entire branch of the family tree below a certain individual.
}
