% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/identify.R
\name{get_famg_children}
\alias{get_famg_children}
\title{Identify all children in a Family Group}
\usage{
get_famg_children(gedcom, famg_xref, birth_only = FALSE, return_name = FALSE)
}
\arguments{
\item{gedcom}{A tidyged object.}

\item{famg_xref}{The xref of a Family Group record to act on if one
is not activated (will override active record).}

\item{birth_only}{Whether to only return biological children.}

\item{return_name}{Whether to return the child name(s) instead of the xref(s).}
}
\value{
A character vector of partner xrefs or names.
}
\description{
Identify all children in a Family Group
}
