% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/identify.R
\name{get_famg_partners}
\alias{get_famg_partners}
\title{Identify all partners in a Family Group}
\usage{
get_famg_partners(gedcom, famg_xref, return_name = FALSE)
}
\arguments{
\item{gedcom}{A tidyged object.}

\item{famg_xref}{The xref of a Family Group record to act on if one
is not activated (will override active record).}

\item{return_name}{Whether to return the partners name(s) instead of the xref(s).}
}
\value{
A character vector of partner xrefs or names.
}
\description{
Identify all partners in a Family Group
}
