% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{get_families_as_child}
\alias{get_families_as_child}
\title{Get all families for an individual where they are a child}
\usage{
get_families_as_child(gedcom, individual = character())
}
\arguments{
\item{gedcom}{A tidygedcom object.}

\item{individual}{The xref or name of an Individual record to act on if one
is not activated (will override active record).}
}
\value{
A character vector of family xrefs.
}
\description{
Get all families for an individual where they are a child
}
