% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/identify.R
\name{get_families_as_partner}
\alias{get_families_as_partner}
\title{Identify all families for an individual where they are a partner}
\usage{
get_families_as_partner(gedcom, indi_xref = character())
}
\arguments{
\item{gedcom}{A tidyged object.}

\item{indi_xref}{The xref of an Individual record to act on if one
is not activated (will override active record).}
}
\value{
A character vector of family xrefs.
}
\description{
Identify all families for an individual where they are a partner
}
\examples{
get_families_as_partner(sample555, "@I2@")
}
