% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interrogate.R
\name{get_family_group_description}
\alias{get_family_group_description}
\title{Get a description of a family}
\usage{
get_family_group_description(gedcom, xref)
}
\arguments{
\item{gedcom}{A tidygedcom object.}

\item{xref}{An xref of a Family group record.}
}
\value{
A character string describing the members of the family group.
}
\description{
Get a description of a family
}
