% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/identify.R
\name{get_indi_cousins}
\alias{get_indi_cousins}
\title{Identify all cousins for an individual}
\usage{
get_indi_cousins(
  gedcom,
  indi_xref = character(),
  degree = 1,
  inc_half_cous = FALSE,
  return_name = FALSE
)
}
\arguments{
\item{gedcom}{A tidyged object.}

\item{indi_xref}{The xref of an Individual record to act on if one
is not activated (will override active record).}

\item{degree}{Whether to return first cousins (degree = 1), second cousins (degree = 2), etc.}

\item{inc_half_cous}{Whether to include half cousins.}

\item{return_name}{Whether to return the parents name(s) instead of the xref(s).}
}
\value{
A character vector of cousin xrefs or names.
}
\description{
Identify all cousins for an individual
}
