% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/identify.R
\name{get_indi_parents}
\alias{get_indi_parents}
\title{Identify all parents for an individual}
\usage{
get_indi_parents(
  gedcom,
  indi_xref = character(),
  birth_only = FALSE,
  return_name = FALSE
)
}
\arguments{
\item{gedcom}{A tidyged object.}

\item{indi_xref}{The xref of an Individual record to act on if one
is not activated (will override active record).}

\item{birth_only}{Whether to only return biological parents.}

\item{return_name}{Whether to return the parents name(s) instead of the xref(s).}
}
\value{
A character vector of parent xrefs or names.
}
\description{
Identify all parents for an individual
}
\examples{
get_indi_parents(sample555, "@I2@")
get_indi_parents(sample555, "@I3@")
get_indi_parents(sample555, "@I3@", return_name = TRUE)
}
