% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/identify.R
\name{get_indi_partners}
\alias{get_indi_partners}
\title{Identify all partners for an individual}
\usage{
get_indi_partners(gedcom, indi_xref = character(), return_name = FALSE)
}
\arguments{
\item{gedcom}{A tidyged object.}

\item{indi_xref}{The xref of an Individual record to act on if one
is not activated (will override active record).}

\item{return_name}{Whether to return the partner's name(s) instead of the xref(s).}
}
\value{
A character vector of partner xrefs or names.
}
\description{
Identify all partners for an individual
}
\examples{
get_indi_partners(sample555, "@I1@")
get_indi_partners(sample555, "@I1@", return_name = TRUE)
get_indi_partners(sample555, "@I3@")
}
