% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/identify.R
\name{get_indi_siblings}
\alias{get_indi_siblings}
\title{Identify all siblings for an individual}
\usage{
get_indi_siblings(
  gedcom,
  indi_xref = character(),
  birth_only = FALSE,
  inc_half_sibs = FALSE,
  return_name = FALSE
)
}
\arguments{
\item{gedcom}{A tidyged object.}

\item{indi_xref}{The xref of an Individual record to act on if one
is not activated (will override active record).}

\item{birth_only}{Whether to only return biological siblings.}

\item{inc_half_sibs}{Whether to include siblings that only share one parent.}

\item{return_name}{Whether to return the parents name(s) instead of the xref(s).}
}
\value{
A character vector of sibling xrefs or names.
}
\description{
Identify all siblings for an individual
}
