% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/identify.R
\name{get_parents}
\alias{get_parents}
\title{Identify all parents for an individual}
\usage{
get_parents(gedcom, indi_xref = character(), return_name = FALSE)
}
\arguments{
\item{gedcom}{A tidyged object.}

\item{indi_xref}{The xref of an Individual record to act on if one
is not activated (will override active record).}

\item{return_name}{Whether to return the parents name(s) instead of the xref(s).}
}
\value{
A character vector of parent xrefs or names.
}
\description{
Identify all parents for an individual
}
\examples{
get_parents(sample555, "@I2@")
get_parents(sample555, "@I3@")
get_parents(sample555, "@I3@", return_name = TRUE)
}
