% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{get_source_title}
\alias{get_source_title}
\title{Retrieve the title of a Source}
\usage{
get_source_title(gedcom, xref)
}
\arguments{
\item{gedcom}{A tidyged object.}

\item{xref}{An xref of a Source record.}
}
\value{
The title found in the Source record. If no title is found, the xref is returned.
}
\description{
Retrieve the title of a Source
}
