% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/identify.R
\name{get_spouses}
\alias{get_spouses}
\title{Identify all spouses for an individual}
\usage{
get_spouses(gedcom, indi_xref = character(), return_name = FALSE)
}
\arguments{
\item{gedcom}{A tidyged object.}

\item{indi_xref}{The xref of an Individual record to act on if one
is not activated (will override active record).}

\item{return_name}{Whether to return the spouse's name(s) instead of the xref(s).}
}
\value{
A character vector of spouse xrefs or names.
}
\description{
Identify all spouses for an individual
}
\examples{
get_spouses(sample555, "@I1@")
get_spouses(sample555, "@I1@", return_name = TRUE)
get_spouses(sample555, "@I3@")
}
