% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{get_valid_xref}
\alias{get_valid_xref}
\title{Derive a valid cross-reference identifier}
\usage{
get_valid_xref(gedcom, xref, record_type, record_type_fn)
}
\arguments{
\item{gedcom}{A tidyged object.}

\item{xref}{A record xref.}

\item{record_type}{A character string describing the record type. Generally one of
the global record_string_* values.}

\item{record_type_fn}{A function to check the record type. Generally one of the is_*
functions.}
}
\value{
A valid xref identifier.
}
\description{
Validate an xref provided explicitly or implicitly (through the active record).
}
\details{
This helper function is designed to derive and run validation checks on an xref
provided explicitly or implicitly. An xref is provided implicitly through the active
record of the tidyged object.

Once found, the xref is checked to ensure it is of the appropriate type.
}
