% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/records.R
\name{HEADER_SECTION}
\alias{HEADER_SECTION}
\title{Construct the HEADER_SECTION tibble}
\usage{
HEADER_SECTION(
  xref_subm,
  approved_system_id,
  character_set,
  xref_subn = character(),
  system_version_number = character(),
  name_of_product = character(),
  name_of_business = character(),
  business_address = ADDRESS_STRUCTURE(character()),
  name_of_source_data = character(),
  publication_date_source_data = date_exact(),
  copyright_source_data = character(),
  receiving_system_name = character(),
  transmission_date = date_exact(),
  transmission_time = character(),
  file_name = character(),
  copyright_gedcom_file = character(),
  character_set_version_number = character(),
  language_of_text = character(),
  place_hierarchy = character(),
  gedcom_content_description = character()
)
}
\arguments{
\item{xref_subm}{An xref ID of a submitter record.}

\item{approved_system_id}{A system identification name which was obtained through the GEDCOM
registration process. This name must be unique from any other product. Spaces within the name
must be substituted with an underscore.}

\item{character_set}{A code value that represents the character set to be used to
interpret this data. Currently, the preferred character set is ANSEL, which
includes ASCII as a subset. UNICODE is not widely supported by most operating systems;
therefore, GEDCOM produced using the UNICODE character set will be limited in its
interchangeability for a while but should eventually provide the international flexibility
that is desired. Allowed values: "ANSEL", "UTF-8", "UNICODE", "ASCII"}

\item{xref_subn}{An xref ID of a submission record.}

\item{system_version_number}{An identifier that represents the version level assigned to the system.
It is defined and changed by the creators of the product.}

\item{name_of_product}{The name of the software product that produced this transmission.}

\item{name_of_business}{Name of the business, corporation, or person that produced or commissioned the product.}

\item{business_address}{An ADDRESS_STRUCTURE() object giving the address of the business.}

\item{name_of_source_data}{The name of the electronic data source that was used to obtain the data in
this transmission. For example, the data may have been obtained from a CD-ROM disc that was named
"U.S. 1880 CENSUS CD-ROM vol. 13."}

\item{publication_date_source_data}{A date_exact() object giving the date this source was published or created.}

\item{copyright_source_data}{A copyright statement required by the owner of data from
which this information was downloaded. For example, when a GEDCOM down-load is requested
from the Ancestral File, this would be the copyright statement to indicate that the data
came from a copyrighted source.}

\item{receiving_system_name}{The name of the system expected to process the GEDCOM-compatible transmission.}

\item{transmission_date}{A date_exact() object giving the date that this transmission was created.}

\item{transmission_time}{The time of a specific event.}

\item{file_name}{The name of the GEDCOM transmission file. If the file name includes a file
extension it must be shown in the form (filename.ext).}

\item{copyright_gedcom_file}{A copyright statement needed to protect the copyrights of the
submitter of this GEDCOM file.}

\item{character_set_version_number}{An identifier that represents the version level assigned to the character set.
It is defined and changed by the creators of the product.}

\item{language_of_text}{The human language in which the data in the transmission is normally read or
written. It is used primarily by programs to select language-specific sorting sequences and phonetic
name matching algorithms.}

\item{place_hierarchy}{This shows the jurisdictional entities that are named in a sequence from the
lowest to the highest jurisdiction. The jurisdictions are separated by commas, and any jurisdiction's
name that is missing is still accounted for by a comma.}

\item{gedcom_content_description}{A note that a user enters to describe the contents of the
lineage-linked file in terms of "ancestors or descendants of" so that the person receiving
the data knows what genealogical information the transmission contains.}
}
\value{
A tidy tibble containing the HEADER part of a GEDCOM file.
}
\description{
This function constructs a tibble representation of the HEADER_SECTION from the GEDCOM 5.5.1
specification.
}
