% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interrogate.R
\name{identify_descendants}
\alias{identify_descendants}
\title{Identify all descendants for an individual}
\usage{
identify_descendants(
  gedcom,
  individual = character(),
  include_individual = FALSE,
  include_spouses = FALSE,
  include_families = FALSE
)
}
\arguments{
\item{gedcom}{A tidyged object.}

\item{individual}{The xref or name of an Individual record to act on if one
is not activated (will override active record).}

\item{include_individual}{Whether to also include the individual themselves.}

\item{include_spouses}{Whether to also include all spouses of this individual (and their descendants).}

\item{include_families}{Whether to also include all Family Group records where this individual is a spouse.}
}
\value{
A vector of xrefs of descendants.
}
\description{
This function identifies records in an entire branch of the family tree below a certain individual.
}
