% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{identify_section}
\alias{identify_section}
\title{Identify the rows of a subrecord in a tidyged object}
\usage{
identify_section(
  gedcom,
  containing_level,
  containing_tag,
  containing_value,
  xrefs = character()
)
}
\arguments{
\item{gedcom}{A tidyged object.}

\item{containing_level}{The level of the first line of the subrecord.}

\item{containing_tag}{The tag of the first line of the subrecord.}

\item{containing_value}{The value of the first line of the subrecord.}

\item{xrefs}{The xrefs of records containing the subrecord (default is all records).}
}
\value{
A vector of rows in the tidyged object of the subrecord(s).
}
\description{
Identify the rows of a subrecord in a tidyged object
}
