% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{identify_unused_records}
\alias{identify_unused_records}
\title{Identify unreferenced records}
\usage{
identify_unused_records(gedcom)
}
\arguments{
\item{gedcom}{A tidyged object.}
}
\value{
A vector of xrefs that are not referenced anywhere else in the tidyged object.
}
\description{
This function identifies records that are not referenced in any other records.
}
\details{
You would expect every record to be referenced by another in some way. For example, Individual
records should reference Family Group records (and vice-versa), Repository records should be referenced
by Source records, and Source records should be cited by other records.

You can use the output of this function with remove_records() to remove them, or describe_records() to
find out more about them.
}
