% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_export_create.R
\name{import_gedcom}
\alias{import_gedcom}
\title{Import a GEDCOM file}
\usage{
import_gedcom(filepath)
}
\arguments{
\item{filepath}{The full filepath of the GEDCOM file}
}
\value{
A tidygedcom object
}
\description{
Imports a *.ged file and creates a tidygedcom object.
}
\examples{
\dontrun{
import_gedcom("C:/my_family.ged")
}
}
