% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/structures.R
\name{individual_attribute_structure}
\alias{individual_attribute_structure}
\title{Constructs the INDIVIDUAL_ATTRIBUTE_STRUCTURE from the GEDCOM specification}
\usage{
individual_attribute_structure(
  attribute_type,
  attribute_description,
  individual_event_details = individual_event_detail()
)
}
\arguments{
\item{attribute_type}{An attribute which may have caused name, addresses, phone numbers,
family listings to be recorded. Its application is in helping to classify sources used for information.}
}
\value{
A tidy tibble containing the INDIVIDUAL_ATTRIBUTE_STRUCTURE part of a GEDCOM file
}
\description{
Constructs the INDIVIDUAL_ATTRIBUTE_STRUCTURE from the GEDCOM specification
}
