% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/structures.R
\name{individual_event_detail}
\alias{individual_event_detail}
\title{Constructs the INDIVIDUAL_EVENT_DETAIL from the GEDCOM specification}
\usage{
individual_event_detail(
  event_details = event_detail(),
  age_at_event = character()
)
}
\arguments{
\item{age_at_event}{A string that indicates the age in years, months, and days that the
principal was at the time of the associated event. Any labels must come after their
corresponding number, for example; 4y 8m 10d.}
}
\value{
A tidy tibble containing the INDIVIDUAL_EVENT_DETAIL part of a GEDCOM file
}
\description{
Constructs the INDIVIDUAL_EVENT_DETAIL from the GEDCOM specification
}
