% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/structures.R
\name{individual_event_structure}
\alias{individual_event_structure}
\title{Construct the INDIVIDUAL_EVENT_STRUCTURE from the GEDCOM specification}
\usage{
individual_event_structure(
  event_type_individual,
  individual_event_details = individual_event_detail(),
  xref_fam = character(),
  adopted_by_which_parent = character()
)
}
\arguments{
\item{event_type_individual}{A code used to indicate the type of individual event.}

\item{individual_event_details}{An individual_event_detail() object giving details of the event.}

\item{xref_fam}{An xref ID of a family record.}

\item{adopted_by_which_parent}{A code which shows which parent in the associated family record adopted this person.
Use "HUSB" for husband, "WIFE" for wife, or "BOTH" for both.}
}
\value{
A tidy tibble containing the INDIVIDUAL_EVENT_STRUCTURE part of a GEDCOM file.
}
\description{
Construct the INDIVIDUAL_EVENT_STRUCTURE from the GEDCOM specification
}
