% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{insert_explicit_marr_types}
\alias{insert_explicit_marr_types}
\title{Insert explicit marriage subrecords for a Family Group record}
\usage{
insert_explicit_marr_types(gedcom, xref)
}
\arguments{
\item{gedcom}{A tidyged object.}

\item{xref}{The xref of a Family Group record.}
}
\value{
The same tidyged object with explicit marriage subrecords in the Family Group record.
}
\description{
Insert explicit marriage subrecords for a Family Group record
}
