% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/identify.R
\name{is_famg_birth_child}
\alias{is_famg_birth_child}
\title{Determine whether an individual is a child of a family by birth}
\usage{
is_famg_birth_child(gedcom, indi_xref, famg_xref)
}
\arguments{
\item{gedcom}{A tidyged object.}

\item{indi_xref}{The xref of an Individual record to act on if one
is not activated (will override active record).}

\item{famg_xref}{The xref of the Family Group record.}
}
\value{
A logical value indicating whether the individual is a child of a family by birth.
}
\description{
Determine whether an individual is a child of a family by birth
}
