% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interrogate.R
\name{is_indi}
\alias{is_indi}
\alias{is_famg}
\alias{is_subm}
\alias{is_repo}
\alias{is_media}
\alias{is_note}
\alias{is_sour}
\title{Check whether a given record is a particular type}
\usage{
is_indi(gedcom, xref)

is_famg(gedcom, xref)

is_subm(gedcom, xref)

is_repo(gedcom, xref)

is_media(gedcom, xref)

is_note(gedcom, xref)

is_sour(gedcom, xref)
}
\arguments{
\item{gedcom}{A tidyged object.}

\item{xref}{The xref of the record.}
}
\value{
A logical indicating whether the record is of a particular type.
}
\description{
Check whether a given record is a particular type
}
\examples{
is_indi(sample555, "@I1@")
is_note(sample555, "@I1@")
is_famg(sample555, "@F2@")
is_repo(sample555, "@R1@")
}
