% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{is_individual}
\alias{is_individual}
\alias{is_family}
\alias{is_submitter}
\alias{is_submission}
\alias{is_repository}
\alias{is_multimedia}
\alias{is_note}
\alias{is_source}
\title{Check whether a given record is a particular type}
\usage{
is_individual(gedcom, xref)

is_family(gedcom, xref)

is_submitter(gedcom, xref)

is_submission(gedcom, xref)

is_repository(gedcom, xref)

is_multimedia(gedcom, xref)

is_note(gedcom, xref)

is_source(gedcom, xref)
}
\arguments{
\item{gedcom}{A tidygedcom object.}

\item{xref}{The xref of the record.}
}
\value{
A logical indicating whether the record is of a particular type.
}
\description{
Check whether a given record is a particular type
}
