% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/structures.R
\name{multimedia_link}
\alias{multimedia_link}
\title{Constructs the MULTIMEDIA_LINK from the GEDCOM specification}
\usage{
multimedia_link(
  xref_obje = character(),
  multimedia_file_refn = character(),
  multimedia_format = character(),
  source_media_type = character(),
  descriptive_title = character()
)
}
\arguments{
\item{xref_obje}{Either a pointer or an unique cross-reference identifier.}

\item{multimedia_format}{Indicates the format of the multimedia data associated with the specific
GEDCOM context. This allows processors to determine whether they can process the data object.
Any linked files should contain the data required, in the indicated format, to process the file data.}

\item{source_media_type}{A code, selected from one of the media classifications choices above,
that indicates the type of material in which the referenced source is stored.}

\item{descriptive_title}{The title of a work, record, item, or object.}
}
\value{

}
\description{
Constructs the MULTIMEDIA_LINK from the GEDCOM specification
}
