% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/records.R
\name{MULTIMEDIA_RECORD}
\alias{MULTIMEDIA_RECORD}
\title{Construct the MULTIMEDIA_RECORD tibble}
\usage{
MULTIMEDIA_RECORD(
  xref_obje,
  multimedia_file_reference,
  multimedia_format,
  source_media_type = character(),
  descriptive_title = character(),
  user_reference_number = character(),
  user_reference_type = character(),
  automated_record_id = character(),
  notes = list(),
  source_citations = list(),
  date_changed = CHANGE_DATE()
)
}
\arguments{
\item{xref_obje}{An xref ID of a multimedia record.}

\item{multimedia_file_reference}{A complete local or remote file reference to the auxiliary data
to be linked to the GEDCOM context. Remote reference would include a network address where the multimedia
data may be obtained.}

\item{multimedia_format}{Indicates the format of the multimedia data associated with the specific
GEDCOM context. This allows processors to determine whether they can process the data object.
Any linked files should contain the data required, in the indicated format, to process the file data. One of:
"AAC", "AVI", "BMP", "ePub", "FLAC", "GIF", "JPEG", "JPG", "MKV", "mobi", "MP3", "PCX", "PDF", "PNG", "TIFF",
"TIF", "WAV".}

\item{source_media_type}{A media classification code that indicates the type of material in which
the referenced source is stored. One of: "audio", "book", "card", "electronic", "fiche", "film",
"magazine", "manuscript", "map", "newspaper", "photo", "tombstone", "video".}

\item{descriptive_title}{The title of a work, record, item, or object.}

\item{user_reference_number}{A user-defined number or text that the submitter uses to identify this record.}

\item{user_reference_type}{A user-defined definition of the user_reference_number.}

\item{automated_record_id}{A unique record identification number assigned to the record by
the source system. This number is intended to serve as a more sure means of identification
of a record for reconciling differences in data between two interfacing systems.}

\item{notes}{A list of NOTE_STRUCTURE() objects.}

\item{source_citations}{A list of SOURCE_CITATION() objects.}

\item{date_changed}{A CHANGE_DATE() object giving the time this record was last modified. If not provided,
the current date is used.}
}
\value{
A tidy tibble containing a MULTIMEDIA_RECORD part of a GEDCOM file.
}
\description{
This function constructs a tibble representation of the MULTIMEDIA_RECORD from the GEDCOM 5.5.5
specification.
}
