% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{mutate_tag_namespace}
\alias{mutate_tag_namespace}
\title{Add a tag namespace column to a tidyged object}
\usage{
mutate_tag_namespace(gedcom)
}
\arguments{
\item{gedcom}{A tidyged object.}
}
\value{
A tidyged object with an additional 'tag_ns' column containing the full namespace of
the tag.
}
\description{
Add a tag namespace column to a tidyged object
}
\details{
This function is useful if you want to find the namespace of a particular value
for the \code{find_xref} function.
}
