% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/structures.R
\name{name_pieces}
\alias{name_pieces}
\title{Define a personal name's components}
\usage{
name_pieces(
  prefix = character(),
  given = character(),
  nickname = character(),
  surname_prefix = character(),
  surname = character(),
  suffix = character(),
  notes = character()
)
}
\arguments{
\item{prefix}{The name prefix, e.g. Cmdr.}

\item{given}{The given name or earned name. Different given names are separated
by a comma.}

\item{nickname}{A descriptive or familiar name used in connection with one's
proper name.}

\item{surname_prefix}{Surname prefix or article used in a family name.
For example in the name "de la Cruz", this value would be "de la".}

\item{surname}{Surname or family name. Different surnames are separated by a comma.}

\item{suffix}{Non-indexing name piece that appears after the given name and surname
parts, e.g. Jr. Different name suffix parts are separated by a comma.}

\item{notes}{A character vector of notes accompanying this name.
These could be xrefs to existing Note records.}
}
\value{
A tibble describing a personal name's components.
}
\description{
Define a personal name's components
}
