% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/records.R
\name{NOTE_RECORD}
\alias{NOTE_RECORD}
\title{Construct the NOTE_RECORD tibble}
\usage{
NOTE_RECORD(
  xref_note,
  user_text,
  user_reference_number = character(),
  user_reference_type = character(),
  automated_record_id = character(),
  source_citations = list(),
  date_changed = CHANGE_DATE()
)
}
\arguments{
\item{xref_note}{An xref ID of a note record.}

\item{user_text}{Free-form user text. Comments, opinions.}

\item{user_reference_number}{A user-defined number or text that the submitter uses to identify this record.}

\item{user_reference_type}{A user-defined definition of the user_reference_number.}

\item{automated_record_id}{A unique record identification number assigned to the record by
the source system. This number is intended to serve as a more sure means of identification
of a record for reconciling differences in data between two interfacing systems.}

\item{source_citations}{A list of SOURCE_CITATION() objects.}

\item{date_changed}{A CHANGE_DATE() object giving the time this record was last modified. If not provided,
the current date is used.}
}
\value{
A tidy tibble containing a NOTE_RECORD part of a GEDCOM file.
}
\description{
This function constructs a tibble representation of the NOTE_RECORD from the GEDCOM 5.5.5
specification.
}
