% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/structures.R
\name{note_structure}
\alias{note_structure}
\title{Constructs the NOTE_STRUCTURE from the GEDCOM specification}
\usage{
note_structure(xref_note = character(), submitter_text = character())
}
\arguments{
\item{xref_note}{Either a pointer or an unique cross-reference identifier.}

\item{submitter_text}{Comments or opinions from the submitter.}
}
\value{
A tidy tibble containing the NOTE_STRUCTURE part of a GEDCOM file
}
\description{
Constructs the NOTE_STRUCTURE from the GEDCOM specification
}
