% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interrogate.R
\name{num_indi}
\alias{num_indi}
\alias{num_famg}
\alias{num_subm}
\alias{num_media}
\alias{num_note}
\alias{num_repo}
\alias{num_sour}
\title{Get the number of records in a tidyged object}
\usage{
num_indi(gedcom)

num_famg(gedcom)

num_subm(gedcom)

num_media(gedcom)

num_note(gedcom)

num_repo(gedcom)

num_sour(gedcom)
}
\arguments{
\item{gedcom}{A tidyged object.}
}
\value{
The number of records of the relevant type.
}
\description{
These functions return the number of records of a particular type in a tidyged object.
}
\examples{
num_indi(sample555)
num_famg(sample555)
num_repo(sample555)
}
