% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{order_famg_children}
\alias{order_famg_children}
\title{Order children in a Family Group record by birth date}
\usage{
order_famg_children(gedcom, xref)
}
\arguments{
\item{gedcom}{A tidyged object.}

\item{xref}{The xref of a Family Group record.}
}
\value{
The same tidyged object with rearranged children rows in the Family Group record.
}
\description{
Order children in a Family Group record by birth date
}
\details{
Any children without a date a birth are placed last.
}
