% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subrecords.R
\name{PERSONAL_NAME_PIECES}
\alias{PERSONAL_NAME_PIECES}
\title{Construct the PERSONAL_NAME_PIECES tibble}
\usage{
PERSONAL_NAME_PIECES(
  name_piece_prefix = character(),
  name_piece_given = character(),
  name_piece_nickname = character(),
  name_piece_surname_prefix = character(),
  name_piece_surname = character(),
  name_piece_suffix = character(),
  notes = list(),
  source_citations = list()
)
}
\arguments{
\item{name_piece_prefix}{Non indexing name piece that appears preceding the given name and surname parts.
Different name prefix parts are separated by a comma.}

\item{name_piece_given}{Given name or earned name. Different given names are separated by a comma.}

\item{name_piece_nickname}{A descriptive or familiar name used in connection with one's proper name.}

\item{name_piece_surname_prefix}{Surname prefix or article used in a family name. A surname prefix that
consists of multiple parts is written as is, and not modified in any way. Thus, the surname prefix for the
surname “de la Cruz” is “de la”.}

\item{name_piece_surname}{Surname or family name. Different surnames are separated by a comma.}

\item{name_piece_suffix}{Non-indexing name piece that appears after the given name and surname parts.
Different name suffix parts are separated by a comma.}

\item{notes}{A list of NOTE_STRUCTURE() objects.}

\item{source_citations}{A list of SOURCE_CITATION() objects.}
}
\value{
A tidy tibble containing the PERSONAL_NAME_PIECES part of a GEDCOM file.
}
\description{
This function constructs a tibble representation of the PERSONAL_NAME_PIECES from the GEDCOM 5.5.5
specification.
}
