% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/structures.R
\name{PERSONAL_NAME_STRUCTURE}
\alias{PERSONAL_NAME_STRUCTURE}
\title{Construct the PERSONAL_NAME_STRUCTURE tibble}
\usage{
PERSONAL_NAME_STRUCTURE(
  name_personal,
  name_type = character(),
  name_pieces = PERSONAL_NAME_PIECES(),
  name_phonetic_variation = character(),
  phonetic_type = character(),
  phonetic_name_pieces = list(),
  name_romanized_variation = character(),
  romanized_type = character(),
  romanized_name_pieces = list()
)
}
\arguments{
\item{name_personal}{The surname of an individual, if known, is enclosed between two slash (/) characters.
The order of the name parts should be the order that the person would, by custom of their culture,
have used when giving it to a recorder.}

\item{name_type}{Indicates the name type, for example the name issued or assumed as an immigrant.}

\item{name_pieces}{A PERSONAL_NAME_PIECES() object giving the components of the name.}

\item{name_phonetic_variation}{A character vector of phonetic variations of the name.}

\item{phonetic_type}{Indicates the method used in transforming the text to the phonetic variation.}

\item{phonetic_name_pieces}{A list of PERSONAL_NAME_PIECES() objects giving the components
of the phonetic name variations.}

\item{name_romanized_variation}{A character vector giving romanized variations of the name.}

\item{romanized_type}{Indicates the method used in transforming the text to a romanized variation.}

\item{romanized_name_pieces}{A list of PERSONAL_NAME_PIECES() objects giving the components
of the romanized name variations.}
}
\value{
A tidy tibble containing the PERSONAL_NAME_STRUCTURE part of a GEDCOM file.
}
\description{
This function constructs a tibble representation of the PERSONAL_NAME_STRUCTURE from the GEDCOM 5.5.1
specification.
}
