% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/structures.R
\name{place}
\alias{place}
\title{Define a place associated with a fact}
\usage{
place(
  name = character(),
  phonetic_var = character(),
  romanised_var = character(),
  latitude = character(),
  longitude = character(),
  notes = character()
)
}
\arguments{
\item{name}{The jurisdictional name of the place.
Jurisdictions are separated by commas, for example, "Cove, Cache, Utah, USA."}

\item{phonetic_var}{A named character vector of phonetic variations of the place name. Element names
must give the phonetisation method used in transforming the text to the corresponding phonetic variation.
i.e. c(method1 = "var1", method2 = "var2")}

\item{romanised_var}{A named character vector of romanised variations of the place name. Element names
must give the romanisation method used in transforming the text to the corresponding romanised variation.
i.e. c(method1 = "var1", method2 = "var2")}

\item{latitude}{The value specifying the latitudinal coordinate of the place.
The latitude coordinate is the direction North or South from the equator in degrees and
fraction of degrees carried out to give the desired accuracy.
For example: 18 degrees, 9 minutes, and 3.4 seconds North would be formatted as "N18.150944"}

\item{longitude}{The value specifying the longitudinal coordinate of the place.
The longitude coordinate is Degrees and fraction of degrees east or west of the zero or
base meridian coordinate. For example:
168 degrees, 9 minutes, and 3.4 seconds East would be formatted as "E168.150944".}

\item{notes}{A character vector of notes accompanying the place.
These could be xrefs to existing Note records.}
}
\value{
A tibble describing a place.
}
\description{
Define a place associated with a fact
}
