% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subrecords.R
\name{PLACE_STRUCTURE}
\alias{PLACE_STRUCTURE}
\title{Construct the PLACE_STRUCTURE tibble}
\usage{
PLACE_STRUCTURE(
  place_name,
  place_phonetic = character(),
  phonetisation_method = character(),
  place_romanised = character(),
  romanisation_method = character(),
  place_latitude = character(),
  place_longitude = character(),
  notes = list()
)
}
\arguments{
\item{place_name}{The jurisdictional name of the place where the event took place.
Jurisdictions are separated by commas. No part of the place name may be replaced by an abbreviation.
Place names are not terminated by a full stop or anything else.}

\item{place_phonetic}{A character vector of phonetic variations of the place name.}

\item{phonetisation_method}{Indicates the method used in transforming the text to the phonetic variation.}

\item{place_romanised}{A character vector of romanised variations of the place name.}

\item{romanisation_method}{Indicates the method used in transforming the text to a romanised variation.}

\item{place_latitude}{The value specifying the latitudinal coordinate of the place name.
The latitude coordinate is the direction North or South from the equator in degrees and fraction of
degrees carried out to give the desired accuracy. For example: 18 degrees, 9 minutes, and 3.4 seconds North
would be formatted as N18.150944.}

\item{place_longitude}{The value specifying the longitudinal coordinate of the place name.
The longitude coordinate is Degrees and fraction of degrees east or west of the zero or base
meridian coordinate. For example: 168 degrees, 9 minutes, and 3.4 seconds East would be formatted as E168.150944.}

\item{notes}{A list of NOTE_STRUCTURE() objects.}
}
\value{
A tidy tibble containing the PLACE_STRUCTURE part of a GEDCOM file.
}
\description{
This function constructs a tibble representation of the PLACE_STRUCTURE from the GEDCOM 5.5.5
specification.
}
