% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/edit_individual_names.R
\name{primary_individual_name}
\alias{primary_individual_name}
\title{Make an Individual name appear first in the Individual record}
\usage{
primary_individual_name(gedcom, name)
}
\arguments{
\item{gedcom}{A tidygedcom object.}

\item{name}{The personal name to move.}
}
\value{
An updated tidygedcom object with the promoted name in the Individual record
}
\description{
Make an Individual name appear first in the Individual record
}
