% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_write_create.R
\name{read_gedcom_encoding}
\alias{read_gedcom_encoding}
\title{Read the Byte Order Mark of the GEDCOM file}
\usage{
read_gedcom_encoding(filepath)
}
\arguments{
\item{filepath}{The full filepath of the GEDCOM file.}
}
\value{
A character string indicating the encoding of the file.
}
\description{
This function reads the Byte Order Mark of a GEDCOM file in order to determine its encoding.
It only checks for UTF-8 or UTF-16 - if neither of these are found it throws an error.
}
