% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{remove_change_dates}
\alias{remove_change_dates}
\title{Remove all CHANge dates}
\usage{
remove_change_dates(gedcom)
}
\arguments{
\item{gedcom}{A tidygedcom object.}
}
\value{
A tidygedcom object with all CHAN structures removed.
}
\description{
Remove all CHANge dates
}
