% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{remove_context_from_tests}
\alias{remove_context_from_tests}
\title{Remove all context lines from test files}
\usage{
remove_context_from_tests()
}
\value{
Nothing.
}
\description{
Remove all context lines from test files
}
\details{
This is a function used in development run after devtools::document().
It removes all context() lines from test files inserted automatically by roxytest.
}
