% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{remove_dates_for_tests}
\alias{remove_dates_for_tests}
\title{Remove all creation dates from a tidygedcom object}
\usage{
remove_dates_for_tests(gedcom)
}
\arguments{
\item{gedcom}{A tidygedcom object.}
}
\value{
The tidygedcom object with creation dates removed.
}
\description{
Remove all creation dates from a tidygedcom object
}
\details{
This is a function used in tests so that the objects created do not
change every time.
}
