% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/edit_family_group.R
\name{remove_empty_family_groups}
\alias{remove_empty_family_groups}
\title{Remove Family Group records with no members}
\usage{
remove_empty_family_groups(gedcom)
}
\arguments{
\item{gedcom}{A tidygedcom object.}
}
\value{
A new tidygedcom object with empty Family Group records removed.
}
\description{
Remove Family Group records with no members
}
