% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/edit_family_group.R
\name{remove_family}
\alias{remove_family}
\title{Remove a Family record from a tidygedcom object}
\usage{
remove_family(gedcom, remove_individuals = FALSE)
}
\arguments{
\item{gedcom}{A tidygedcom object.}

\item{remove_individuals}{Whether to also remove the individual records for all individuals
in the family.}
}
\value{
An updated tidygedcom object excluding the active Family record (and potentially the
individuals within it).
}
\description{
This function removes a record containing details of a family.
}
\details{
This function will also automatically remove references to this record in other
individual records. If remove_individuals is set to TRUE, it will also remove
all records for individuals in this family (including associations, but not aliases).
}
