% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/edit_family_group.R
\name{remove_family_group}
\alias{remove_family_group}
\title{Remove a Family group record from a tidyged object}
\usage{
remove_family_group(
  gedcom,
  family_xref = character(),
  remove_individuals = FALSE
)
}
\arguments{
\item{gedcom}{A tidyged object.}

\item{family_xref}{The xref of a Family Group record to act on if one is not
activated (will override active record).}

\item{remove_individuals}{Whether to also remove the records for all Individuals
in the family.}
}
\value{
An updated tidyged object excluding the selected Family group record
(and potentially the individuals within it).
}
\description{
This function removes a record containing details of a family group.
}
\details{
This function will also automatically remove references to this record in other
individual records. If remove_individuals is set to TRUE, it will also remove
all records for individuals in this family (including associations).
}
