% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/individual.R
\name{remove_indi}
\alias{remove_indi}
\title{Remove an Individual record from a tidyged object}
\usage{
remove_indi(gedcom, individual = character(), remove_associations = TRUE)
}
\arguments{
\item{gedcom}{A tidyged object.}

\item{individual}{The xref or name of an Individual record to act on if one
is not activated (will override active record).}

\item{remove_associations}{Whether to also remove associations with this individual in
other individual records. Defaults to TRUE.}
}
\value{
An updated tidyged object excluding the selected Individual record.
}
\description{
This function removes an active Individual record from the tidyged object.
}
\details{
At a minimum it will also remove references to this individual in Family group records.
If remove_associations is TRUE (default) it will remove associations with this
individual in other Individual records.
}
