% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/edit_individual_names.R
\name{remove_indi_name}
\alias{remove_indi_name}
\title{Remove a personal name (and components) from an Individual record}
\usage{
remove_indi_name(gedcom, name)
}
\arguments{
\item{gedcom}{A tidyged object.}

\item{name}{The personal name to remove.}
}
\value{
An updated tidyged object with an Individual record excluding
this personal name (and components).
}
\description{
Remove a personal name (and components) from an Individual record
}
