% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/edit_individual_names.R
\name{remove_individual_name}
\alias{remove_individual_name}
\title{Remove a personal name (and components) from an Individual record}
\usage{
remove_individual_name(gedcom, name, xref = character())
}
\arguments{
\item{gedcom}{A tidyged object.}

\item{name}{The personal name to remove.}

\item{xref}{The xref of a record to act on if one is not activated (will override active record).}
}
\value{
An updated tidyged object with an Individual record excluding
this personal name (and components).
}
\description{
Remove a personal name (and components) from an Individual record
}
