% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/edit_individual_names.R
\name{remove_individual_name_var}
\alias{remove_individual_name_var}
\title{Remove a variation of a personal name from an Individual record}
\usage{
remove_individual_name_var(gedcom, variation_name, phonetic_variation = TRUE)
}
\arguments{
\item{gedcom}{A tidygedcom object.}

\item{variation_name}{The personal name variation to remove.}

\item{phonetic_variation}{Whether the name variation is a phonetic variation
(TRUE, default) or a romanized variation (FALSE).}
}
\value{
An updated tidygedcom object with an Individual record excluding
this personal name variation (and components).
}
\description{
Remove a variation of a personal name from an Individual record
}
