% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/edit_multimedia.R
\name{remove_media}
\alias{remove_media}
\title{Remove a Multimedia record from a tidyged object}
\usage{
remove_media(gedcom, multimedia = character())
}
\arguments{
\item{gedcom}{A tidyged object.}

\item{multimedia}{The xref or file reference of a Multimedia record to act
on if one is not activated (will override active record).}
}
\value{
An updated tidyged object excluding the selected Multimedia record.
}
\description{
Remove a Multimedia record from a tidyged object
}
