% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/note.R
\name{remove_note}
\alias{remove_note}
\title{Remove a Note record from a tidyged object}
\usage{
remove_note(gedcom, note = character())
}
\arguments{
\item{gedcom}{A tidyged object.}

\item{note}{The xref or excerpt of a Note record to act on if one is
not activated (will override active record).}
}
\value{
An updated tidyged object excluding the selected Note record.
}
\description{
Remove a Note record from a tidyged object
}
