% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{remove_records}
\alias{remove_records}
\title{Remove multiple records at once}
\usage{
remove_records(gedcom, xrefs)
}
\arguments{
\item{gedcom}{A tidyged object.}

\item{xrefs}{A vector of xrefs to remove.}
}
\value{
An updated tidyged object with the records removed.
}
\description{
Remove multiple records at once
}
