% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/edit_repository.R
\name{remove_repo}
\alias{remove_repo}
\title{Remove a Repository record from a tidyged object}
\usage{
remove_repo(gedcom, repository = character())
}
\arguments{
\item{gedcom}{A tidyged object.}

\item{repository}{The xref or name of a Repository record to act on if one is
not activated (will override active record).}
}
\value{
An updated tidyged object excluding the selected Repository record.
}
\description{
Remove a Repository record from a tidyged object
}
