% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{remove_section}
\alias{remove_section}
\title{Remove a subrecord in a tidygedcom object}
\usage{
remove_section(
  gedcom,
  containing_level,
  containing_tag,
  containing_value,
  xrefs = character()
)
}
\arguments{
\item{gedcom}{A tidygedcom object.}

\item{containing_level}{The level of the first line of the subrecord.}

\item{containing_tag}{The tag of the first line of the subrecord.}

\item{containing_value}{The value of the first line of the subrecord.}

\item{xrefs}{The xrefs of records containing the subrecord (default is all records).}
}
\value{
The tidygedcom object with the subrecord(s) removed.
}
\description{
Remove a subrecord in a tidygedcom object
}
