% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/edit_source.R
\name{remove_source}
\alias{remove_source}
\title{Remove a Source record from a tidyged object}
\usage{
remove_source(gedcom, source = character())
}
\arguments{
\item{gedcom}{A tidyged object.}

\item{source}{The xref or title of a Source record to act on if one is not
activated (will override active record).}
}
\value{
An updated tidyged object excluding the selected Source record.
}
\description{
Remove a Source record from a tidyged object
}
