% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/edit_submission.R
\name{remove_submission}
\alias{remove_submission}
\title{Remove a Submission record from a tidygedcom object}
\usage{
remove_submission(gedcom)
}
\arguments{
\item{gedcom}{A tidygedcom object.}
}
\value{
An updated tidygedcom object excluding the Submission record.
}
\description{
Remove a Submission record from a tidygedcom object
}
