% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/edit_submitter.R
\name{remove_submitter}
\alias{remove_submitter}
\title{Remove a Submitter record from a tidygedcom object}
\usage{
remove_submitter(gedcom)
}
\arguments{
\item{gedcom}{A tidygedcom object.}
}
\value{
An updated tidygedcom object excluding the active Submitter record.
}
\description{
Remove a Submitter record from a tidygedcom object
}
