% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/records.R
\name{REPOSITORY_RECORD}
\alias{REPOSITORY_RECORD}
\title{Construct the REPOSITORY_RECORD tibble}
\usage{
REPOSITORY_RECORD(
  xref_repo,
  name_of_repository,
  address = ADDRESS_STRUCTURE(character()),
  notes = list(),
  user_reference_number = character(),
  user_reference_type = character(),
  automated_record_id = character(),
  date_changed = CHANGE_DATE()
)
}
\arguments{
\item{xref_repo}{An xref ID of a repository record.}

\item{name_of_repository}{The official name of the archive in which the stated source material is stored.}

\item{address}{An ADDRESS_STRUCTURE() object giving details of the repository address.}

\item{notes}{A list of NOTE_STRUCTURE() objects.}

\item{user_reference_number}{A user-defined number or text that the submitter uses to identify this record.}

\item{user_reference_type}{A user-defined definition of the user_reference_number.}

\item{automated_record_id}{A unique record identification number assigned to the record by
the source system. This number is intended to serve as a more sure means of identification
of a record for reconciling differences in data between two interfacing systems.}

\item{date_changed}{A CHANGE_DATE() object giving the time this record was last modified. If not provided,
the current date is used.}
}
\value{
A tidy tibble containing a REPOSITORY_RECORD part of a GEDCOM file.
}
\description{
This function constructs a tibble representation of the REPOSITORY_RECORD from the GEDCOM 5.5.1
specification.
}
