% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{sample555}
\alias{sample555}
\title{A sample GEDCOM file in tidyged format}
\format{
A tibble with 97 rows and 4 variables:
\describe{
\item{level}{The level in the GEDCOM structural hierarchy}
\item{record}{The unique cross reference for the record}
\item{tag}{The tag describing the nature of the line value given}
\item{value}{The GEDCOM line value}
}
}
\source{
\url{https://www.gedcom.org/samples.html}
}
\usage{
sample555
}
\description{
This is the official GEDCOM 5.5.5 sample file.
}
\keyword{datasets}
