% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/activate.R
\name{set_active_record}
\alias{set_active_record}
\title{Flag a record as being active}
\usage{
set_active_record(gedcom, xref)
}
\arguments{
\item{gedcom}{A tidyged object}

\item{xref}{The xref of the record to activate}
}
\value{
The same tidyged object with an "active_record" attribute set to the xref of the record
}
\description{
This allows an easy mechanism to edit particular records without specifying them each time.
}
