% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/structures.R
\name{source_citation}
\alias{source_citation}
\title{Constructs the SOURCE_CITATION from the GEDCOM specification}
\usage{
source_citation(
  source_ref,
  page = character(),
  event_type = character(),
  role = character(),
  entry_recording_date = character(),
  source_text = character(),
  certainty_assessment = character(),
  multimedia_links = list(),
  notes = list()
)
}
\arguments{
\item{entry_recording_date}{The date that this event data was entered into the original source document.}

\item{certainty_assessment}{A quantitative evaluation of the credibility of a piece of
information, based upon its supporting evidence. Some systems use this feature to rank multiple
conflicting opinions for display of most likely information first. It is not intended to
eliminate the receiver's need to evaluate the evidence for themselves.
0 = unreliable/estimated data
1 = Questionable reliability of evidence
2 = Secondary evidence, data officially recorded sometime after event
3 = Direct and primary evidence used, or by dominance of the evidence}
}
\value{

}
\description{
Constructs the SOURCE_CITATION from the GEDCOM specification
}
