% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/records.R
\name{SOURCE_RECORD}
\alias{SOURCE_RECORD}
\title{Construct the SOURCE_RECORD tibble}
\usage{
SOURCE_RECORD(
  xref_sour,
  events_recorded = character(),
  date_period_covered = date_period(),
  source_jurisdiction_place = character(),
  responsible_agency = character(),
  data_notes = list(),
  source_originator = character(),
  source_descriptive_title = character(),
  source_filed_by_entry = character(),
  source_publication_facts = character(),
  text_from_source = character(),
  source_repository_citations = list(),
  user_reference_number = character(),
  user_reference_type = character(),
  automated_record_id = character(),
  date_changed = CHANGE_DATE(),
  notes = list(),
  multimedia_links = list()
)
}
\arguments{
\item{xref_sour}{An xref ID of a source record.}

\item{events_recorded}{An enumeration of the different kinds of events that were recorded in a
particular source. Each enumeration is separated by a comma. Such as a parish register of births,
deaths, and marriages would be BIRT, DEAT, MARR. These can be enumerated over more than one
vector element.}

\item{date_period_covered}{A date_period() object associated with the period covered by the course.}

\item{source_jurisdiction_place}{The name of the lowest jurisdiction that encompasses all lower-level places
named in this source.}

\item{responsible_agency}{The organization, institution, corporation, person, or other entity that has
responsibility for the associated context. For example, an employer of a person of an associated occupation,
or a church that administered rites or events, or an organization responsible for creating and/or archiving
records.}

\item{data_notes}{A list of NOTE_STRUCTURE() objects associated with the data in this source.}

\item{source_originator}{The person, agency, or entity who created the record. For a published work,
this could be the author, compiler, transcriber, abstractor, or editor. For an unpublished source,
this may be an individual, a government agency, church organization, or private organization, etc.}

\item{source_descriptive_title}{The title of the work, record, or item and, when appropriate,
the title of the larger work or series of which it is a part.}

\item{source_filed_by_entry}{This entry is to provide a short title used for sorting, filing,
and retrieving source records.}

\item{source_publication_facts}{When and where the record was created.}

\item{text_from_source}{A verbatim copy of any description contained within the source.
This indicates notes or text that are actually contained in the source document, not the submitter's
opinion about the source.}

\item{source_repository_citations}{A list of SOURCE_REPOSITORY_CITATION() objects.}

\item{user_reference_number}{A user-defined number or text that the submitter uses to identify this record.}

\item{user_reference_type}{A user-defined definition of the user_reference_number.}

\item{automated_record_id}{A unique record identification number assigned to the record by
the source system. This number is intended to serve as a more sure means of identification
of a record for reconciling differences in data between two interfacing systems.}

\item{date_changed}{A CHANGE_DATE() object giving the time this record was last modified. If not provided,
the current date is used.}

\item{notes}{A list of NOTE_STRUCTURE() objects.}

\item{multimedia_links}{A list of MULTIMEDIA_LINK() objects}
}
\value{
A tidy tibble containing a SOURCE_RECORD part of a GEDCOM file.
}
\description{
This function constructs a tibble representation of the SOURCE_RECORD from the GEDCOM 5.5.1
specification.
}
