% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/structures.R
\name{source_repository_citation}
\alias{source_repository_citation}
\title{Construct the SOURCE_REPOSITORY_CITATION from the GEDCOM specification}
\usage{
source_repository_citation(
  xref_repo,
  notes = list(),
  source_call_number = character(),
  source_media_type = character()
)
}
\arguments{
\item{xref_repo}{An xref ID of a repository record.}

\item{notes}{A list of note_structure() objects.}

\item{source_call_number}{An identification or reference description used to file and retrieve items
from the holdings of a repository.}

\item{source_media_type}{A code, selected from one of the media classifications choices above,
that indicates the type of material in which the referenced source is stored.}
}
\value{
A tidy tibble containing the SOURCE_REPOSITORY_CITATION part of a GEDCOM file.
}
\description{
Construct the SOURCE_REPOSITORY_CITATION from the GEDCOM specification
}
