% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create.R
\name{split_gedcom}
\alias{split_gedcom}
\title{Split a tidygedcom object into two}
\usage{
split_gedcom(gedcom, xrefs, remove_dead_refs = TRUE)
}
\arguments{
\item{gedcom}{A tidygedcom object to split.}

\item{xrefs}{A vector of xrefs to put into the new tidyged object.}

\item{remove_dead_refs}{Whether to remove references to records not in the new tidyged object.}
}
\value{
A new tidyged object containing the xrefs specified. It will also have the same
header and submitter information as the input tidyged object.
}
\description{
Split a tidygedcom object into two
}
