% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_write_create.R
\name{split_gedcom_values}
\alias{split_gedcom_values}
\title{Convert the GEDCOM form to GEDCOM grammar}
\usage{
split_gedcom_values(gedcom, char_limit)
}
\arguments{
\item{gedcom}{A tidyged object.}

\item{char_limit}{Maximum string length of values.}
}
\value{
A tidyged object in the GEDCOM grammar ready to export.
}
\description{
This function introduces CONC/CONT lines for line values that exceed the given number of characters.
This function only uses the CONC(atenation) tag for splitting values, and does
not use the CONT(inuation) tag. This is because it is easier to implement.
}
