% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{split_text}
\alias{split_text}
\title{Split a vector of free text into separate rows of a GEDCOM file}
\usage{
split_text(text, top_tag, char_limit = 248, start_level = 0)
}
\arguments{
\item{text}{A character vector of free text.}

\item{top_tag}{The GEDCOM tag of the first line.}

\item{char_limit}{The maximum number of characters allowed on each line. Defaults to 248.}

\item{start_level}{The level of the first line. Defaults to 0.}
}
\value{
A tidy tibble containing the GEDCOM representation of the character vector.
}
\description{
This function only uses the CONC(atenation) tag for splitting free text, and does
not use the CONT(inuation) tag. This is because it is easier to implement.
}
