% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/structures.R
\name{spouse_to_family_link}
\alias{spouse_to_family_link}
\title{Constructs the SPOUSE_TO_FAMILY_LINK from the GEDCOM specification}
\usage{
spouse_to_family_link(family_ref, notes = list())
}
\value{

}
\description{
Constructs the SPOUSE_TO_FAMILY_LINK from the GEDCOM specification
}
