% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interrogate.R
\name{str.tidyged}
\alias{str.tidyged}
\title{Get the structure of a tidyged object}
\usage{
\method{str}{tidyged}(object, ...)
}
\arguments{
\item{object}{A tidyged object.}

\item{...}{Not used.}
}
\value{
A printed summary of records in the tidyged object.
}
\description{
This function gives a breakdown of record counts in the GEDCOM file.
}
