% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interrogate.R
\name{str.tidygedcom}
\alias{str.tidygedcom}
\title{Get the structure of a tidygedcom object}
\usage{
\method{str}{tidygedcom}(object, ...)
}
\arguments{
\item{object}{A tidygedcom object.}

\item{...}{Not used.}
}
\value{
A printed summary of records in the tidygedcom object.
}
\description{
This function gives a breakdown of record counts in the GEDCOM file.
}
