% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create.R
\name{subm}
\alias{subm}
\title{Define a Submitter record for a new tidyged object}
\usage{
subm(
  name = unname(Sys.info()["user"]),
  local_address_lines = character(),
  city = character(),
  state = character(),
  postal_code = character(),
  country = character(),
  phone_number = character(),
  email = character(),
  fax = character(),
  web_page = character(),
  subm_notes = character(),
  multimedia_links = character()
)
}
\arguments{
\item{name}{The name of the submitter.}

\item{local_address_lines}{The first lines of the submitter address.}

\item{city}{The city of the submitter.}

\item{state}{The state/county of the submitter.}

\item{postal_code}{The postal code of the submitter.}

\item{country}{The country of the submitter.}

\item{phone_number}{A character vector containing up to three phone numbers of the submitter.}

\item{email}{A character vector containing up to three email addresses of the submitter.}

\item{fax}{A character vector containing up to three fax numbers of the submitter.}

\item{web_page}{A character vector containing up to three web pages of the submitter.}

\item{subm_notes}{A character vector of notes accompanying this Submitter record.
These could be xrefs to existing Note records.}

\item{multimedia_links}{TODO}
}
\value{
A Submitter record to be incorporated into a new tidyged object.
}
\description{
Define a Submitter record for a new tidyged object
}
\details{
This function is supposed to be used in the gedcom() function to define a
new tidyged object.

This submitter record identifies the individual or organization that contributed
information contained in the GEDCOM file.
}
