% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/records.R
\name{SUBMISSION_RECORD}
\alias{SUBMISSION_RECORD}
\title{Construct the SUBMISSION_RECORD tibble}
\usage{
SUBMISSION_RECORD(
  xref_subn,
  xref_subm = character(),
  name_of_family_file = character(),
  temple_code = character(),
  generations_of_ancestors = character(),
  generations_of_descendants = character(),
  ordinance_process_flag = character(),
  automated_record_id = character(),
  notes = list(),
  date_changed = CHANGE_DATE()
)
}
\arguments{
\item{xref_subn}{An xref ID of a submission record.}

\item{xref_subm}{An xref ID of a submitter record.}

\item{name_of_family_file}{Name under which family names for ordinances are stored in the temple's family file.}

\item{temple_code}{An abbreviation of the temple in which LDS temple ordinances were performed.}

\item{generations_of_ancestors}{The number of generations of ancestors included in this transmission.
This value is usually provided when FamilySearch programs build a GEDCOM file for a patron requesting
a download of ancestors.}

\item{generations_of_descendants}{The number of generations of descendants included in this transmission.
This value is usually provided when FamilySearch programs build a GEDCOM file for a patron requesting
a download of descendants.}

\item{ordinance_process_flag}{A flag that indicates whether submission should be processed for clearing
temple ordinances.}

\item{automated_record_id}{A unique record identification number assigned to the record by
the source system. This number is intended to serve as a more sure means of identification
of a record for reconciling differences in data between two interfacing systems.}

\item{notes}{A list of NOTE_STRUCTURE() objects.}

\item{date_changed}{A CHANGE_DATE() object giving the time this record was last modified. If not provided,
the current date is used.}
}
\value{
A tidy tibble containing a SUBMISSION_RECORD part of a GEDCOM file.
}
\description{
This function constructs a tibble representation of the SUBMISSION_RECORD from the GEDCOM 5.5.1
specification.
}
