% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/records.R
\name{SUBMITTER_RECORD}
\alias{SUBMITTER_RECORD}
\title{Construct the SUBMITTER_RECORD tibble}
\usage{
SUBMITTER_RECORD(
  xref_subm,
  submitter_name,
  address = ADDRESS_STRUCTURE(),
  multimedia_links = list(),
  automated_record_id = character(),
  notes = list(),
  date_changed = CHANGE_DATE()
)
}
\arguments{
\item{xref_subm}{An xref ID of a submitter record.}

\item{submitter_name}{The name of the submitter formatted for display and address generation.}

\item{address}{An ADDRESS_STRUCTURE() object giving address details of the submitter.}

\item{multimedia_links}{A list of MULTIMEDIA_LINK() objects}

\item{automated_record_id}{A unique record identification number assigned to the record by
the source system. This number is intended to serve as a more sure means of identification
of a record for reconciling differences in data between two interfacing systems.}

\item{notes}{A list of NOTE_STRUCTURE() objects.}

\item{date_changed}{A CHANGE_DATE() object giving the time this record was last modified. If not provided,
the current date is used.}
}
\value{
A tidy tibble containing a SUBMITTER_RECORD part of a GEDCOM file.
}
\description{
This function constructs a tibble representation of the SUBMITTER_RECORD from the GEDCOM 5.5.5
specification.
}
