% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarise.R
\name{summary.tidyged}
\alias{summary.tidyged}
\title{Get a summary of a tidyged object}
\usage{
\method{summary}{tidyged}(object, ...)
}
\arguments{
\item{object}{A tidyged object.}

\item{...}{Not used.}
}
\value{
A printed summary of the tidyged object.
}
\description{
This function shows key information from the header of a tidyged object, including submitter
and description.
}
\examples{
summary(sample555)
}
