% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interrogate.R
\name{summary.tidygedcom}
\alias{summary.tidygedcom}
\title{Get a summary of a tidygedcom object}
\usage{
\method{summary}{tidygedcom}(object, ...)
}
\arguments{
\item{object}{A tidygedcom object.}

\item{...}{Not used.}
}
\value{
A printed summary of the tidygedcom object.
}
\description{
This function shows key information from the header of a tidygedcom object, including submitter
and description.
}
