% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{temporarily_remove_name_slashes}
\alias{temporarily_remove_name_slashes}
\title{Temporarily remove forward slashes from surnames}
\usage{
temporarily_remove_name_slashes(gedcom)
}
\arguments{
\item{gedcom}{A tidygedcom object.}
}
\value{
A tidygedcom object with all forward slashes removed from surnames.
}
\description{
Temporarily remove forward slashes from surnames
}
