% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{update_change_date}
\alias{update_change_date}
\title{Update a record's change date}
\usage{
update_change_date(gedcom, xref)
}
\arguments{
\item{gedcom}{A tidyged object.}

\item{xref}{The xref of a record.}
}
\value{
An updated tidyged object where the specified record has a change date of today.
}
\description{
Update a record's change date
}
