% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_write_create.R
\name{update_header_with_filename}
\alias{update_header_with_filename}
\title{Update GEDCOM header with filename}
\usage{
update_header_with_filename(gedcom, filename)
}
\arguments{
\item{gedcom}{A tidygedcom object.}

\item{filename}{The name of the file (with extension).}
}
\value{
An updated tidygedcom object with the updated filename.
}
\description{
Update GEDCOM header with filename
}
