% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate_gedcom.R
\name{validate_gedcom}
\alias{validate_gedcom}
\title{Validate a GEDCOM file}
\usage{
validate_gedcom(gedcom)
}
\arguments{
\item{gedcom}{A tidygedcom object}
}
\value{
Nothing
}
\description{
Conduct some simple (but not exhaustive) checks on a GEDCOM file
}
\details{
This function is called when importing a GEDCOM file. The checks contained within are relatively
simple since there are a wealth of GEDCOM validators already available.
}
