% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate_gedcom.R
\name{validate_header}
\alias{validate_header}
\title{Validate a GEDCOM header}
\usage{
validate_header(gedcom, expected_encoding)
}
\arguments{
\item{gedcom}{A tidygedcom object}

\item{expected_encoding}{A character string given the expected file encoding. One of
"UTF-8", "UTF-16LE", or "UTF-16BE".}
}
\value{
Nothing
}
\description{
Check the GEDCOM header is properly formed.
}
