% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_write_create.R
\name{write_gedcom}
\alias{write_gedcom}
\title{Save a tidygedcom object to disk as a GEDCOM file}
\usage{
write_gedcom(gedcom, filepath)
}
\arguments{
\item{gedcom}{A tidygedcom object.}

\item{filepath}{The full filepath to write to.}
}
\value{
Nothing
}
\description{
Save a tidygedcom object to disk as a GEDCOM file
}
\details{
This function prepares the tidygedcom object and then writes it to the filepath.
Steps taken include escaping "@" signs (with another "@") and splitting long lines onto
separate lines.
}
