% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/identify.R
\name{xrefs_indi}
\alias{xrefs_indi}
\alias{xrefs_famg}
\alias{xrefs_subm}
\alias{xrefs_sour}
\alias{xrefs_repo}
\alias{xrefs_note}
\alias{xrefs_media}
\title{Get the xrefs of particular record types}
\usage{
xrefs_indi(gedcom)

xrefs_famg(gedcom)

xrefs_subm(gedcom)

xrefs_sour(gedcom)

xrefs_repo(gedcom)

xrefs_note(gedcom)

xrefs_media(gedcom)
}
\arguments{
\item{gedcom}{A tidyged object.}
}
\value{
A vector of xrefs of records of the relevant type.
}
\description{
These functions return the xrefs of all records of a particular type in a tidyged object.
}
