% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/activate.R
\name{activate_family_group_record}
\alias{activate_family_group_record}
\title{Activate a Family group record}
\usage{
activate_family_group_record(gedcom, xref)
}
\arguments{
\item{gedcom}{A tidygedcom object}

\item{xref}{The xref of the Family group record to activate}
}
\value{
The same tidygedcom object with "active_record" attribute set to the specific
Family group record to allow easy editing.
}
\description{
This allows a mechanism to easily edit particular records. For Family group records,
the explicit xref is required.
}
