% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/activate.R
\name{activate_individual_record}
\alias{activate_individual_record}
\title{Activate an Individual record}
\usage{
activate_individual_record(gedcom, individual)
}
\arguments{
\item{gedcom}{A tidygedcom object.}

\item{individual}{The xref or name of the Individual record to be activated.}
}
\value{
The same tidygedcom object with "active_record" attribute set to the specific
Individual record to allow easy editing.
}
\description{
This allows a mechanism to easily edit particular records. Either the individual name can be provided
(can be a regex pattern) or the xref can be provided explicitly. Providing the name instead of the
xref can make your code more readable, provided that the name returns a unique record!
}
