% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/activate.R
\name{activate_individual_record}
\alias{activate_individual_record}
\title{Activate an Individual record}
\usage{
activate_individual_record(
  gedcom,
  individual_name = character(),
  xref = character()
)
}
\arguments{
\item{gedcom}{A tidygedcom object}

\item{individual_name}{The name of the individual whose record should be activated}

\item{xref}{The xref of the Individual record to activate}
}
\value{
The same tidygedcom object with "active_record" attribute set to the specific
Individual record to allow easy editing.
}
\description{
This allows a mechanism to easily edit particular records. Either the individual name can be provided
(can be a partial name) or the xref can be provided explicitly. Providing the name instead of the
xref can make your code more readable, provided that the name returns a unique record!
}
