% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/activate.R
\name{activate_source_record}
\alias{activate_source_record}
\title{Activate a Source record}
\usage{
activate_source_record(gedcom, source)
}
\arguments{
\item{gedcom}{A tidyged object.}

\item{source}{The xref or title of the Source record to be activated.}
}
\value{
The same tidyged object with "active_record" attribute set to the specific
Source record to allow easy editing.
}
\description{
This allows a mechanism to easily edit particular records. Either the source title can be provided
or the xref can be provided explicitly. Providing the source title instead of the
xref can make your code more readable, provided that the name returns a unique record!
}
