% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/family_group.R
\name{add_famg}
\alias{add_famg}
\title{Add a Family Group record to a tidyged object}
\usage{
add_famg(
  gedcom,
  husband = character(),
  wife = character(),
  children = character(),
  child_linkage_types = rep("birth", length(children)),
  number_of_children = character(),
  user_reference_number = character(),
  user_reference_type = character(),
  family_notes = character(),
  multimedia_links = character()
)
}
\arguments{
\item{gedcom}{A tidyged object.}

\item{husband}{A character string identifying the husband of this family. This can either
be an xref or a regular expression to match to an individual name.}

\item{wife}{A character string identifying the wife of this family. This can either
be an xref or a regular expression to match to an individual name.}

\item{children}{A character vector of other Individual records that are children of this
family. Children can either be referenced by an xref or by a regular expression
to match to an individual name.}

\item{child_linkage_types}{Codes used to indicate the child to family relationships. If defined,
this must be a character vector the same size as children. Values must be one of:
"birth" (default), "adopted", or "foster".}

\item{number_of_children}{The reported number of children known to belong to this family,
regardless of whether the associated children are represented here.}

\item{user_reference_number}{A unique user-defined number or text that the submitter
uses to identify this record. You can supply more than one in a vector.}

\item{user_reference_type}{A user-defined definition of the user_reference_number(s). If this
parameter is used, there must be a reference type for every reference number defined.}

\item{family_notes}{A character vector of notes accompanying this Family group record. These could be
xrefs to existing Note records.}

\item{multimedia_links}{A character vector of multimedia file references accompanying this
Family group record. These could be xrefs to existing Multimedia records.}
}
\value{
An updated tidyged object including the Family group record.
}
\description{
Add a Family Group record to a tidyged object
}
\details{
If you need to add further information about this family (e.g. events), use the
add_family_event_*() functions.

The function will automatically add links to this family to the respective Individual
records of the wife, husband, and children.
}
