% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/edit_individual.R
\name{add_indi}
\alias{add_indi}
\title{Add an Individual record to a tidyged object}
\usage{
add_indi(
  gedcom,
  sex = "U",
  user_reference_number = character(),
  user_reference_type = character(),
  indi_notes = character(),
  multimedia_links = character(),
  qn = character()
)
}
\arguments{
\item{gedcom}{A tidyged object.}

\item{sex}{The sex of the individual. Either "M" (male), "F" (female), "U" (undetermined),
"X" (intersex), or "N" (not recorded).}

\item{user_reference_number}{A unique user-defined number or text that the submitter
uses to identify this record. You can supply more than one in a vector.}

\item{user_reference_type}{A user-defined definition of the user_reference_number(s). If this
parameter is used, there must be a reference type for every reference number defined.}

\item{indi_notes}{A character vector of notes accompanying this Individual record.
These could be xrefs to existing Note records.}

\item{multimedia_links}{A character vector of multimedia file references accompanying this
Individual record. These could be xrefs to existing Multimedia records.}

\item{qn}{A shortcut to quickly define a name for this individual. This is a shortcut for
the add_indi_names() function (which you should really use instead), but this is useful
for quick demonstrations.}
}
\value{
An updated tidyged object including the Individual record.
}
\description{
Add an Individual record to a tidyged object
}
\details{
If you need to add further information about this individual (e.g. names), use the
add_indi_* functions.
}
