% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/individual_links.R
\name{add_indi_links_to_families}
\alias{add_indi_links_to_families}
\title{Add family links as a child or spouse}
\usage{
add_indi_links_to_families(
  gedcom,
  parents = character(),
  linkage_type = "birth",
  linkage_notes = character(),
  spouse = character(),
  children = character(),
  xref = character(),
  update_date_changed = TRUE
)
}
\arguments{
\item{gedcom}{A tidyged object.}

\item{parents}{A character vector of parents so that this person can be linked to a Family Group
record as a child.}

\item{linkage_type}{A code used to indicate the relationship with the parent(s). Either,
"birth" (default), "foster", or "adopted".}

\item{linkage_notes}{A character vector of notes accompanying this linkage.
These could be xrefs to existing Note records.}

\item{spouse}{A character vector (of length 1) of the spouse so that this person can be linked to
a Family Group record as a spouse.}

\item{children}{A character vector of children so that this person can be linked to a Family Group
record as a spouse.}

\item{xref}{The xref of a record to act on if one is not activated (will override active record).}

\item{update_date_changed}{Whether to add/update the change date for the record.}
}
\value{
An updated tidyged object with an expanded Individual record including
the family link(s) and expanded Family Group record(s) linking to this individual.
}
\description{
This function adds links connecting an Individual record to existing Family Group records. Family
links will be added to the Individual record, and the Family Group records will be updated to link
to this individual.
}
