% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/edit_individual_names.R
\name{add_indi_names}
\alias{add_indi_names}
\title{Add a personal name (and components) to an Individual record}
\usage{
add_indi_names(
  gedcom,
  name,
  type = character(),
  prefix = character(),
  given = character(),
  nickname = character(),
  surname_prefix = character(),
  surname = character(),
  suffix = character(),
  name_notes = character(),
  xref = character(),
  update_date_changed = TRUE
)
}
\arguments{
\item{gedcom}{A tidyged object.}

\item{name}{The full name of the individual. The order of the name parts should
be the order that the person would, by custom of their culture, have used when
giving it to a recorder. The surname, if known, should be enclosed between two
forward slash (/) characters.}

\item{type}{The name type, e.g. "birth", "aka", "maiden".}

\item{prefix}{The name prefix, e.g. Cmdr.}

\item{given}{The given name or earned name. Different given names are separated
by a comma.}

\item{nickname}{A descriptive or familiar name used in connection with one's
proper name.}

\item{surname_prefix}{Surname prefix or article used in a family name.
Different surname articles are separated by a comma, for example in the name
"de la Cruz", this value would be "de, la".}

\item{surname}{Surname or family name. Different surnames are separated by a comma.}

\item{suffix}{Non-indexing name piece that appears after the given name and surname
parts, e.g. Jr. Different name suffix parts are separated by a comma.}

\item{name_notes}{A character vector of notes accompanying this name.
These could be xrefs to existing Note records.}

\item{xref}{The xref of a record to act on if one is not activated (will override active record).}

\item{update_date_changed}{Whether to add/update the change date for the record.}
}
\value{
An updated tidyged object with an expanded Individual record including
these names.
}
\description{
This function can be applied to an Individual record several times to record
personal names.
}
