% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/individual_names.R
\name{add_indi_names_var}
\alias{add_indi_names_var}
\title{Add a variation of a personal name to an Individual record}
\usage{
add_indi_names_var(
  gedcom,
  primary_name,
  method,
  prefix = character(),
  given = character(),
  nickname = character(),
  surname_prefix = character(),
  surname = character(),
  suffix = character(),
  phonetic_variation = TRUE,
  variation_notes = character(),
  xref = character(),
  update_date_changed = TRUE
)
}
\arguments{
\item{gedcom}{A tidyged object.}

\item{primary_name}{The name for which this is a variation. This is treated as a
regex pattern to match to existing names.}

\item{method}{Indicates the method used in transforming the text to the variation.}

\item{prefix}{The name prefix, e.g. Cmdr.}

\item{given}{The given name or earned name. Different given names are separated
by a comma.}

\item{nickname}{A descriptive or familiar name used in connection with one's
proper name.}

\item{surname_prefix}{Surname prefix or article used in a family name.
For example in the name "de la Cruz", this value would be "de la".}

\item{surname}{Surname or family name. Different surnames are separated by a comma.}

\item{suffix}{Non-indexing name piece that appears after the given name and surname
parts, e.g. Jr. Different name suffix parts are separated by a comma.}

\item{phonetic_variation}{Whether the name variation is a phonetic variation
(TRUE, default) or a romanised variation (FALSE).}

\item{variation_notes}{A character vector of notes accompanying this name variation.
These could be xrefs to existing Note records.}

\item{xref}{The xref of a record to act on if one is not activated (will override active record).}

\item{update_date_changed}{Whether to add/update the change date for the record.}
}
\value{
An updated tidyged object with an expanded Individual record including
these name variants.
}
\description{
Add a variation of a personal name to an Individual record
}
