% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/edit_individual.R
\name{add_individual}
\alias{add_individual}
\title{Add an Individual record to a tidyged object}
\usage{
add_individual(
  gedcom,
  quick_name = character(),
  sex = "U",
  user_reference_number = character(),
  user_reference_type = character(),
  individual_notes = character(),
  multimedia_links = character()
)
}
\arguments{
\item{gedcom}{A tidyged object.}

\item{quick_name}{A shortcut to quickly define a name for this individual. This is a shortcut for
the add_individual_names() function (which you should really use instead), but this is useful
for quick demonstrations.}

\item{sex}{The sex of the individual. Either "M" (male), "F" (female), "U" (undetermined),
"X" (intersex), or "N" (not recorded).}

\item{user_reference_number}{A unique user-defined number or text that the submitter
uses to identify this record. You can supply more than one in a vector.}

\item{user_reference_type}{A user-defined definition of the user_reference_number(s). If this
parameter is used, there must be a reference type for every reference number defined.}

\item{individual_notes}{A character vector of notes accompanying this Individual record.
These could be xrefs to existing Note records.}

\item{multimedia_links}{A character vector of multimedia file references accompanying this
Individual record. These could be xrefs to existing Multimedia records.}
}
\value{
An updated tidyged object including the Individual record.
}
\description{
Add an Individual record to a tidyged object
}
\details{
If you need to add further information about this individual (e.g. names), use the
add_individual_* functions.
}
