% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/edit_individual.R
\name{add_individual}
\alias{add_individual}
\title{Add an Individual record to a tidyged object}
\usage{
add_individual(
  gedcom,
  sex = "U",
  user_reference_number = character(),
  user_reference_type = character(),
  automated_record_id = character(),
  individual_notes = character(),
  multimedia_links = character()
)
}
\arguments{
\item{gedcom}{A tidyged object.}

\item{sex}{The sex of the individual. Either "M" (male), "F" (female), "U" (undetermined),
"X" (intersex), or "N" (not recorded).}

\item{user_reference_number}{A user-defined number or text that the submitter uses to identify
this record. See the Gedcom 5.5.5 Specification for more details.}

\item{user_reference_type}{A user-defined definition of the user_reference_number.}

\item{automated_record_id}{A unique record identification number assigned to the record by
the source system.}

\item{individual_notes}{A character vector of notes accompanying this Individual record.
These could be xrefs to existing Note records.}

\item{multimedia_links}{A character vector of multimedia file references accompanying this
Individual record. These could be xrefs to existing Multimedia records.}
}
\value{
An updated tidyged object including the Individual record.
}
\description{
This function adds a new record for an individual.
}
\details{
This function will automatically assign a unique xref for this record. Most users
will only need to use the sex and individual_notes parameters (and of course gedcom).

If you need to add further information about this individual (e.g. names), use the
add_individual_* functions.
}
