% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/edit_individual_attributes.R
\name{add_individual_attribute}
\alias{add_individual_attribute}
\alias{add_individual_attribute_caste}
\alias{add_individual_attribute_phys_descr}
\alias{add_individual_attribute_education}
\alias{add_individual_attribute_national_id}
\alias{add_individual_attribute_nationality}
\alias{add_individual_attribute_num_children}
\alias{add_individual_attribute_num_marriages}
\alias{add_individual_attribute_occupation}
\alias{add_individual_attribute_possessions}
\alias{add_individual_attribute_religion}
\alias{add_individual_attribute_residence}
\alias{add_individual_attribute_social_sec_num}
\alias{add_individual_attribute_nobility_title}
\alias{add_individual_attribute_other}
\title{Add an attribute associated with an individual}
\usage{
add_individual_attribute(
  gedcom,
  attribute_type,
  attribute_descriptor,
  fact_classification = character(),
  event_date = date_value(),
  event_cause = character(),
  age_at_event = character(),
  event_notes = character(),
  place_name = character(),
  place_hierarchy = character(),
  place_phonetic_variation = character(),
  phonetic_type = character(),
  place_romanized_variation = character(),
  romanized_type = character(),
  place_latitude = character(),
  place_longitude = character(),
  place_notes = character(),
  address_first_line = character(),
  city = character(),
  state = character(),
  postal_code = character(),
  country = character(),
  phone_number = character(),
  email = character(),
  fax = character(),
  web_page = character(),
  responsible_agency = character(),
  religious_affiliation = character(),
  restriction_notice = character()
)

add_individual_attribute_caste(
  gedcom,
  attribute_type = "CAST",
  attribute_descriptor,
  fact_classification = character(),
  event_date = date_value(),
  event_cause = character(),
  age_at_event = character(),
  event_notes = character(),
  place_name = character(),
  place_hierarchy = character(),
  place_phonetic_variation = character(),
  phonetic_type = character(),
  place_romanized_variation = character(),
  romanized_type = character(),
  place_latitude = character(),
  place_longitude = character(),
  place_notes = character(),
  address_first_line = character(),
  city = character(),
  state = character(),
  postal_code = character(),
  country = character(),
  phone_number = character(),
  email = character(),
  fax = character(),
  web_page = character(),
  responsible_agency = character(),
  religious_affiliation = character(),
  restriction_notice = character()
)

add_individual_attribute_phys_descr(
  gedcom,
  attribute_type = "DSCR",
  attribute_descriptor,
  fact_classification = character(),
  event_date = date_value(),
  event_cause = character(),
  age_at_event = character(),
  event_notes = character(),
  place_name = character(),
  place_hierarchy = character(),
  place_phonetic_variation = character(),
  phonetic_type = character(),
  place_romanized_variation = character(),
  romanized_type = character(),
  place_latitude = character(),
  place_longitude = character(),
  place_notes = character(),
  address_first_line = character(),
  city = character(),
  state = character(),
  postal_code = character(),
  country = character(),
  phone_number = character(),
  email = character(),
  fax = character(),
  web_page = character(),
  responsible_agency = character(),
  religious_affiliation = character(),
  restriction_notice = character()
)

add_individual_attribute_education(
  gedcom,
  attribute_type = "EDUC",
  attribute_descriptor,
  fact_classification = character(),
  event_date = date_value(),
  event_cause = character(),
  age_at_event = character(),
  event_notes = character(),
  place_name = character(),
  place_hierarchy = character(),
  place_phonetic_variation = character(),
  phonetic_type = character(),
  place_romanized_variation = character(),
  romanized_type = character(),
  place_latitude = character(),
  place_longitude = character(),
  place_notes = character(),
  address_first_line = character(),
  city = character(),
  state = character(),
  postal_code = character(),
  country = character(),
  phone_number = character(),
  email = character(),
  fax = character(),
  web_page = character(),
  responsible_agency = character(),
  religious_affiliation = character(),
  restriction_notice = character()
)

add_individual_attribute_national_id(
  gedcom,
  attribute_type = "IDNO",
  attribute_descriptor,
  fact_classification = character(),
  event_date = date_value(),
  event_cause = character(),
  age_at_event = character(),
  event_notes = character(),
  place_name = character(),
  place_hierarchy = character(),
  place_phonetic_variation = character(),
  phonetic_type = character(),
  place_romanized_variation = character(),
  romanized_type = character(),
  place_latitude = character(),
  place_longitude = character(),
  place_notes = character(),
  address_first_line = character(),
  city = character(),
  state = character(),
  postal_code = character(),
  country = character(),
  phone_number = character(),
  email = character(),
  fax = character(),
  web_page = character(),
  responsible_agency = character(),
  religious_affiliation = character(),
  restriction_notice = character()
)

add_individual_attribute_nationality(
  gedcom,
  attribute_type = "NATI",
  attribute_descriptor,
  fact_classification = character(),
  event_date = date_value(),
  event_cause = character(),
  age_at_event = character(),
  event_notes = character(),
  place_name = character(),
  place_hierarchy = character(),
  place_phonetic_variation = character(),
  phonetic_type = character(),
  place_romanized_variation = character(),
  romanized_type = character(),
  place_latitude = character(),
  place_longitude = character(),
  place_notes = character(),
  address_first_line = character(),
  city = character(),
  state = character(),
  postal_code = character(),
  country = character(),
  phone_number = character(),
  email = character(),
  fax = character(),
  web_page = character(),
  responsible_agency = character(),
  religious_affiliation = character(),
  restriction_notice = character()
)

add_individual_attribute_num_children(
  gedcom,
  attribute_type = "NCHI",
  attribute_descriptor,
  fact_classification = character(),
  event_date = date_value(),
  event_cause = character(),
  age_at_event = character(),
  event_notes = character(),
  place_name = character(),
  place_hierarchy = character(),
  place_phonetic_variation = character(),
  phonetic_type = character(),
  place_romanized_variation = character(),
  romanized_type = character(),
  place_latitude = character(),
  place_longitude = character(),
  place_notes = character(),
  address_first_line = character(),
  city = character(),
  state = character(),
  postal_code = character(),
  country = character(),
  phone_number = character(),
  email = character(),
  fax = character(),
  web_page = character(),
  responsible_agency = character(),
  religious_affiliation = character(),
  restriction_notice = character()
)

add_individual_attribute_num_marriages(
  gedcom,
  attribute_type = "NMR",
  attribute_descriptor,
  fact_classification = character(),
  event_date = date_value(),
  event_cause = character(),
  age_at_event = character(),
  event_notes = character(),
  place_name = character(),
  place_hierarchy = character(),
  place_phonetic_variation = character(),
  phonetic_type = character(),
  place_romanized_variation = character(),
  romanized_type = character(),
  place_latitude = character(),
  place_longitude = character(),
  place_notes = character(),
  address_first_line = character(),
  city = character(),
  state = character(),
  postal_code = character(),
  country = character(),
  phone_number = character(),
  email = character(),
  fax = character(),
  web_page = character(),
  responsible_agency = character(),
  religious_affiliation = character(),
  restriction_notice = character()
)

add_individual_attribute_occupation(
  gedcom,
  attribute_type = "OCCU",
  attribute_descriptor,
  fact_classification = character(),
  event_date = date_value(),
  event_cause = character(),
  age_at_event = character(),
  event_notes = character(),
  place_name = character(),
  place_hierarchy = character(),
  place_phonetic_variation = character(),
  phonetic_type = character(),
  place_romanized_variation = character(),
  romanized_type = character(),
  place_latitude = character(),
  place_longitude = character(),
  place_notes = character(),
  address_first_line = character(),
  city = character(),
  state = character(),
  postal_code = character(),
  country = character(),
  phone_number = character(),
  email = character(),
  fax = character(),
  web_page = character(),
  responsible_agency = character(),
  religious_affiliation = character(),
  restriction_notice = character()
)

add_individual_attribute_possessions(
  gedcom,
  attribute_type = "PROP",
  attribute_descriptor,
  fact_classification = character(),
  event_date = date_value(),
  event_cause = character(),
  age_at_event = character(),
  event_notes = character(),
  place_name = character(),
  place_hierarchy = character(),
  place_phonetic_variation = character(),
  phonetic_type = character(),
  place_romanized_variation = character(),
  romanized_type = character(),
  place_latitude = character(),
  place_longitude = character(),
  place_notes = character(),
  address_first_line = character(),
  city = character(),
  state = character(),
  postal_code = character(),
  country = character(),
  phone_number = character(),
  email = character(),
  fax = character(),
  web_page = character(),
  responsible_agency = character(),
  religious_affiliation = character(),
  restriction_notice = character()
)

add_individual_attribute_religion(
  gedcom,
  attribute_type = "RELI",
  attribute_descriptor,
  fact_classification = character(),
  event_date = date_value(),
  event_cause = character(),
  age_at_event = character(),
  event_notes = character(),
  place_name = character(),
  place_hierarchy = character(),
  place_phonetic_variation = character(),
  phonetic_type = character(),
  place_romanized_variation = character(),
  romanized_type = character(),
  place_latitude = character(),
  place_longitude = character(),
  place_notes = character(),
  address_first_line = character(),
  city = character(),
  state = character(),
  postal_code = character(),
  country = character(),
  phone_number = character(),
  email = character(),
  fax = character(),
  web_page = character(),
  responsible_agency = character(),
  religious_affiliation = character(),
  restriction_notice = character()
)

add_individual_attribute_residence(
  gedcom,
  attribute_type = "RESI",
  attribute_descriptor,
  fact_classification = character(),
  event_date = date_value(),
  event_cause = character(),
  age_at_event = character(),
  event_notes = character(),
  place_name = character(),
  place_hierarchy = character(),
  place_phonetic_variation = character(),
  phonetic_type = character(),
  place_romanized_variation = character(),
  romanized_type = character(),
  place_latitude = character(),
  place_longitude = character(),
  place_notes = character(),
  address_first_line = character(),
  city = character(),
  state = character(),
  postal_code = character(),
  country = character(),
  phone_number = character(),
  email = character(),
  fax = character(),
  web_page = character(),
  responsible_agency = character(),
  religious_affiliation = character(),
  restriction_notice = character()
)

add_individual_attribute_social_sec_num(
  gedcom,
  attribute_type = "SSN",
  attribute_descriptor,
  fact_classification = character(),
  event_date = date_value(),
  event_cause = character(),
  age_at_event = character(),
  event_notes = character(),
  place_name = character(),
  place_hierarchy = character(),
  place_phonetic_variation = character(),
  phonetic_type = character(),
  place_romanized_variation = character(),
  romanized_type = character(),
  place_latitude = character(),
  place_longitude = character(),
  place_notes = character(),
  address_first_line = character(),
  city = character(),
  state = character(),
  postal_code = character(),
  country = character(),
  phone_number = character(),
  email = character(),
  fax = character(),
  web_page = character(),
  responsible_agency = character(),
  religious_affiliation = character(),
  restriction_notice = character()
)

add_individual_attribute_nobility_title(
  gedcom,
  attribute_type = "TITL",
  attribute_descriptor,
  fact_classification = character(),
  event_date = date_value(),
  event_cause = character(),
  age_at_event = character(),
  event_notes = character(),
  place_name = character(),
  place_hierarchy = character(),
  place_phonetic_variation = character(),
  phonetic_type = character(),
  place_romanized_variation = character(),
  romanized_type = character(),
  place_latitude = character(),
  place_longitude = character(),
  place_notes = character(),
  address_first_line = character(),
  city = character(),
  state = character(),
  postal_code = character(),
  country = character(),
  phone_number = character(),
  email = character(),
  fax = character(),
  web_page = character(),
  responsible_agency = character(),
  religious_affiliation = character(),
  restriction_notice = character()
)

add_individual_attribute_other(
  gedcom,
  attribute_type = "FACT",
  attribute_descriptor,
  fact_classification = character(),
  event_date = date_value(),
  event_cause = character(),
  age_at_event = character(),
  event_notes = character(),
  place_name = character(),
  place_hierarchy = character(),
  place_phonetic_variation = character(),
  phonetic_type = character(),
  place_romanized_variation = character(),
  romanized_type = character(),
  place_latitude = character(),
  place_longitude = character(),
  place_notes = character(),
  address_first_line = character(),
  city = character(),
  state = character(),
  postal_code = character(),
  country = character(),
  phone_number = character(),
  email = character(),
  fax = character(),
  web_page = character(),
  responsible_agency = character(),
  religious_affiliation = character(),
  restriction_notice = character()
)
}
\arguments{
\item{gedcom}{A tidygedcom object.}

\item{attribute_type}{The type of attribute. This should be automatically populated with the appropriate
attribute function.}

\item{attribute_descriptor}{The value of this attribute.}

\item{fact_classification}{A descriptive word or phrase used to further classify this
attribute. This should be used whenever the 'other' attribute is used (but can also be used
with others).}

\item{event_date}{A date_value() object giving the date of the event.}

\item{event_cause}{Used in special cases to record the reasons which precipitated an event.
Normally this will be used for a death event to show cause of death, such as might be listed
on a death certificate.}

\item{age_at_event}{A character string that indicates the age in years, months, and days
that the individual was at the time of the event. Any combination of these is permitted.
Any labels must come after their corresponding number, for example; "4y 8m 10d".
The string can also be "CHILD", "INFANT", or "STILLBORN".}

\item{event_notes}{A character vector of notes accompanying the event.
These could be xrefs to existing Note records.}

\item{place_name}{The jurisdictional name of the place where the event took place.
Jurisdictions are separated by commas, for example, "Cove, Cache, Utah, USA."}

\item{place_hierarchy}{This shows the jurisdictional entities that are named in a sequence
from the lowest to the highest jurisdiction. The jurisdictions are separated by commas,
and any jurisdiction's name that is missing is still accounted for by a comma.}

\item{place_phonetic_variation}{A character vector of phonetic variations of the place name.}

\item{phonetic_type}{A character vector giving the method used in transforming the text to
the corresponding phonetic variation. If this argument is used, it must be the same size
as the place_phonetic_variation argument.}

\item{place_romanized_variation}{A character vector of romanized variations of the place name.}

\item{romanized_type}{A character vector giving the method used in transforming the text to
the corresponding romanized variation. If this argument is used, it must be the same size
as the place_romanized_variation argument.}

\item{place_latitude}{The value specifying the latitudinal coordinate of the event place.
The latitude coordinate is the direction North or South from the equator in degrees and
fraction of degrees carried out to give the desired accuracy.
For example: 18 degrees, 9 minutes, and 3.4 seconds North would be formatted as "N18.150944"}

\item{place_longitude}{The value specifying the longitudinal coordinate of the event place.
The longitude coordinate is Degrees and fraction of degrees east or west of the zero or
base meridian coordinate. For example:
168 degrees, 9 minutes, and 3.4 seconds East would be formatted as "E168.150944".}

\item{place_notes}{A character vector of notes accompanying the event place.
These could be xrefs to existing Note records.}

\item{address_first_line}{The first line of the event address.}

\item{city}{The city of the event address.}

\item{state}{The state/county of the event address.}

\item{postal_code}{The postal code of the event address.}

\item{country}{The country of the event address.}

\item{phone_number}{A character vector containing up to three phone numbers of the event address.}

\item{email}{A character vector containing up to three email addresses of the event address.}

\item{fax}{A character vector containing up to three fax numbers of the event address.}

\item{web_page}{A character vector containing up to three web pages of the event address.}

\item{responsible_agency}{The organisation, institution, corporation, person, or other
entity that has responsibility for the event data.}

\item{religious_affiliation}{A name of the religion with which this event was affiliated.}

\item{restriction_notice}{Only for Ancestral File usage. See the Gedcom 5.5.1 Standard for more
details.}
}
\value{
An updated tidygedcom object with an expanded Individual record including
this attribute.
}
\description{
Add an attribute associated with an individual
}
