% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/edit_multimedia.R
\name{add_media}
\alias{add_media}
\title{Add a Multimedia record to a tidyged object}
\usage{
add_media(
  gedcom,
  file_reference,
  format,
  source_media = character(),
  title = character(),
  user_reference_number = character(),
  user_reference_type = character(),
  media_notes = character()
)
}
\arguments{
\item{gedcom}{A tidyged object.}

\item{file_reference}{A reference for the file, typically a filepath or URL.}

\item{format}{A string indicating the format of the multimedia file.
Currently limited to one of: "AAC", "AVI", "BMP", "ePub", "FLAC", "GIF", "JPEG", "JPG",
"MKV", "mobi", "MP3", "PCX", "PDF", "PNG", "TIFF", "TIF", "WAV".}

\item{source_media}{A code that indicates the type of material in which the referenced
source is stored. Must be one of: audio, book, card, electronic, fiche, film, magazine,
manuscript, map, newspaper, photo, tombstone, video.}

\item{title}{The title of the multimedia file.}

\item{user_reference_number}{A unique user-defined number or text that the submitter
uses to identify this record. You can supply more than one in a vector.}

\item{user_reference_type}{A user-defined definition of the user_reference_number(s). If this
parameter is used, there must be a reference type for every reference number defined.}

\item{media_notes}{A character vector of notes accompanying this Multimedia record. These could be
xrefs to existing Note records.}
}
\value{
An updated tidyged object including the Multimedia record.
}
\description{
Add a Multimedia record to a tidyged object
}
