% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/edit_note.R
\name{add_note}
\alias{add_note}
\title{Add a Note record to a tidyged object}
\usage{
add_note(
  gedcom,
  text,
  user_reference_number = character(),
  user_reference_type = character()
)
}
\arguments{
\item{gedcom}{A tidyged object.}

\item{text}{A character string containing the text of the note.}

\item{user_reference_number}{A unique user-defined number or text that the submitter
uses to identify this record. You can supply more than one in a vector.}

\item{user_reference_type}{A user-defined definition of the user_reference_number(s). If this
parameter is used, there must be a reference type for every reference number defined.}
}
\value{
An updated tidyged object including the Note record.
}
\description{
Add a Note record to a tidyged object
}
