% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/note.R
\name{add_note}
\alias{add_note}
\title{Add a Note record to a tidyged object}
\usage{
add_note(gedcom, text, user_reference_numbers = character())
}
\arguments{
\item{gedcom}{A tidyged object.}

\item{text}{A character string containing the text of the note.}

\item{user_reference_numbers}{A unique user-defined number or text that the submitter
uses to identify this record. You can supply more than one in a vector. You can also define a
user reference type by using a named vector (e.g c(id1 = "123A", id2 = "456B")).}
}
\value{
An updated tidyged object including the Note record.
}
\description{
Add a Note record to a tidyged object
}
