% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/repository.R
\name{add_repo}
\alias{add_repo}
\title{Add a Repository record to a tidyged object}
\usage{
add_repo(
  gedcom,
  name,
  repo_address = address(),
  user_reference_number = character(),
  user_reference_type = character(),
  repo_notes = character()
)
}
\arguments{
\item{gedcom}{A tidyged object.}

\item{name}{The name of the repository.}

\item{user_reference_number}{A unique user-defined number or text that the submitter
uses to identify this record. You can supply more than one in a vector.}

\item{user_reference_type}{A user-defined definition of the user_reference_number(s). If this
parameter is used, there must be a reference type for every reference number defined.}

\item{repo_notes}{A character vector of notes accompanying this Repository record.
These could be xrefs to existing Note records.}

\item{address}{An address() object giving the address of the repository.}
}
\value{
An updated tidyged object including the Repository record.
}
\description{
Add a Repository record to a tidyged object
}
