% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/edit_source.R
\name{add_source_repository_citation}
\alias{add_source_repository_citation}
\title{Add a source repository citation to a Source record}
\usage{
add_source_repository_citation(
  gedcom,
  repository,
  call_number = character(),
  update_date_changed = TRUE
)
}
\arguments{
\item{gedcom}{A tidygedcom object.}

\item{repository}{A character string identifying the repository. This can either
be an xref or a regular expression to match to a repository name.}

\item{call_number}{An identification or reference description used to file
and retrieve items from the holdings of a repository.}

\item{update_date_changed}{Whether to add/update the change date for the record.}
}
\value{
An updated tidygedcom object with an expanded Source record including
this repository citation.
}
\description{
This structure is used within a source record to point to a name and
address record of the holder of the source document.
}
