% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/structures.R
\name{association_structure}
\alias{association_structure}
\title{Construct the ASSOCIATION_STRUCTURE from the GEDCOM specification}
\usage{
association_structure(
  xref_indi,
  relation_is_descriptor,
  source_citations = list(),
  notes = list()
)
}
\arguments{
\item{xref_indi}{An xref ID of an individual record.}

\item{relation_is_descriptor}{A word or phrase that states object 1's relation is object 2.}

\item{source_citations}{A list of source_citation() objects.}

\item{notes}{A list of note_structure() objects.}
}
\value{
A tidy tibble containing the ASSOCIATION_STRUCTURE part of a GEDCOM file.
}
\description{
Construct the ASSOCIATION_STRUCTURE from the GEDCOM specification
}
