% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/structures.R
\name{change_date}
\alias{change_date}
\title{Construct the CHANGE_DATE from the GEDCOM specification}
\usage{
change_date(
  change_date = date_exact(),
  time_value = character(),
  notes = list()
)
}
\arguments{
\item{change_date}{A date_exact() object giving the date that this data was changed.}

\item{time_value}{The time of a specific event.}

\item{notes}{A list of note_structure() objects.}
}
\value{
A tidy tibble containing the CHANGE_DATE part of a GEDCOM file.
}
\description{
Construct the CHANGE_DATE from the GEDCOM specification
}
