% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/structures.R
\name{change_date}
\alias{change_date}
\title{Constructs the CHANGE_DATE from the GEDCOM specification}
\usage{
change_date(
  change_date = date_exact(),
  time_value = character(),
  notes = list()
)
}
\arguments{
\item{change_date}{The date that this data was changed.}

\item{time_value}{The time of a specific event.}
}
\value{

}
\description{
Constructs the CHANGE_DATE from the GEDCOM specification
}
