% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/structures.R
\name{child_to_family_link}
\alias{child_to_family_link}
\title{Constructs the CHILD_TO_FAMILY_LINK from the GEDCOM specification}
\usage{
child_to_family_link(
  xref_fam,
  pedigree_linkage_type = character(),
  child_linkage_status = character(),
  notes = list()
)
}
\arguments{
\item{xref_fam}{Either a pointer or an unique cross-reference identifier.}

\item{pedigree_linkage_type}{A code used to indicate the child to family relationship for pedigree navigation
purposes.}

\item{child_linkage_status}{A status code that allows passing on the users opinion of the
status of a child to family link.
"challenged": Linking this child to this family is suspect, but the linkage has been
neither proven nor disproven.
"disproven": There has been a claim by some that this child belongs to this family, but
the linkage has been disproven.
"proven" = There has been a claim by some that this child does not belongs to this family,
but the linkage has been proven.}
}
\value{
A tidy tibble containing the CHILD_TO_FAMILY_LINK part of a GEDCOM file
}
\description{
Constructs the CHILD_TO_FAMILY_LINK from the GEDCOM specification
}
