% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{construct_full_name}
\alias{construct_full_name}
\title{Construct a full personal name}
\usage{
construct_full_name(
  given = character(),
  surname_prefix = character(),
  surname = character(),
  suffix = character()
)
}
\arguments{
\item{given}{The given name(s).}

\item{surname_prefix}{The surname prefix.}

\item{surname}{The surname.}

\item{suffix}{The name suffix.}
}
\value{
The full name with all name pieces combined.
}
\description{
This function constructs a full personal name from individual name pieces.
}
