% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interrogate.R
\name{describe_family_group}
\alias{describe_family_group}
\title{Get a description of a family group}
\usage{
describe_family_group(gedcom, xref)
}
\arguments{
\item{gedcom}{A tidyged object.}

\item{xref}{An xref of a Family group record.}
}
\value{
A character string describing the members of the family group.
}
\description{
Get a description of a family group
}
