% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/structures.R
\name{event_detail}
\alias{event_detail}
\title{Constructs the EVENT_DETAIL from the GEDCOM specification}
\usage{
event_detail(
  event_or_fact_classification = character(),
  date = date_value(),
  place = place_structure(character()),
  address = address_structure(character()),
  responsible_agency = character(),
  religious_affiliation = character(),
  cause_of_event = character(),
  restriction_notice = character(),
  notes = list(),
  source_citations = list(),
  multimedia_links = list()
)
}
\arguments{
\item{event_or_fact_classification}{A descriptive word or phrase used to further classify the parent
event or attribute tag. This should be used whenever either of the generic EVEN or FACT tags are used.
The value of this primative is responsible for classifying the generic event or fact being cited.}

\item{date}{A date_value() object giving the date of the event.}

\item{place}{A place_structure() object giving the location of the event.}

\item{address}{An address_structure() object giving the address of the event.}

\item{responsible_agency}{The organization, institution, corporation, person, or other entity that has
responsibility for the associated context. For example, an employer of a person of an associated occupation,
or a church that administered rites or events, or an organization responsible for creating and/or archiving
records.}

\item{religious_affiliation}{A name of the religion with which this person, event, or record was affiliated.}

\item{cause_of_event}{Used in special cases to record the reasons which precipitated an event.
Normally this will be used subordinate to a death event to show cause of death, such as
might be listed on a death certificate.}

\item{restriction_notice}{The restriction notice is defined for Ancestral File usage. Ancestral File
download GEDCOM files may contain this data.}

\item{notes}{A list of note_structure() objects.}

\item{source_citations}{A list of source_citation() objects.}

\item{multimedia_links}{A list of multimedia_link() objects}
}
\value{
A tidy tibble containing the EVENT_DETAIL part of a GEDCOM file
}
\description{
Constructs the EVENT_DETAIL from the GEDCOM specification
}
