% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/structures.R
\name{family_event_detail}
\alias{family_event_detail}
\title{Constructs the FAMILY_EVENT_DETAIL from the GEDCOM specification}
\usage{
family_event_detail(
  husband_age_at_event = character(),
  wife_age_at_event = character(),
  event_details = event_detail()
)
}
\value{

}
\description{
Constructs the FAMILY_EVENT_DETAIL from the GEDCOM specification
}
