% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/structures.R
\name{FAMILY_EVENT_STRUCTURE}
\alias{FAMILY_EVENT_STRUCTURE}
\title{Construct the FAMILY_EVENT_STRUCTURE tibble}
\usage{
FAMILY_EVENT_STRUCTURE(
  event_type_family,
  event_descriptor = character(),
  family_event_details = FAMILY_EVENT_DETAIL()
)
}
\arguments{
\item{event_type_family}{A code used to indicate the type of family event.}

\item{event_descriptor}{Text describing a particular event pertaining to the individual or family.
This event value is usually assigned to the EVEN tag. The classification as to the difference
between this specific event and other occurrences of the EVENt tag is indicated by the use of
a subordinate TYPE tag selected from the EVENT_DETAIL structure.}

\item{family_event_details}{A FAMILY_EVENT_DETAIL() object giving details of the event.}
}
\value{
A tidy tibble containing the FAMILY_EVENT_STRUCTURE part of a GEDCOM file.
}
\description{
This function constructs a tibble representation of the FAMILY_EVENT_STRUCTURE from the GEDCOM 5.5.1
specification.
}
