% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{find_insertion_point}
\alias{find_insertion_point}
\title{Find a particular row position in a tidyged object.}
\usage{
find_insertion_point(
  gedcom,
  xref,
  parent_level,
  parent_tag,
  parent_value = NULL
)
}
\arguments{
\item{gedcom}{A tidyged object.}

\item{xref}{The xref of the record where the insertion point will be.}

\item{parent_level}{The level of the row where the insertion point will be.}

\item{parent_tag}{The tag of the row where the insertion point will be.}

\item{parent_value}{The value of the row where the insertion point will be.}
}
\value{
The row after the insertion point in the tidyged object.
}
\description{
Find a particular row position in a tidyged object.
}
