% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/activate.R
\name{find_xref}
\alias{find_xref}
\title{Find an xref of a record given a particular value.}
\usage{
find_xref(gedcom, record_ids, tags, search_string, allow_broken = FALSE)
}
\arguments{
\item{gedcom}{A tidygedcom object}

\item{record_ids}{A list of potential xrefs to consider}

\item{tags}{The tags to look at when comparing values}

\item{search_string}{The identifying information to search for}

\item{allow_broken}{Whether to allow the search_string to be matched piece-wise or exactly.
For example, if TRUE, "Joe Bloggs" will match "Joe Ian Bloggs".}
}
\value{
A single xref for the given record
}
\description{
This is a helper function to identify the xref of a record given a piece of information such
as a name or reference number.
}
