% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{gedcom_value}
\alias{gedcom_value}
\title{Extract a particular value from a tidygedcom object}
\usage{
gedcom_value(gedcom, record_xref, tag, level, after_tag = NULL)
}
\arguments{
\item{gedcom}{A tidygedcom object}

\item{record_xref}{The xref of the record in which the value may exist}

\item{tag}{The tag associated with the value}

\item{level}{The level number of the value}

\item{after_tag}{Whether the tag should be subordinate to this parent tag}
}
\value{
The particular value fitting the criteria of the input arguments. If no value is found,
an empty string is returned.
}
\description{
Extract a particular value from a tidygedcom object
}
