% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/identify.R
\name{get_ancestors}
\alias{get_ancestors}
\title{Identify all ancestors for an individual}
\usage{
get_ancestors(
  gedcom,
  individual = character(),
  include_individual = FALSE,
  include_siblings = FALSE,
  include_families = FALSE
)
}
\arguments{
\item{gedcom}{A tidyged object.}

\item{individual}{The xref or name of an Individual record to act on if one
is not activated (will override active record).}

\item{include_individual}{Whether to also include the individual themselves.}

\item{include_siblings}{Whether to also include all siblings of ancestors (siblings of this individual will only be
included if the individual is included).}

\item{include_families}{Whether to also include all Family Group records where this individual is a child
(and all ancestors' Family Group records).}
}
\value{
A vector of xrefs of ancestors.
}
\description{
This function identifies records in an entire branch of the family tree above a certain individual.
}
