% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/identify.R
\name{get_spouses}
\alias{get_spouses}
\title{Get all spouses for an individual}
\usage{
get_spouses(gedcom, individual = character(), return_name = FALSE)
}
\arguments{
\item{gedcom}{A tidyged object.}

\item{individual}{The xref or name of an Individual record to act on if one
is not activated (will override active record).}

\item{return_name}{Whether to return the spouse's name(s) instead of the xref(s).}
}
\value{
A character vector of spouse xrefs or names.
}
\description{
Get all spouses for an individual
}
