% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/structures.R
\name{INDIVIDUAL_ATTRIBUTE_STRUCTURE}
\alias{INDIVIDUAL_ATTRIBUTE_STRUCTURE}
\title{Construct the INDIVIDUAL_ATTRIBUTE_STRUCTURE tibble}
\usage{
INDIVIDUAL_ATTRIBUTE_STRUCTURE(
  attribute_type,
  attribute_descriptor,
  individual_event_details = INDIVIDUAL_EVENT_DETAIL()
)
}
\arguments{
\item{attribute_type}{An attribute which may have caused name, addresses, phone numbers,
family listings to be recorded. Its application is in helping to classify sources used for information.}

\item{attribute_descriptor}{Text describing a particular characteristic or attribute assigned to an individual.}

\item{individual_event_details}{An INDIVIDUAL_EVENT_DETAIL() object giving details of the attribute.}
}
\value{
A tidy tibble containing the INDIVIDUAL_ATTRIBUTE_STRUCTURE part of a GEDCOM file.
}
\description{
This function constructs a tibble representation of the INDIVIDUAL_ATTRIBUTE_STRUCTURE from the GEDCOM 5.5.1
specification.
}
