% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/structures.R
\name{individual_attribute_structure}
\alias{individual_attribute_structure}
\title{Construct the INDIVIDUAL_ATTRIBUTE_STRUCTURE from the GEDCOM specification}
\usage{
individual_attribute_structure(
  attribute_type,
  attribute_descriptor,
  individual_event_details = individual_event_detail()
)
}
\arguments{
\item{attribute_type}{An attribute which may have caused name, addresses, phone numbers,
family listings to be recorded. Its application is in helping to classify sources used for information.}

\item{attribute_descriptor}{Text describing a particular characteristic or attribute assigned to an individual.}

\item{individual_event_details}{An individual_event_detail() object giving details of the attribute.}
}
\value{
A tidy tibble containing the INDIVIDUAL_ATTRIBUTE_STRUCTURE part of a GEDCOM file.
}
\description{
Construct the INDIVIDUAL_ATTRIBUTE_STRUCTURE from the GEDCOM specification
}
