% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subrecords.R
\name{INDIVIDUAL_EVENT_DETAIL}
\alias{INDIVIDUAL_EVENT_DETAIL}
\title{Construct the INDIVIDUAL_EVENT_DETAIL tibble}
\usage{
INDIVIDUAL_EVENT_DETAIL(
  event_details = EVENT_DETAIL(),
  age_at_event = character()
)
}
\arguments{
\item{event_details}{An EVENT_DETAIL() object giving details of the event.}

\item{age_at_event}{A string that indicates the age in years, months, and days that the
principal was at the time of the associated event. Any labels must come after their
corresponding number, for example; 4y 8m 10d.
The line value should be normalised; it should for example not specify 2y 13m, but 3y 1m
instead. Number of days is allowed to be 365 because of leap years.
The YYY, MM and DDD values must not be zero; if a value equals zero, that part is left off.
The values may not contain leading zeroes either.}
}
\value{
A tidy tibble containing the INDIVIDUAL_EVENT_DETAIL part of a GEDCOM file.
}
\description{
This function constructs a tibble representation of the INDIVIDUAL_EVENT_DETAIL from the GEDCOM 5.5.5
specification.
}
