% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/structures.R
\name{individual_event_structure}
\alias{individual_event_structure}
\title{Constructs the INDIVIDUAL_EVENT_STRUCTURE from the GEDCOM specification}
\usage{
individual_event_structure(
  event_type_individual,
  individual_event_details = individual_event_detail(),
  xref_fam = character(),
  adopted_by_which_parent = character()
)
}
\arguments{
\item{event_type_individual}{A code used to indicate the type of individual event.}

\item{xref_fam}{Either a pointer or an unique cross-reference identifier.}
}
\value{

}
\description{
Constructs the INDIVIDUAL_EVENT_STRUCTURE from the GEDCOM specification
}
