% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subrecords.R
\name{INDIVIDUAL_EVENT_STRUCTURE}
\alias{INDIVIDUAL_EVENT_STRUCTURE}
\title{Construct the INDIVIDUAL_EVENT_STRUCTURE tibble}
\usage{
INDIVIDUAL_EVENT_STRUCTURE(
  event_type_individual,
  individual_event_details = INDIVIDUAL_EVENT_DETAIL(),
  xref_fam = character(),
  adopted_by_which_parent = character()
)
}
\arguments{
\item{event_type_individual}{A code used to indicate the type of individual event.}

\item{individual_event_details}{An INDIVIDUAL_EVENT_DETAIL() object giving details of the event.}

\item{xref_fam}{An xref ID of a family record.}

\item{adopted_by_which_parent}{A code which shows which parent in the associated family record adopted this person.
Use "HUSB" for husband, "WIFE" for wife, or "BOTH" for both.}
}
\value{
A tidy tibble containing the INDIVIDUAL_EVENT_STRUCTURE part of a GEDCOM file.
}
\description{
This function constructs a tibble representation of the INDIVIDUAL_EVENT_STRUCTURE from the GEDCOM 5.5.5
specification.
}
