% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/records.R
\name{INDIVIDUAL_RECORD}
\alias{INDIVIDUAL_RECORD}
\title{Construct the INDIVIDUAL_RECORD tibble}
\usage{
INDIVIDUAL_RECORD(
  xref_indi,
  names = list(),
  sex_value = character(),
  events = list(),
  attributes = list(),
  child_to_family_links = list(),
  spouse_to_family_links = list(),
  associations = list(),
  user_reference_number = character(),
  user_reference_type = character(),
  automated_record_id = character(),
  date_changed = CHANGE_DATE(),
  notes = list(),
  source_citations = list(),
  multimedia_links = list()
)
}
\arguments{
\item{xref_indi}{An xref ID of an individual record.}

\item{names}{A list of PERSONAL_NAME_STRUCTURE() objects giving the names associated with this individual.}

\item{sex_value}{A code that indicates the sex of the individual.}

\item{events}{A list of INDIVIDUAL_EVENT_STRUCTURE() objects giving the events associated with
this individual.}

\item{attributes}{A list of INDIVIDUAL_ATTRIBUTE_STRUCTURE() objects giving the attributes associated
with this individual.}

\item{child_to_family_links}{A list of CHILD_TO_FAMILY_LINK() objects giving the details of families
this individual is a child of.}

\item{spouse_to_family_links}{A list of SPOUSE_TO_FAMILY_LINK() objects giving the details of families
this individual is a spouse of.}

\item{associations}{A list of ASSOCIATION_STRUCTURE() objects giving the details of individuals this
individual is associated with.}

\item{user_reference_number}{A user-defined number or text that the submitter uses to identify this record.}

\item{user_reference_type}{A user-defined definition of the user_reference_number.}

\item{automated_record_id}{A unique record identification number assigned to the record by
the source system. This number is intended to serve as a more sure means of identification
of a record for reconciling differences in data between two interfacing systems.}

\item{date_changed}{A CHANGE_DATE() object giving the time this record was last modified. If not provided,
the current date is used.}

\item{notes}{A list of NOTE_STRUCTURE() objects.}

\item{source_citations}{A list of SOURCE_CITATION() objects.}

\item{multimedia_links}{A list of MULTIMEDIA_LINK() objects}
}
\value{
A tidy tibble containing an INDIVIDUAL_RECORD part of a GEDCOM file.
}
\description{
This function constructs a tibble representation of the INDIVIDUAL_RECORD from the GEDCOM 5.5.5
specification.
}
