% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subrecords.R
\name{MULTIMEDIA_LINK}
\alias{MULTIMEDIA_LINK}
\title{Construct the MULTIMEDIA_LINK tibble}
\usage{
MULTIMEDIA_LINK(xref_obje)
}
\arguments{
\item{xref_obje}{An xref ID of a Multimedia record.}
}
\value{
A tidy tibble containing the MULTIMEDIA_LINK part of a GEDCOM file.
}
\description{
This function constructs a tibble representation of the MULTIMEDIA_LINK from the GEDCOM 5.5.5
specification.
}
