% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subrecords.R
\name{NOTE_STRUCTURE}
\alias{NOTE_STRUCTURE}
\title{Construct the NOTE_STRUCTURE tibble}
\usage{
NOTE_STRUCTURE(xref_note = character(), user_text = character())
}
\arguments{
\item{xref_note}{An xref ID of a note record.}

\item{user_text}{Free-form user text. Comments, opinions.}
}
\value{
A tidy tibble containing the NOTE_STRUCTURE part of a GEDCOM file.
}
\description{
This function constructs a tibble representation of the NOTE_STRUCTURE from the GEDCOM 5.5.5
specification.
}
