% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/structures.R
\name{note_structure}
\alias{note_structure}
\title{Constructs the NOTE_STRUCTURE from the GEDCOM specification}
\usage{
note_structure(xref_note = character(), submitter_text = character())
}
\arguments{
\item{submitter_text}{Comments or opinions from the submitter.}
}
\value{

}
\description{
Constructs the NOTE_STRUCTURE from the GEDCOM specification
}
