% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interrogate.R
\name{num_indi}
\alias{num_indi}
\alias{num_fam}
\alias{num_subm}
\alias{num_subn}
\alias{num_media}
\alias{num_note}
\alias{num_repo}
\alias{num_sour}
\title{Get the number of records in a tidygedcom object}
\usage{
num_indi(gedcom)

num_fam(gedcom)

num_subm(gedcom)

num_subn(gedcom)

num_media(gedcom)

num_note(gedcom)

num_repo(gedcom)

num_sour(gedcom)
}
\arguments{
\item{gedcom}{A tidygedcom object.}
}
\value{
The number of records of the relevant type.
}
\description{
These functions return the number of records of a particular type in a tidygedcom object.
}
