% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/structures.R
\name{personal_name_pieces}
\alias{personal_name_pieces}
\title{Construct the PERSONAL_NAME_PIECES from the GEDCOM specification}
\usage{
personal_name_pieces(
  name_piece_prefix = character(),
  name_piece_given = character(),
  name_piece_nickname = character(),
  name_piece_surname_prefix = character(),
  name_piece_surname = character(),
  name_piece_suffix = character(),
  notes = list(),
  source_citations = list()
)
}
\arguments{
\item{name_piece_prefix}{Non indexing name piece that appears preceding the given name and surname parts.
Different name prefix parts are separated by a comma.}

\item{name_piece_given}{Given name or earned name. Different given names are separated by a comma.}

\item{name_piece_nickname}{A descriptive or familiar name used in connection with one's proper name.}

\item{name_piece_surname_prefix}{Surname prefix or article used in a family name. Different surname
articles are separated by a comma, for example in the name "de la Cruz", this value would be "de, la".}

\item{name_piece_surname}{Surname or family name. Different surnames are separated by a comma.}

\item{name_piece_suffix}{Non-indexing name piece that appears after the given name and surname parts.
Different name suffix parts are separated by a comma.}

\item{notes}{A list of note_structure() objects.}

\item{source_citations}{A list of source_citation() objects.}
}
\value{
A tidy tibble containing the PERSONAL_NAME_PIECES part of a GEDCOM file.
}
\description{
Construct the PERSONAL_NAME_PIECES from the GEDCOM specification
}
