% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/structures.R
\name{personal_name_pieces}
\alias{personal_name_pieces}
\title{Constructs the PERSONAL_NAME_PIECES from the GEDCOM specification}
\usage{
personal_name_pieces(
  prefix = character(),
  given = character(),
  nickname = character(),
  surname_prefix = character(),
  surname = character(),
  suffix = character(),
  notes = list(),
  source_citations = list()
)
}
\value{

}
\description{
Constructs the PERSONAL_NAME_PIECES from the GEDCOM specification
}
