% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subrecords.R
\name{PERSONAL_NAME_STRUCTURE}
\alias{PERSONAL_NAME_STRUCTURE}
\title{Construct the PERSONAL_NAME_STRUCTURE tibble}
\usage{
PERSONAL_NAME_STRUCTURE(
  name_personal,
  name_type = character(),
  name_pieces = PERSONAL_NAME_PIECES(),
  name_phonetic = character(),
  phonetisation_method = character(),
  phonetic_name_pieces = list(),
  name_romanised = character(),
  romanisation_method = character(),
  romanised_name_pieces = list()
)
}
\arguments{
\item{name_personal}{The full name formed in the manner the name is normally spoken.
The surname of an individual, if known, is enclosed between two slash (/) characters.
The order of the name parts should be the order that the person would, by custom of their culture,
have used when giving it to a recorder.}

\item{name_type}{Indicates the name type, for example the name issued or assumed as an immigrant.}

\item{name_pieces}{A PERSONAL_NAME_PIECES() object giving the components of the name.}

\item{name_phonetic}{A character vector of phonetic variations of the full name.}

\item{phonetisation_method}{Indicates the method used in transforming the text to the phonetic variation.}

\item{phonetic_name_pieces}{A list of PERSONAL_NAME_PIECES() objects giving the components
of the phonetic name variations.}

\item{name_romanised}{A character vector giving romanised variations of the name.}

\item{romanisation_method}{Indicates the method used in transforming the text to a romanised variation.}

\item{romanised_name_pieces}{A list of PERSONAL_NAME_PIECES() objects giving the components
of the romanised name variations.}
}
\value{
A tidy tibble containing the PERSONAL_NAME_STRUCTURE part of a GEDCOM file.
}
\description{
This function constructs a tibble representation of the PERSONAL_NAME_STRUCTURE from the GEDCOM 5.5.5
specification.
}
