% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/structures.R
\name{personal_name_structure}
\alias{personal_name_structure}
\title{Constructs the PERSONAL_NAME_STRUCTURE from the GEDCOM specification}
\usage{
personal_name_structure(
  name,
  type = character(),
  name_pieces = personal_name_pieces(),
  phonetic_variation = character(),
  phonetic_type = character(),
  phonetic_name_pieces = list(),
  romanized_variation = character(),
  romanized_type = character(),
  romanized_name_pieces = list()
)
}
\arguments{
\item{phonetic_type}{Indicates the method used in transforming the text to the phonetic variation.}

\item{romanized_type}{Indicates the method used in transforming the text to a romanized variation.}
}
\value{

}
\description{
Constructs the PERSONAL_NAME_STRUCTURE from the GEDCOM specification
}
