% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/structures.R
\name{PLACE_STRUCTURE}
\alias{PLACE_STRUCTURE}
\title{Construct the PLACE_STRUCTURE tibble}
\usage{
PLACE_STRUCTURE(
  place_name,
  place_hierarchy = character(),
  place_phonetic_variation = character(),
  phonetic_type = character(),
  place_romanized_variation = character(),
  romanized_type = character(),
  place_latitude = character(),
  place_longitude = character(),
  notes = list()
)
}
\arguments{
\item{place_name}{The jurisdictional name of the place where the event took place.
Jurisdictions are separated by commas.}

\item{place_hierarchy}{This shows the jurisdictional entities that are named in a sequence from the
lowest to the highest jurisdiction. The jurisdictions are separated by commas, and any jurisdiction's
name that is missing is still accounted for by a comma.}

\item{place_phonetic_variation}{A character vector of phonetic variations of the place name.}

\item{phonetic_type}{Indicates the method used in transforming the text to the phonetic variation.}

\item{place_romanized_variation}{A character vector of romanized variations of the place name.}

\item{romanized_type}{Indicates the method used in transforming the text to a romanized variation.}

\item{place_latitude}{The value specifying the latitudinal coordinate of the place name.
The latitude coordinate is the direction North or South from the equator in degrees and fraction of
degrees carried out to give the desired accuracy. For example: 18 degrees, 9 minutes, and 3.4 seconds North
would be formatted as N18.150944.}

\item{place_longitude}{The value specifying the longitudinal coordinate of the place name.
The longitude coordinate is Degrees and fraction of degrees east or west of the zero or base
meridian coordinate. For example: 168 degrees, 9 minutes, and 3.4 seconds East would be formatted as E168.150944.}

\item{notes}{A list of NOTE_STRUCTURE() objects.}
}
\value{
A tidy tibble containing the PLACE_STRUCTURE part of a GEDCOM file.
}
\description{
This function constructs a tibble representation of the PLACE_STRUCTURE from the GEDCOM 5.5.1
specification.
}
