% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/structures.R
\name{place_structure}
\alias{place_structure}
\title{Constructs the PLACE_STRUCTURE from the GEDCOM specification}
\usage{
place_structure(
  place,
  place_hierarchy = character(),
  phonetic_variation = character(),
  phonetic_type = character(),
  romanized_variation = character(),
  romanized_type = character(),
  latitude = character(),
  longitude = character(),
  notes = list()
)
}
\arguments{
\item{place_hierarchy}{This shows the jurisdictional entities that are named in a sequence from the
lowest to the highest jurisdiction. The jurisdictions are separated by commas, and any jurisdiction's
name that is missing is still accounted for by a comma.}

\item{phonetic_type}{Indicates the method used in transforming the text to the phonetic variation.}

\item{romanized_type}{Indicates the method used in transforming the text to a romanized variation.}
}
\value{

}
\description{
Constructs the PLACE_STRUCTURE from the GEDCOM specification
}
