% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_write_create.R
\name{read_gedcom}
\alias{read_gedcom}
\title{Import a GEDCOM file}
\usage{
read_gedcom(filepath)
}
\arguments{
\item{filepath}{The full filepath of the GEDCOM file}
}
\value{
A tidygedcom object
}
\description{
Imports a *.ged file and creates a tidygedcom object.
}
\examples{
\dontrun{
read_gedcom("C:/my_family.ged")
}
}
