% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/edit_family.R
\name{remove_family}
\alias{remove_family}
\title{Remove a Family record from a tidygedcom object}
\usage{
remove_family(gedcom, remove_individuals = FALSE)
}
\arguments{
\item{gedcom}{A tidygedcom object.}

\item{remove_individuals}{Whether to also remove the individual records for all individuals
in the family.}
}
\value{
An updated tidygedcom object excluding the active Family record (and potentially the
individuals within it).
}
\description{
Remove a Family record from a tidygedcom object
}
