% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/edit_individual.R
\name{remove_individual}
\alias{remove_individual}
\title{Remove an Individual record from a tidygedcom object}
\usage{
remove_individual(gedcom, xref = character(), remove_associations = TRUE)
}
\arguments{
\item{gedcom}{A tidygedcom object.}

\item{xref}{The xref of a record to act on if one is not activated (will override active record).}

\item{remove_associations}{Whether to also remove associations with this individual in
other individual records. Defaults to TRUE.}
}
\value{
An updated tidygedcom object excluding the active Individual record.
}
\description{
This function removes an active Individual record from the tidygedcom object.
}
\details{
At a minimum it will also remove references to this individual in Family group records.
If remove_associations is TRUE (default) it will remove associations with this
individual in other Individual records.
}
