% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/edit_individual.R
\name{remove_individual}
\alias{remove_individual}
\title{Remove an Individual record from a tidygedcom object}
\usage{
remove_individual(gedcom, remove_aliases = FALSE, remove_associations = TRUE)
}
\arguments{
\item{gedcom}{A tidygedcom object.}

\item{remove_aliases}{Whether to also remove the individual records given as aliases of
this individual. Defaults to FALSE.}

\item{remove_associations}{Whether to also remove associations with this individual in
other individual records. Defaults to TRUE.}
}
\value{
An updated tidygedcom object excluding the active Individual record.
}
\description{
Remove an Individual record from a tidygedcom object
}
