% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/edit_note.R
\name{remove_note}
\alias{remove_note}
\title{Remove a Note record from a tidygedcom object}
\usage{
remove_note(gedcom)
}
\arguments{
\item{gedcom}{A tidygedcom object.}
}
\value{
An updated tidygedcom object excluding the active Note record.
}
\description{
Remove a Note record from a tidygedcom object
}
