% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/edit_repository.R
\name{remove_repository}
\alias{remove_repository}
\title{Remove a Repository record from a tidygedcom object}
\usage{
remove_repository(gedcom)
}
\arguments{
\item{gedcom}{A tidygedcom object.}
}
\value{
An updated tidygedcom object excluding the active Repository record.
}
\description{
Remove a Repository record from a tidygedcom object
}
