% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subrecords.R
\name{SOURCE_CITATION}
\alias{SOURCE_CITATION}
\title{Construct the SOURCE_CITATION tibble}
\usage{
SOURCE_CITATION(
  xref_sour,
  where_within_source = character(),
  event_type_cited_from = character(),
  role_in_event = character(),
  entry_recording_date = date_value(),
  text_from_source = character(),
  certainty_assessment = character(),
  multimedia_links = list(),
  notes = list()
)
}
\arguments{
\item{xref_sour}{An xref ID of a source record.}

\item{where_within_source}{Specific location with in the information referenced.}

\item{event_type_cited_from}{A code that indicates the type of event which was responsible for the
source entry being recorded. For example, if the entry was created to record a birth of a child,
then the type would be BIRT regardless of the assertions made from that record, such as the mother's
name or mother's birth date. This will allow a prioritized best view choice and a determination of
the certainty associated with the source used in asserting the cited fact.}

\item{role_in_event}{Indicates what role this person played in the event that is being cited in this context.}

\item{entry_recording_date}{A date_value() object giving the date that this event data
was entered into the original source document.}

\item{text_from_source}{A verbatim copy of any description contained within the source.
This indicates notes or text that are actually contained in the source document, not the submitter's
opinion about the source.}

\item{certainty_assessment}{A quantitative evaluation of the credibility of a piece of
information, based upon its supporting evidence. Some systems use this feature to rank multiple
conflicting opinions for display of most likely information first. It is not intended to
eliminate the receiver's need to evaluate the evidence for themselves.
0 = unreliable/estimated data
1 = Questionable reliability of evidence
2 = Secondary evidence, data officially recorded sometime after event
3 = Direct and primary evidence used, or by dominance of the evidence}

\item{multimedia_links}{A list of MULTIMEDIA_LINK() objects}

\item{notes}{A list of NOTE_STRUCTURE() objects.}
}
\value{
A tidy tibble containing the SOURCE_CITATION part of a GEDCOM file.
}
\description{
This function constructs a tibble representation of the SOURCE_CITATION from the GEDCOM 5.5.5
specification.
}
