% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subrecords.R
\name{SOURCE_REPOSITORY_CITATION}
\alias{SOURCE_REPOSITORY_CITATION}
\title{Construct the SOURCE_REPOSITORY_CITATION tibble}
\usage{
SOURCE_REPOSITORY_CITATION(xref_repo, source_call_number = character())
}
\arguments{
\item{xref_repo}{An xref ID of a Repository record.}

\item{source_call_number}{An identification or reference description used to file and retrieve items
from the holdings of a repository.}
}
\value{
A tidy tibble containing the SOURCE_REPOSITORY_CITATION part of a GEDCOM file.
}
\description{
This function constructs a tibble representation of the SOURCE_REPOSITORY_CITATION from the GEDCOM 5.5.5
specification.
}
