% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/structures.R
\name{source_repository_citation}
\alias{source_repository_citation}
\title{Constructs the SOURCE_REPOSITORY_CITATION from the GEDCOM specification}
\usage{
source_repository_citation(
  repo_ref,
  notes = list(),
  call_numbers = character(),
  media_type = character()
)
}
\value{

}
\description{
Constructs the SOURCE_REPOSITORY_CITATION from the GEDCOM specification
}
