% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_export_create.R
\name{split_gedcom_values}
\alias{split_gedcom_values}
\title{Convert the GEDCOM form to GEDCOM grammar}
\usage{
split_gedcom_values(gedcom, char_limit)
}
\arguments{
\item{gedcom}{A tidygedcom object.}
}
\value{
A tidygedcom object potentially expanded with CONC/CONT rows.
}
\description{
This function introduces CONC/CONT lines for line values that exceed the given number of characters.
This function only uses the CONC(atenation) tag for splitting values, and does
not use the CONT(inuation) tag. This is because it is easier to implement.
}
