% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subrecords.R
\name{SPOUSE_TO_FAMILY_LINK}
\alias{SPOUSE_TO_FAMILY_LINK}
\title{Construct the SPOUSE_TO_FAMILY_LINK tibble}
\usage{
SPOUSE_TO_FAMILY_LINK(xref_fam, notes = list())
}
\arguments{
\item{xref_fam}{An xref ID of a family record.}

\item{notes}{A list of NOTE_STRUCTURE() objects.}
}
\value{
A tidy tibble containing the SPOUSE_TO_FAMILY_LINK part of a GEDCOM file.
}
\description{
This function constructs a tibble representation of the SPOUSE_TO_FAMILY_LINK from the GEDCOM 5.5.5
specification.
}
