% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interrogate.R
\name{str.tidygedcom}
\alias{str.tidygedcom}
\title{Get the structure of a tidygedcom object}
\usage{
\method{str}{tidygedcom}(gedcom)
}
\arguments{
\item{gedcom}{A tidygedcom object.}
}
\value{
A printed summary of the GEDCOM file structure.
}
\description{
This function gives a breakdown of record counts in the GEDCOM file.
}
