% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create.R
\name{subm}
\alias{subm}
\title{Define a Submitter record for a new tidyged object}
\usage{
subm(
  name = unname(Sys.info()["user"]),
  subm_address = address(),
  subm_notes = character()
)
}
\arguments{
\item{name}{The name of the submitter.}

\item{subm_address}{An address() object containing the submitter address.}

\item{subm_notes}{A character vector of notes accompanying this Submitter record.}
}
\value{
A Submitter record to be incorporated into a new tidyged object.
}
\description{
Define a Submitter record for a new tidyged object
}
\details{
This function is supposed to be used in the gedcom() function to define a
new tidyged object.

This submitter record identifies the individual or organization that contributed
information contained in the GEDCOM file.

This function deviates from the specification slightly by omitting multimedia links.
The logic for this is that if the file is just being created, then there will be
no multimedia records to link to.
}
