% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/edit_submitter.R
\name{subm}
\alias{subm}
\title{Define a Submitter record for a new tidygedcom object}
\usage{
subm(
  name = unname(Sys.info()["user"]),
  local_address_lines = character(),
  city = character(),
  state = character(),
  postal_code = character(),
  country = character(),
  phone_number = character(),
  email = character(),
  fax = character(),
  web_page = character(),
  automated_record_id = character(),
  submitter_notes = character()
)
}
\arguments{
\item{name}{The name of the submitter.}

\item{city}{The city of the submitter.}

\item{state}{The state/county of the submitter.}

\item{postal_code}{The postal code of the submitter.}

\item{country}{The country of the submitter.}

\item{phone_number}{A character vector containing up to three phone numbers of the submitter.}

\item{email}{A character vector containing up to three email addresses of the submitter.}

\item{fax}{A character vector containing up to three fax numbers of the submitter.}

\item{web_page}{A character vector containing up to three web pages of the submitter.}

\item{automated_record_id}{A unique record identification number assigned to the record by
the source system.}

\item{submitter_notes}{A character vector of notes accompanying this Submitter record.
These could be xrefs to existing Note records.}

\item{address_first_line}{The first line of the submitter address.}

\item{language_preference}{The language in which the submitter prefers to communicate.}

\item{submitter_registered_rfn}{See the Gedcom 5.5.1 Standard for more details.}
}
\value{
A Submitter record to be incorporated into a new tidygedcom object.
}
\description{
Define a Submitter record for a new tidygedcom object
}
\details{
This function is supposed to be used in the gedcom() function to define a
new tidygedcom object.

This submitter record identifies the individual or organization that contributed
information contained in the GEDCOM file.

The function will automatically split the submitter_notes onto separate lines if the
character limit in the Gedcom standard is exceeded.
}
